% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_facet_call_pkg_prefix}
\alias{add_facet_call_pkg_prefix}
\title{Add the package prefix for \code{ggplot2::facet_grid} and \code{dplyr::vars}}
\usage{
add_facet_call_pkg_prefix(fun_call)
}
\description{
Add the package prefix for \code{ggplot2::facet_grid} and \code{dplyr::vars}
}
