% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{add_selector}
\alias{add_selector}
\title{Add data extract input to ui}
\usage{
add_selector(input, output, session, idx, label, data_extract)
}
\arguments{
\item{input}{shiny object}

\item{output}{shiny object}

\item{session}{shiny object}

\item{idx}{id of input}

\item{label}{label to be attached to data extract input}

\item{data_extract}{\code{data_extract_spect} to be rendered in ui}
}
\description{
Add data extract input to ui
}
