% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_choices.R
\name{choices_selected}
\alias{choices_selected}
\title{Choices function}
\usage{
choices_selected(choices, selected = choices[1])
}
\description{
Choices function
}
\examples{

choices_selected(LETTERS[1:5])


}
