% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_mosaic.R
\name{create_mosaic_layers}
\alias{create_mosaic_layers}
\title{Minimal mosaic plot}
\usage{
create_mosaic_layers(data_name, x_var, y_var, reduce_plot_call)
}
\arguments{
\item{data_name}{Name of the data frame to use.}

\item{x_var}{Name of the variable to use on the x-axis.}

\item{y_var}{Name of the variable to use for fill colors.}

\item{reduce_plot_call}{Function that takes multiple ggplot2 layers and combines them into a single plot call.}
}
\value{
An expression that creates a mosaic plot when evaluated.
}
\description{
Provides a minimal mosaic plot implementation using ggplot2.
}
\keyword{internal}
