% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{create_sparklines.default}
\alias{create_sparklines.default}
\title{Default method for \code{\link{create_sparklines}}}
\usage{
\method{create_sparklines}{default}(arr, width = 150, ...)
}
\arguments{
\item{arr}{vector of any type and length}

\item{width}{\code{numeric} the width of the \code{sparkline} widget (pixels)}

\item{...}{\code{list} additional options passed to bar plots of \code{jquery.sparkline};  see
\href{http://omnipotent.net/jquery.sparkline/#common}{\code{jquery.sparkline docs}}}
}
\description{
Default method for \code{\link{create_sparklines}}
}
\seealso{
\code{\link{create_sparklines}}
}
