% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{create_sparklines.POSIXct}
\alias{create_sparklines.POSIXct}
\title{Generates the HTML code for the \code{sparkline} widget}
\usage{
\method{create_sparklines}{POSIXct}(arr, ...)
}
\arguments{
\item{arr}{vector of any type and length}

\item{...}{\code{list} additional options passed to bar plots of \code{jquery.sparkline};  see
\href{http://omnipotent.net/jquery.sparkline/#common}{\code{jquery.sparkline docs}}}
}
\value{
\code{character} with HTML code for the \code{sparkline} widget
}
\description{
Generates the HTML code for the \code{sparkline} widget
}
\seealso{
\code{\link{create_sparklines}}
}
