% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{create_sparklines.POSIXlt}
\alias{create_sparklines.POSIXlt}
\title{Generates the HTML code for the \code{sparkline} widget}
\usage{
\method{create_sparklines}{POSIXlt}(arr, width = 150, bar_spacing = 5, bar_width = 20, ...)
}
\arguments{
\item{arr}{vector of any type and length}

\item{width}{\code{numeric} the width of the \code{sparkline} widget (pixels)}

\item{bar_spacing}{\code{numeric} the spacing between the bars (in pixels)}

\item{bar_width}{\code{numeric} the width of the bars (in pixels)}

\item{...}{\code{list} additional options passed to bar plots of \code{jquery.sparkline};  see
\href{http://omnipotent.net/jquery.sparkline/#common}{\code{jquery.sparkline docs}}}
}
\value{
\code{character} with HTML code for the \code{sparkline} widget
}
\description{
Generates the HTML code for the \code{sparkline} widget
}
\seealso{
\code{\link{create_sparklines}}
}
\keyword{internal}
