% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{create_sparklines}
\alias{create_sparklines}
\title{S3 generic for \code{sparkline} widget HTML}
\usage{
create_sparklines(arr, width = 150, ...)
}
\arguments{
\item{arr}{vector of any type and length}

\item{width}{\code{numeric} the width of the \code{sparkline} widget (pixels)}

\item{...}{\code{list} additional options passed to bar plots of \code{jquery.sparkline};  see
\href{http://omnipotent.net/jquery.sparkline/#common}{\code{jquery.sparkline docs}}}
}
\value{
character variable containing the HTML code of the \code{sparkline} HTML widget
}
\description{
Generates the \code{sparkline} HTML code corresponding to the input array.
For numeric variables creates a box plot, for character and factors - bar plot.
Produces an empty string for variables of other types.
}
