% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{current_extract_selectors}
\alias{current_extract_selectors}
\title{Current data extract selectors}
\usage{
current_extract_selectors(input)
}
\arguments{
\item{input}{shiny object}
}
\value{
list of selectors defined by dataname and selected filter
}
\description{
List of unique data extract selectors basing on the dataname and applied filters.
}
