% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_colur_picker.R
\name{.color_palette_discrete}
\alias{.color_palette_discrete}
\title{Color palette discrete}
\usage{
.color_palette_discrete(levels, color)
}
\arguments{
\item{levels}{(\code{character}) values of possible variable levels}

\item{color}{(\verb{named character}) valid color names (see \code{\link[=colors]{colors()}}) or hex-colors named by \code{levels}.}
}
\value{
\code{character} with hex colors named by \code{levels}.
}
\description{
To specify custom discrete colors to \code{plotly} or \code{ggplot} elements one needs to specify a vector named by
levels of variable used for coloring. This function allows to specify only some or none of the colors/levels
as the rest will be filled automatically.
}
