% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_reactable.R
\name{.make_reactable_columns_call}
\alias{.make_reactable_columns_call}
\title{Makes \code{reactable::colDef} call containing:
name = \if{html}{\out{<column label attribute>}}
cell = \if{html}{\out{<url formatter>}}
Arguments of \code{\link[reactable:colDef]{reactable::colDef()}} are specified only if necessary}
\usage{
.make_reactable_columns_call(dataset)
}
\arguments{
\item{dataset}{(\code{data.frame})}
}
\value{
named list of \code{colDef} calls
}
\description{
Makes \code{reactable::colDef} call containing:
name = \if{html}{\out{<column label attribute>}}
cell = \if{html}{\out{<url formatter>}}
Arguments of \code{\link[reactable:colDef]{reactable::colDef()}} are specified only if necessary
}
\keyword{internal}
