% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.plotly_selected_filter_children}
\alias{.plotly_selected_filter_children}
\title{Filter children on \code{plotly_selected}}
\usage{
.plotly_selected_filter_children(
  data,
  plot_dataname,
  xvar,
  yvar,
  plotly_selected,
  children_datanames
)
}
\arguments{
\item{data}{(\verb{reactive teal_data})}

\item{plot_dataname}{(\code{character(1)})}

\item{xvar}{(\code{character(1)})}

\item{yvar}{(\code{character(1)})}

\item{plotly_selected}{(\code{reactive})}

\item{children_datanames}{(\code{character})}
}
\description{
Filters children datanames according to:
\itemize{
\item selected x and y values on the plot (based on the parent dataset)
\item \code{\link[teal.data:join_keys]{teal.data::join_keys}} relationship between \code{children_datanames}
}
}
