% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mosaic.R
\name{.scale_x_mosaic}
\alias{.scale_x_mosaic}
\title{Determining scales for mosaics}
\usage{
.scale_x_mosaic(
  breaks = unique,
  minor_breaks = NULL,
  labels = unique,
  na.value = NA_real_,
  position = "bottom",
  ...
)
}
\arguments{
\item{breaks, labels, minor_breaks}{One of:
\itemize{
\item \code{NULL} for no breaks / labels.
\item \code{\link[ggplot2:waiver]{ggplot2::waiver()}} for the default breaks / labels computed by the scale.
\item A numeric / character vector giving the positions of the breaks / labels.
\item A function.
See \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}} for more details.
}}

\item{na.value}{The value to be used for \code{NA} values.}

\item{position}{For position scales, The position of the axis.
left or right for y axes, top or bottom for x axes.}

\item{...}{other arguments passed to \code{continuous_scale()}.}
}
\description{
Determining scales for mosaics
}
\keyword{internal}
