% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{establish_updating_selection}
\alias{establish_updating_selection}
\title{Creates observers updating the currently selected column}
\usage{
establish_updating_selection(datanames, input, plot_var, columns_names)
}
\arguments{
\item{datanames}{(\code{character}) the name of the dataset}

\item{input}{(\code{session$input}) the shiny session input}

\item{plot_var}{(\code{list}) the list containing the currently selected dataset (tab) and its column names}

\item{columns_names}{(\code{environment}) the environment containing bindings for each dataset}
}
\description{
The created observers update the column currently selected in the left-hand side
tabset panel.
}
\note{
Creates an observer for each dataset (each tab in the tabset panel).
}
\keyword{internal}
