% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{extract_filters_choices}
\alias{extract_filters_choices}
\title{Filter choices from \code{data_extract_spec}}
\usage{
extract_filters_choices(x)
}
\arguments{
\item{x}{\code{data_extract_spec} object}
}
\value{
lists with one element per choice alongside with dataname
}
\description{
Creates combination of filter choices from \code{data_extract_spec}. Each choice is being
converted into separate list with \code{dataname} and \code{filter} elements
}
