% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_input}
\alias{extract_input}
\title{Extract html id for \code{data_extract_ui}}
\usage{
extract_input(varname, dataname, filter = FALSE)
}
\arguments{
\item{varname}{character original html id.
This will be mostly retrieved with \code{ns("original id")} in \code{ui} or
\code{session$ns("original id")} in server function.}

\item{dataname}{character \code{dataname} from data_extract input.
This might be retrieved like \code{teal.transform::data_extract_spec(...)[[1]]$dataname}.}

\item{filter}{logical if the connected \code{extract_data_spec} is used with \code{filter} option.}
}
\description{
The \code{data_extract_ui} is located under extended html id.
We could not use \code{ns("original id")} for reference, as it is extended with specific suffixes.
}
\keyword{internal}
