% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{extract_input_id}
\alias{extract_input_id}
\title{HTML id of the extract input contained within another UI}
\usage{
extract_input_id(idx, session)
}
\arguments{
\item{idx}{index or identifier of the data extract input element}

\item{session}{shiny session object}
}
\value{
the html id of the element given the \code{idx}
}
\description{
HTML id of the extract input contained within another UI
}
