% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{extract_remove_id}
\alias{extract_remove_id}
\title{HTML id of the element to remove the whole extract UI}
\usage{
extract_remove_id(idx, session)
}
\arguments{
\item{idx}{index or identifier of the button to remove the data extract ui}

\item{session}{shiny session object}
}
\value{
the html id of the element given the \code{idx}
}
\description{
HTML id of the element to remove the whole extract UI
}
