% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{extract_ui_id}
\alias{extract_ui_id}
\title{HTML id of the UI element that contains the extract input}
\usage{
extract_ui_id(idx, session)
}
\arguments{
\item{idx}{index or identifier of the data extract UI that contains the input element}

\item{session}{shiny session object}
}
\value{
the html id of the element given the \code{idx}
}
\description{
HTML id of the UI element that contains the extract input
}
