% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{filter_outliers}
\alias{filter_outliers}
\title{Logical vector}
\usage{
filter_outliers(var, outlier_definition)
}
\arguments{
\item{var}{(\code{numeric}) a numeric vector}

\item{outlier_definition}{(\code{numeric}) if \code{0} then no outliers are removed, otherwise
outliers (those more than \verb{outlier_definition*IQR below/above Q1/Q3}) are removed}
}
\description{
Returns a logical vector.
Suitable for \code{dplyr::filter()} and data.frames.
}
\keyword{internal}
