% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_bivariate.R
\name{g_bp}
\alias{g_bp}
\title{Bivariate Plot}
\usage{
g_bp(x = NULL, y = NULL, freq = TRUE)
}
\description{
Bivariate Plot
}
\examples{

c1 <- rnorm(100)
c2 <- rnorm(100)
d1 <- factor(sample(letters[1:5], 100, TRUE), levels = letters[1:5])
d2 <- factor(sample(LETTERS[24:26], 100, TRUE), levels = LETTERS[24:26])

g_bp(x = c1)
g_bp(x = c1, freq = FALSE)
g_bp(y = c1)
g_bp(y = c1, freq = FALSE)
g_bp(x = d1)
g_bp(x = d1, freq = FALSE)
g_bp(y = d1)
g_bp(y = d1, freq = FALSE)


g_bp(x = c1, y = c2)
g_bp(x = c1, y = d1)
g_bp(x = d1, y = c1)
g_bp(x = d1, y = d2) ## no color




x <- d1
library(scales)
ggplot() + aes(x = x) + geom_bar(aes(y = ..prop.., group = 1))




}
