% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_mosaic.R
\name{geom_mosaic}
\alias{geom_mosaic}
\title{Mosaic Rectangles Layer for ggplot2}
\usage{
geom_mosaic(
  mapping = NULL,
  data = NULL,
  stat = "mosaic",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. Must specify \code{x} and \code{fill}.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data. Defaults to \code{"rects"}.}

\item{position}{Position adjustment. Defaults to \code{"identity"}.}

\item{...}{Other arguments passed to \code{layer()}.}

\item{na.rm}{Logical. Should missing values be removed?}

\item{show.legend}{Logical. Should this layer be included in the legends?}

\item{inherit.aes}{Logical. If \code{FALSE}, overrides default aesthetics.}
}
\value{
A ggplot2 layer that adds mosaic rectangles to the plot.
}
\description{
Adds a mosaic-style rectangles layer to a ggplot, visualizing the joint distribution of categorical variables.
Each rectangle's size reflects the proportion of observations for combinations of \code{x} and \code{fill}.
}
\examples{
df <- data.frame(RACE = c("Black", "White", "Black", "Asian"), SEX = c("M", "M", "F", "F"))
library(ggplot2)
ggplot(df, aes(x = RACE, fill = SEX)) + geom_rects()
}
