% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{get_active_inputs}
\alias{get_active_inputs}
\title{Get list of non-empty shiny inputs}
\usage{
get_active_inputs(input, inverse = FALSE)
}
\arguments{
\item{input}{shiny input object}

\item{inverse}{\code{logical} \code{TRUE} for non-active input names}
}
\value{
names of the inputs which are not \code{NULL}
}
\description{
Get list of non-empty shiny inputs
}
