% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_data_table.R
\name{get_datanames_selected}
\alias{get_datanames_selected}
\title{a tool for ui and server for getting datanames taking into account the datasets_selected vector}
\usage{
get_datanames_selected(datasets, datasets_selected)
}
\arguments{
\item{datasets}{teal datasets object}

\item{datasets_selected}{(\code{character}) a character vector that says which datasets should be
shown and in what order. Names in a vector have to correspond with datasets names.}
}
\value{
(\code{character}) a character vector
}
\description{
a tool for ui and server for getting datanames taking into account the datasets_selected vector
}
\keyword{internal}
