% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_expr_table_args}
\alias{get_expr_table_args}
\title{Transform \code{basic_table_args} into \code{basic_table} expression}
\usage{
get_expr_table_args(
  basic_table_args,
  table_name = "default",
  nest_table_args = NULL
)
}
\arguments{
\item{basic_table_args}{(`basic_table_args`)}

\item{table_name}{(`character`)}

\item{nest_table_args}{(`basic_table_args`)}
}
\description{
a function
}
