% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{get_extract_active_inputs}
\alias{get_extract_active_inputs}
\title{Active inputs}
\usage{
get_extract_active_inputs(idx, input, inverse = FALSE)
}
\arguments{
\item{idx}{pattern to be found in names of the input object}

\item{input}{shiny input object}

\item{inverse}{\code{logical = TRUE} for non-active inputs}
}
\value{
vector of input names matching id pattern
}
\description{
Active inputs
}
