% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplot2_params}
\alias{ggplot2_params}
\title{Additional ggplot2 arguments}
\arguments{
\item{ggplot2_args}{(\code{list}) A list of lists for each module plot.}
}
\description{
Contains arguments that are shared between multiple functions
  in the package to avoid repetition using `inheritParams`.
  It has to be separated from `shared_params` as not all NEST modules is using ggplot2.
}
