% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplot_args}
\alias{ggplot_args}
\title{Building a list of the class ggplot_args}
\usage{
ggplot_args(labs = list(), theme = list())
}
\arguments{
\item{labs}{(named `list`) where all fields have to match `ggplot2::labs` arguments.}

\item{theme}{(named `list`) where all fields have to match `ggplot2::theme` arguments.}
}
\value{
(`list`) with the class `ggplot_args`.
}
\description{
this function has to be used to build an input for a `ggplot2_args` argument.
The `ggplot2_args` argument is part of every module which contains any `ggplot2` graphics.
The input is validated to match its `ggplot2` equivalent.
}
