% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tab_active_js}
\alias{is_tab_active_js}
\title{Get a string with java-script code checking if the specific tab is clicked}
\usage{
is_tab_active_js(id, name)
}
\arguments{
\item{id}{\code{character(1)} the id of the tab panel with tabs.}

\item{name}{\code{character(1)} the name of the tab.}
}
\description{
will be the input for \code{shiny::conditionalPanel()}
}
\keyword{internal}
