% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_landing_popup.R
\name{landing_popup}
\alias{landing_popup}
\title{Landing Popup}
\usage{
landing_popup(title = NULL, content = NULL, buttons = modalButton("Accept"))
}
\arguments{
\item{title}{\code{character(1)} the text to be displayed as a title of the popup.}

\item{content}{\code{character(1)} the content of the popup. Passed to \code{...} of \code{shiny::modalDialog}. Can be a \code{character}
or a text input control (like \code{textInput}) or a list of \code{shiny} tags. See examples.}

\item{buttons}{\code{shiny} tag or a list of tags (\code{tagList}). Typically a \code{modalButton} or \code{actionButton}. See examples.}
}
\description{
This function creates a simple landing welcome popup for \code{teal} applications and
can be used in a \code{teal::init(extra_server = )} parameter.
}
\examples{
app1 <- teal::init(
  data = teal.data::dataset("iris", iris),
  modules = teal::modules(
    teal.modules.general::tm_front_page("A")
  ),
  extra_server = teal.modules.general::landing_popup(
    title = "Welcome",
    content = "A place for the welcome message or a disclaimer statement.",
    buttons = modalButton("Proceed")
  )
)
if (interactive()) {
  shinyApp(app1$ui, app1$server)
}

app2 <- teal::init(
  data = teal.data::dataset("iris", iris),
  modules = teal::modules(
    teal.modules.general::tm_front_page("A")
  ),
  extra_server = teal.modules.general::landing_popup(
    title = "Welcome",
    content = div(tags$b("A place for the welcome message or a disclaimer statement.", style = "color: red;")),
    buttons = tagList(modalButton("Proceed"), actionButton("close", "Read more", onclick = "window.open('http://google.com', '_blank')"))
  )
)

if (interactive()) {
  shinyApp(app2$ui, app2$server)
}

}
