% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_or_null}
\alias{list_or_null}
\title{Returns NULL or an argument coerced to a list}
\usage{
list_or_null(obj)
}
\arguments{
\item{obj}{object to be tested for NULL}
}
\value{
: NULL if obj is NULL, list(obj) otherwise
}
\description{
Returns NULL or an argument coerced to a list
}
\examples{
\dontrun{
a <- NULL
b <- c(1, 2)
list_or_null(a) # returns NULL
l <- list_or_null(b) # return list(b)
print(l)
}
}
