% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_last_chunk_outputs_attributes}
\alias{modify_last_chunk_outputs_attributes}
\title{Set the attributes of the last chunk outputs}
\usage{
modify_last_chunk_outputs_attributes(
  teal_card,
  attributes,
  n = 1,
  inner_classes = NULL,
  quiet = FALSE
)
}
\arguments{
\item{teal_card}{(\code{teal_card}) object to modify.}

\item{attributes}{(\code{list}) of attributes to set on the last chunk outputs.}

\item{n}{(\code{integer(1)}) number of the last element of \code{teal_card} to modify.
it will only change \code{chunk_output} objects.}

\item{inner_classes}{(\code{character}) classes within \code{chunk_output} that should be modified.
This can be used to only change \code{recordedplot}, \code{ggplot2} or other type of objects.}
}
\description{
Set the attributes of the last chunk outputs
}
\keyword{internal}
