% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_decorators}
\alias{normalize_decorators}
\title{Convert flat list of \code{teal_transform_module} to named lists}
\usage{
normalize_decorators(decorators, available_decorators = NULL)
}
\arguments{
\item{decorators}{(list of \code{teal_transformodules}) to normalize.}

\item{available_decorators}{(\code{character}) a character vector of available decorator names.
When \code{NULL} it assumes that all decorators should be applied, which can be used
when module only has 1 object.}
}
\value{
A named list of lists with \code{teal_transform_module} objects.
}
\description{
Convert flat list of \code{teal_transform_module} to named lists
}
\keyword{internal}
