% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{plot_var_summary}
\alias{plot_var_summary}
\title{Plot variable}
\usage{
plot_var_summary(var, var_lab)
}
\arguments{
\item{var}{vector of any type to be plotted. For numeric variables it produces histogram with
density line, for factors it creates frequency plot}

\item{var_lab}{text describing selected variable to be displayed on the plot}
}
\value{
plot
}
\description{
Creates summary plot with statistics relevant to data type.
}
