% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{possible_extract_selectors}
\alias{possible_extract_selectors}
\title{Possible data extracts}
\usage{
possible_extract_selectors(data_extracts, dataname)
}
\arguments{
\item{data_extracts}{list of \code{data_extract_spec}}

\item{dataname}{name of dataset}
}
\value{
list of possible data extract spec - only dataname and selected filter
}
\description{
Possible combination of dataname and selected filters from available \code{data_extract_spec}.
}
