% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADLB}
\alias{rADLB}
\title{Random lab analysis}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 102 columns.
}
\source{
internal
}
\usage{
rADLB
}
\description{
Random lab analysis
}
\keyword{datasets}
\keyword{internal}
