% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{remove_extract_ui_id}
\alias{remove_extract_ui_id}
\title{HTML id of the element to remove the whole extract UI}
\usage{
remove_extract_ui_id(idx)
}
\arguments{
\item{idx}{index or identifier of the button to remove the data extract ui}
}
\value{
the html id of the element given the idx
}
\description{
HTML id of the element to remove the whole extract UI
}
