% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{remove_selector}
\alias{remove_selector}
\title{Removes data extract input from ui}
\usage{
remove_selector(input, output, session, rv, idx)
}
\arguments{
\item{input}{shiny object}

\item{output}{shiny object}

\item{session}{shiny object}

\item{rv}{reactive object containing \code{data_extracts} slot}

\item{idx}{id of input}
}
\description{
Removes data extract input from DOM and set input reactive value to \code{NULL}
}
