% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{render_single_tab}
\alias{render_single_tab}
\title{Renders a single tab in the left-hand side tabset panel}
\usage{
render_single_tab(
  dataset_name,
  output,
  datasets,
  input,
  columns_names,
  plot_var
)
}
\arguments{
\item{dataset_name}{(`character`) the name of the dataset contained in the rendered tab}

\item{output}{(`session$output`) the shiny session output}

\item{datasets}{(`FilteredData`) the object containing all datasets}

\item{input}{(`session$input`) the shiny session input}

\item{columns_names}{(`environment`) the environment containing bindings for each dataset}

\item{plot_var}{(`list`) the list containing the currently selected dataset (tab) and its column names}
}
\description{
Renders a single tab in the left-hand side tabset panel. The rendered tab contains
information about one dataset out of many presented in the module.
}
