% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{render_tab_header}
\alias{render_tab_header}
\title{Renders the text headlining a single tab in the left-hand side tabset panel}
\usage{
render_tab_header(dataset_name, output, datasets)
}
\arguments{
\item{dataset_name}{(\code{character}) the name of the dataset of the tab}

\item{output}{(\code{session$output}) the shiny session output}

\item{datasets}{(\code{FilteredData}) the object containing all datasets}
}
\description{
Renders the text headlining a single tab in the left-hand side tabset panel
}
\keyword{internal}
