% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{render_tab_table}
\alias{render_tab_table}
\title{Renders the table for a single dataset in the left-hand side tabset panel}
\usage{
render_tab_table(dataset_name, output, datasets, input, columns_names)
}
\arguments{
\item{dataset_name}{(`character`) the name of the dataset}

\item{output}{(`session$output`) the shiny session output}

\item{datasets}{(`FilteredData`) the object containing all datasets}

\item{input}{(`session$input`) the shiny session input}

\item{columns_names}{(`environment`) the environment containing bindings for each dataset}
}
\description{
The table contains column names, column labels,
small summary about NA values and a sparkline (if appropriate).
}
