% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{render_tabset_panel_content}
\alias{render_tabset_panel_content}
\title{Renders the left-hand side `tabset` panel of the module}
\usage{
render_tabset_panel_content(
  datanames,
  output,
  datasets,
  input,
  columns_names,
  plot_var
)
}
\arguments{
\item{datanames}{(`character`) the name of the dataset}

\item{output}{(`session$output`) the shiny session output}

\item{datasets}{(`FilteredData`) the object containing all datasets}

\item{input}{(`session$input`) the shiny session input}

\item{columns_names}{(`environment`) the environment containing bindings for each dataset}

\item{plot_var}{(`list`) the list containing the currently selected dataset (tab) and its column names}
}
\description{
Renders the left-hand side `tabset` panel of the module
}
