% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_regression.R
\name{safe_chunks_eval}
\alias{safe_chunks_eval}
\title{Suppresses warning that chunk was already evaluated and validates that chunks are okay}
\usage{
safe_chunks_eval()
}
\value{
value of \code{chunks_eval}
}
\description{
Suppresses warning that chunk was already evaluated and validates that chunks are okay
}
