% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_p_scatterplot.R
\name{scatterplotly}
\alias{scatterplotly}
\title{Generate Scatter Plotly Code}
\usage{
scatterplotly(
  df,
  x_var,
  y_var,
  color_var,
  id_var,
  colors,
  source,
  tooltip_vars = NULL
)
}
\arguments{
\item{df}{(\code{language}) Symbol representing the data frame to plot}

\item{x_var}{(\code{character(1)}) Name of the variable to be used for x-axis}

\item{y_var}{(\code{character(1)}) Name of the variable to be used for y-axis}

\item{color_var}{(\code{character(1)}) Name of the variable to be used for coloring points}

\item{id_var}{(\code{character(1)}) Name of the identifier variable}

\item{colors}{(\code{character}) Named vector of colors for color_var levels}

\item{source}{(\code{character(1)}) Source identifier for plotly events}

\item{tooltip_vars}{(\code{character} or \code{NULL}) A vector of column names to be displayed in the tooltip.
If \code{NULL}, default tooltip is created showing id, x, y, and color variables.}
}
\value{
A code expression that when evaluated creates a plotly plot object
}
\description{
Creates code expression that generates a scatter plot with tooltips using plotly.
This function includes all the data manipulation and plot creation logic
from tm_p_scatterplot module, including label extraction, tooltip generation,
and event registration.
}
\examples{
# Generate code for a scatter plot
code <- scatterplotly(
  df = quote(iris_data),
  x_var = "Sepal.Length",
  y_var = "Petal.Length",
  color_var = "Species",
  id_var = "row_id",
  colors = c("setosa" = "red", "versicolor" = "blue", "virginica" = "green"),
  source = "scatterplot",
  tooltip_vars = c("Sepal.Width", "Petal.Width")
)

}
