% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_choices.R
\name{select_choices}
\alias{select_choices}
\title{Choices function}
\usage{
select_choices(choices, selected = choices[1])
}
\description{
Choices function
}
\examples{

select_choices(LETTERS[1:5])


}
