% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{selector_id}
\alias{selector_id}
\title{HTML id of the extract input contained within another UI}
\usage{
selector_id(idx, session)
}
\arguments{
\item{idx}{index of the selector}

\item{session}{shiny session object}
}
\value{
the html id of the element given the \code{idx}
}
\description{
HTML id of the extract input contained within another UI
}
