% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{selectors_waiting}
\alias{selectors_waiting}
\title{Index of the UI in rendering queue}
\usage{
selectors_waiting(idx, input, inverse = FALSE)
}
\arguments{
\item{idx}{index of the selector}

\item{input}{shiny input object}

\item{inverse}{if opposite search should be applied}
}
\value{
id which are not rendered yet (found in shiny input names)
or not-found if \code{inverse = TRUE}
}
\description{
Index of the UI in rendering queue
}
