% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{srv_decorate_teal_data}
\alias{srv_decorate_teal_data}
\alias{ui_decorate_teal_data}
\title{Wrappers around \code{srv_transform_teal_data} that allows to decorate the data}
\usage{
srv_decorate_teal_data(id, data, decorators, expr, keep_output = NULL)

ui_decorate_teal_data(id, decorators, ...)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module instance id.}

\item{data}{(\code{reactive} returning \code{teal_data})}

\item{expr}{(\code{reactive}) with expression to evaluate on the output of the
decoration. It must be compatible with \code{code} argument of \code{\link[teal.code:eval_code]{teal.code::eval_code()}}.
Default is \code{NULL} which won't evaluate any appending code.}

\item{keep_output}{(\code{character} or \code{NULL}) Names of output objects in the environment
that are will be added in the card for the reporter.
These are shown in the card via the \code{\link[tools:toHTML]{tools::toHTML()}} and \code{\link[teal.reporter:to_rmd]{to_rmd()}} implementations.}

\item{...}{(\code{\link{dots}}) additional arguments passed to future methods.}
}
\description{
Wrappers around \code{srv_transform_teal_data} that allows to decorate the data
}
\details{
\code{srv_decorate_teal_data} is a wrapper around \code{srv_transform_teal_data} that
allows to decorate the data with additional expressions.
When original \code{teal_data} object is in error state, it will show that error
first.

\code{ui_decorate_teal_data} is a wrapper around \code{ui_transform_teal_data}.
}
\keyword{internal}
