% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_response.R
\name{srv_g_response}
\alias{srv_g_response}
\title{ANL_filtered %>% ggplot() +
  aes(x = SEX) +
  geom_bar(aes(fill = AVALC), position =  "fill") +
  geom_text(stat='count', aes(label=..count.., vjust = -1), position =  "fill") +
  #scale_y_continuous(limits = c(0, 1.2)) +
  facet_grid(cols = vars(ARM)) +
  ylab("Distribution")}
\usage{
srv_g_response(input, output, session, datasets, dataname,
  code_data_processing)
}
\description{
ANL_filtered %>% ggplot() +
  aes(x = fct_rev(SEX)) +
  geom_bar(aes(fill = AVALC), position =  "fill") +
  geom_text(stat='count', aes(label=..count.., hjust = -0.3), position =  "fill") +
  scale_y_continuous(limits = c(0, 1.4)) +
  facet_grid(cols = vars(ARM)) +
  ylab("Distribution") +
  coord_flip()
}
