% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_mosaic.R
\docType{data}
\name{StatMosaic}
\alias{StatMosaic}
\title{Mosaic Statistic for ggplot2}
\format{
An object of class \code{StatMosaic} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatMosaic
}
\description{
Implements a custom statistic for mosaic plots in ggplot2,
calculating coordinates and axis breaks/labels for categorical
variables.
}
\details{
This statistic processes input data to compute the positions
and sizes of mosaic rectangles, as well as the appropriate
axis breaks and labels for the plot.
}
\keyword{internal}
