% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_args}
\alias{table_args}
\title{Building a list of the class \code{basic_table_args}}
\usage{
table_args(...)
}
\arguments{
\item{...}{arguments compatible with \code{rtables::basic_table()}.}
}
\value{
list of the \code{basic_table_args} class.
}
\description{
This function has to be used to build an input for a \code{basic_table_args} argument.
The \code{basic_table_args} argument is part of every module which contains any \code{rtables} tables
Arguments are validated to match its \code{rtables} equivalent.
}
