% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.modules.general-package.R
\docType{package}
\name{teal.devel}
\alias{teal.devel}
\alias{teal.devel-package}
\title{teal.modules.general: Enabling you to write modules in teal}
\description{
PLEASE UPDATE TEXT
}
