% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.modules.general.R
\docType{package}
\name{teal.modules.general-package}
\alias{teal.modules.general}
\alias{teal.modules.general-package}
\title{teal.modules.general: General modules to add to a teal application}
\description{
The modules in this package are generic modules that should work with any data set
(not necessarily for clinical trials data).
}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}

Authors:
\itemize{
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Maciej Nasinski
  \item Konrad Pagacz
  \item Nikolas Burkoff
}

Other contributors:
\itemize{
  \item Ondrej Slama \email{ondrej.slama@roche.com} [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
