% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.modules.general-package.R
\docType{package}
\name{teal.modules.general}
\alias{teal.modules.general}
\alias{teal.modules.general-package}
\title{teal.modules.general: Enabling you to write modules in teal}
\description{
The modules in this package are generic modules that should work with any data set
(not necessarily for clinical trials data). We are prototyping these modules and they are not
meant for wide adoption at the moment.
}
