% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_a_pca.R
\name{tm_a_pca}
\alias{tm_a_pca}
\title{Principal component analysis module}
\usage{
tm_a_pca(
  label = "Principal component analysis",
  dat,
  plot_height = c(600, 200, 2000),
  pre_output = NULL,
  post_output = NULL
)
}
\arguments{
\item{label}{(\code{character}) Label of the module}

\item{dat}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}).
Dataset used to compute PCA.}

\item{plot_height}{(\code{numeric}) \code{c(value, min and max)} of plot height slider}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
Principal component analysis module
}
\examples{

# ADSL example

library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

app <- teal::init(
  data = cdisc_data(cdisc_dataset("ADSL", ADSL)),
  modules = root_modules(
    tm_a_pca("PCA",
             data_extract_spec(
               dataname = "ADSL",
               select = select_spec(
                 choices = variable_choices(data = ADSL),
                 selected = c("BMRKR1", "AGE"),
                 multiple = TRUE
               ),
               filter = NULL
             )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}

}
