% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_barplot.R
\name{tm_bep_safety0}
\alias{tm_bep_safety0}
\title{Bar plot for safety events}
\usage{
tm_bep_safety0(label, dataname, arm_var, arm_var_choices = arm_var,
  bep_var, bep_var_choices = bep_var, cov_var,
  cov_var_choices = cov_var, response_var,
  response_var_choices = response_var, responder_var,
  responder_var_choices = responder_var, non_responder_var,
  non_responder_var_choices = non_responder_var, exclude_var,
  exclude_var_choices = exclude_var, color_coding_var,
  color_coding_var_choices = color_coding_var, plot_height = c(1200,
  400, 5000), pre_output = NULL, post_output = NULL,
  code_data_processing = NULL)
}
\description{
Bar plot for safety events
}
\examples{
\dontrun{
library(random.cdisc.data)

ASL <- radsl(N = 600)
ADTE <- radaette(ASL)

app <- teal::init(
 data = list(ASL = ASL, ADTE = ADTE),
 modules = root_modules(
   teal.modules.general::tm_bep_safety0(label = "My label", dataname = "ADTE",
       arm_var=c("STUDYID"), bep_var = "", cov_var = "", response_var = c("AETTE1"),responder_var_choices = c(
           "AETTE2","AETTE1"
           ),
       responder_var = "", non_responder_var = "", exclude_var = "",
       color_coding_var = ""
       )
 )
)

shinyApp(app$ui, app$server)
}

}
