% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_cross_table.R
\name{tm_cross_table}
\alias{tm_cross_table}
\title{Create a simple cross-table}
\usage{
tm_cross_table(label = "Cross table", x, y, useNA = c("ifany", "no",
  "always"), pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{(\code{chracter}) menu label}

\item{x}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Specification how the user can select data to get encoded in the rows of the cross table.
Please just use single selections inside the \code{select_spec}.}

\item{y}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Specification how the user can select data to get encoded in the columns of the cross table.
Please just use single selections inside the \code{select_spec}.}

\item{useNA}{(\code{character}) optional pre-selected option indicating how to utilize NA in
table display. One of \code{'ifany'}, \code{'always'}, \code{'no'}. If
missing then \code{'ifany'} will be used. If vector then only the first
one will be used.}

\item{pre_output}{(\code{shiny.tag}) html tags appended below the output}

\item{post_output}{(\code{shiny.tag}) html tags appended after the output}
}
\description{
Create a table with the \code{\link{table}[base]} function
}
\examples{
library(random.cdisc.data)

ADSL <- cadsl

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- cadsl",
    check = TRUE
  ),
  root_modules(
    tm_cross_table(
      "Table Choices",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = names(ADSL),
          selected = "BMRKR2",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = names(ADSL),
          selected = "ARM",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
