% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_file_viewer.R
\name{tm_file_viewer}
\alias{tm_file_viewer}
\title{File Viewer Teal Module}
\usage{
tm_file_viewer(label = "File Viewer Module", input_path = NULL)
}
\arguments{
\item{label}{(\code{character}) Label shown in the navigation item for the module.}

\item{input_path}{(`list`) A list of the input path to either specific files of accepted formats or a directory.}
}
\description{
The file viewer module provides a tool to upload and view static files.
Supported formats include text formats, PDF, PNG, APNG, JPEG, SVG, WEBP, GIF and BMP.
}
\examples{
library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl")
  ),
  modules = root_modules(
    tm_file_viewer(input_path = list("./"))
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

}
