% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_front_page.R
\name{tm_front_page}
\alias{tm_front_page}
\title{Front page module}
\usage{
tm_front_page(
  label = "Front page",
  bold_header_text = NULL,
  header_text = NULL,
  tables = list(),
  footnotes = character(0),
  show_metadata = FALSE
)
}
\arguments{
\item{label}{(\code{character}) Label shown in the navigation item for the module.}

\item{bold_header_text}{\code{string} text to be shown in bold at the top of the module}

\item{header_text}{\code{string} text to be shown (not in bold) at the top of the module}

\item{tables}{\verb{named list of dataframes} tables to be shown in the module}

\item{footnotes}{\code{character} a vector of text to be shown at the bottom of the module, the first
word of each element is shown in bold.}

\item{show_metadata}{\code{logical} should the metadata of the datasets be available on the module?}
}
\value{
A \code{teal} module to be used in \code{teal} applications
}
\description{
This \code{teal} module creates a simple front page for your application
}
\examples{
library(scda)

table_1 <- data.frame(Info = c("A", "B"), Text = c("A", "B"))
table_2 <- data.frame(`Column 1` = c("C", "D"), `Column 2` = c(5.5, 6.6), `Column 3`= c("A", "B"))
table_3 <- data.frame(Info = c("E", "F"), Text = c("G", "H"))

table_input <- list(
  "Table 1" = table_1,
  "Table 2" = table_2,
  "Table 3" = table_3
)

ADSL <- synthetic_cdisc_data("latest")$adsl
app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL,
      code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl",
      metadata = list("Author" = "NEST team", "data_source" = "synthetic data")
    ),
    check = TRUE
  ),
  modules = modules(
    tm_front_page(
      bold_header_text = "Important information here",
      header_text = "Other information added here",
      tables = table_input,
      footnotes = c("X is the first footnote", "Y is the second footnote"),
      show_metadata = TRUE
    )
  ),
  header = tags$h1("Sample Application"),
  footer = tags$p("Application footer"),
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
