% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_association.R
\name{tm_g_association}
\alias{tm_g_association}
\title{Stack Plots of variables and show association with reference variable}
\usage{
tm_g_association(label = "Association", dataname, var,
  show_association = TRUE, plot_height = c(600, 400, 5000),
  pre_output = NULL, post_output = NULL, code_data_processing = NULL)
}
\description{
Stack Plots of variables and show association with reference variable
}
\examples{

\dontrun{

library(random.cdisc.data)
ASL <- radsl()
ASL$BOOLEAN <- sample(c(TRUE, FALSE), nrow(ASL), TRUE)

attr(ASL, "source") <- "# asl import"
x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_g_association(
      dataname = "ASL",
      var = choices_selected(names(ASL), "AGE")
    )
  )
)

shinyApp(x$ui, x$server)
}

}
