% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_regression.R
\name{tm_g_regression}
\alias{tm_g_regression}
\title{Scatterplot and Regression Model}
\usage{
tm_g_regression(label = "Regression Analysis", dataname)
}
\description{
Scatterplot and Regression Model
}
\examples{

ASL <- data.frame(
  USUBJID = paste("id", 1:100, sep = "-")

)

}
