% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix_dummy.R
\name{tm_g_scatterplotmatrix_dummy}
\alias{tm_g_scatterplotmatrix_dummy}
\title{Create a scatterplot matrix}
\usage{
tm_g_scatterplotmatrix_dummy(label = "Scatterplot matrix", dataname,
  selected, plot_height = c(600, 200, 2000), pre_output = NULL,
  post_output = NULL)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{dataname}{Name of datasets used to generate the regression plot (just used for labeling).}

\item{selected}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Plotting variables from an incoming dataset with filtering and selecting.}

\item{plot_height}{(\code{numeric}) A vector of length three with \code{c(value, min and max)} for a slider
encoding the plot height.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
Create a scatterplot matrix
}
