% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_spiderplot.R
\name{tm_g_spiderplot}
\alias{tm_g_spiderplot}
\title{\code{teal} module: Spider Plot}
\usage{
tm_g_spiderplot(
  label = "Spiderplot",
  plot_dataname,
  time_var,
  value_var,
  subject_var,
  event_var,
  color_var,
  point_colors,
  point_symbols,
  plot_height = 600,
  table_datanames = character(0),
  reactable_args = list(),
  transformator = transformator
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{plot_dataname}{(\code{character(1)}) name of the dataset which visualization is builded on.}

\item{time_var}{(\code{character(1)}) name of the \code{numeric} column in \code{plot_dataname} to be used as x-axis.}

\item{value_var}{(\code{character(1)}) name of the \code{numeric} column in \code{plot_dataname} to be used as y-axis.
column.}

\item{subject_var}{(\code{character(1)}) name of the \code{factor} or \code{character} column in \code{plot_dataname}
to be used as grouping variable for displayed lines/points.}

\item{color_var}{(\code{character(1)}) name of the \code{factor} or \code{character} column in \code{plot_dataname}
to be used to differentiate colors and symbols.}

\item{point_colors}{(\verb{named character}) valid color names (see \code{\link[=colors]{colors()}}) or hex-colors named
by levels of \code{color_var} column.}

\item{point_symbols}{(\verb{named character}) valid plotly symbol name named  by levels of \code{color_var}
column.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{table_datanames}{(\code{character}) names of the datasets which should be listed below the plot
when some data points are selected. Objects named after \code{table_datanames} will be pulled from
\code{data} so it is important that data actually contains these datasets. Please be aware that
table datasets must be linked with \code{plot_dataname} by the relevant \code{\link[=join_keys]{join_keys()}}.
See section "Decorating Module" below for more details.}
}
\description{
Module visualizes value development in time grouped by subjects.
}
