% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_swimlane.R
\name{tm_g_swimlane}
\alias{tm_g_swimlane}
\title{\code{teal} module: Swimlane plot}
\usage{
tm_g_swimlane(
  label = "Swimlane",
  plot_dataname,
  time_var,
  subject_var,
  color_var,
  group_var,
  sort_var = NULL,
  point_colors = character(0),
  point_symbols,
  plot_height = 700,
  table_datanames = character(0),
  reactable_args = list()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{plot_dataname}{(\code{character(1)} or \code{choices_selected}) name of the dataset which visualization is builded on.}

\item{time_var}{(\code{character(1)} or \code{choices_selected}) name of the \code{numeric} column in \code{plot_dataname} to be used as x-axis.}

\item{subject_var}{(\code{character(1)} or \code{choices_selected}) name of the \code{factor} or \code{character}  column in \code{plot_dataname}
to be used as y-axis.}

\item{color_var}{(\code{character(1)} or \code{choices_selected}) name of the \code{factor} or \code{character}  column in \code{plot_dataname}
to name and color subject events in time.}

\item{group_var}{(\code{character(1)} or \code{choices_selected}) name of the \code{factor} or \code{character} column in \code{plot_dataname}
to categorize type of event.
(legend is sorted according to this variable, and used in toolip to display type of the event)
todo: this can be fixed by ordering factor levels}

\item{sort_var}{(\code{character(1)} or \code{choices_selected}) name(s) of the column in \code{plot_dataname} which
value determines order of the subjects displayed on the y-axis.}

\item{point_colors}{(\verb{named character}) valid color names (see \code{\link[=colors]{colors()}}) or hex-colors named
by levels of \code{color_var} column.}

\item{point_symbols}{(\verb{named character}) valid plotly symbol name named  by levels of \code{color_var}
column.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{table_datanames}{(\code{character}) names of the datasets which should be listed below the plot
when some data points are selected. Objects named after \code{table_datanames} will be pulled from
\code{data} so it is important that data actually contains these datasets. Please be aware that
table datasets must be linked with \code{plot_dataname} by the relevant \code{\link[=join_keys]{join_keys()}}.
See section "Decorating Module" below for more details.}
}
\description{
Module visualizes subjects' events in time.
}
