% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_waterfall.R
\name{tm_g_waterfall}
\alias{tm_g_waterfall}
\title{\code{teal} module: Waterfall plot}
\usage{
tm_g_waterfall(
  label = "Waterfall",
  plot_dataname,
  subject_var,
  value_var,
  sort_var = NULL,
  color_var = NULL,
  bar_colors = list(),
  value_arbitrary_hlines = c(0.2, -0.3),
  plot_title = "Waterfall plot",
  plot_height = 700,
  table_datanames = character(0),
  reactable_args = list()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{plot_dataname}{(\code{character(1)}) name of the dataset which visualization is builded on.}

\item{subject_var}{(\code{character(1)} or \code{choices_selected}) name of the \code{factor} or \code{character} column
in \code{plot_dataname} to be used as x-axis.}

\item{value_var}{(\code{character(1)} or \code{choices_selected}) name of the \code{numeric} column
in \code{plot_dataname} to be used as y-axis.}

\item{color_var}{(\code{character(1)} or \code{choices_selected}) name of the \code{factor} or \code{character} column in \code{plot_dataname}
to be used to differentiate bar colors.}

\item{bar_colors}{(\verb{named character}) valid color names (see \code{\link[=colors]{colors()}}) or hex-colors named
by levels of \code{color_var} column.}

\item{value_arbitrary_hlines}{(\code{numeric}) values in the same scale as \code{value_var} to horizontal
lines on the plot.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{table_datanames}{(\code{character}) names of the datasets which should be listed below the plot
when some data points are selected. Objects named after \code{table_datanames} will be pulled from
\code{data} so it is important that data actually contains these datasets. Please be aware that
table datasets must be linked with \code{plot_dataname} by the relevant \code{\link[=join_keys]{join_keys()}}.
See section "Decorating Module" below for more details.}
}
\description{
Module visualizes subjects sorted decreasingly by y-values.
}
