% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_landing_popup.R
\name{tm_landing_popup}
\alias{tm_landing_popup}
\title{Landing Popup module}
\usage{
tm_landing_popup(title = NULL, text = NULL, button = NULL, icon = NULL)
}
\arguments{
\item{title}{tba}

\item{text}{tba}

\item{button}{tba}

\item{icon}{tba}
}
\description{
This \code{teal} module creates a simple landing welcome popup for \code{teal} applications and
can be used in \code{teal::init(extra_server = )} parameter
}
