% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_p_waterfall.R
\name{tm_p_waterfall}
\alias{tm_p_waterfall}
\title{\code{teal} module: Waterfall plot}
\usage{
tm_p_waterfall(
  label = "Waterfall",
  plot_dataname,
  subject_var,
  value_var,
  sort_var = NULL,
  color_var = NULL,
  tooltip_vars = NULL,
  bar_colors = character(0),
  value_arbitrary_hlines = c(0.2, -0.3),
  plot_title = "Waterfall plot",
  plot_height = c(600, 400, 1200)
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{plot_dataname}{(\code{character(1)}) name of the dataset which visualization is builded on.}

\item{subject_var}{(\code{character(1)} or \code{choices_selected}) name of the \code{factor} or \code{character} column
in \code{plot_dataname} to be used as x-axis.}

\item{value_var}{(\code{character(1)} or \code{choices_selected}) name of the \code{numeric} column
in \code{plot_dataname} to be used as y-axis.}

\item{color_var}{(\code{character(1)} or \code{choices_selected}) name of the \code{factor} or \code{character} column in \code{plot_dataname}
to be used to differentiate bar colors.}

\item{tooltip_vars}{(\code{character} or \code{NULL}) A vector of column names to be displayed in the tooltip.
If \code{NULL}, default tooltip is created.}

\item{bar_colors}{(\verb{named character}) valid color names (see \code{\link[=colors]{colors()}}) or hex-colors named
by levels of \code{color_var} column.}

\item{value_arbitrary_hlines}{(\code{numeric}) values in the same scale as \code{value_var} to horizontal
lines on the plot.}

\item{plot_title}{(\code{character}) Title of the plot.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}
}
\description{
Module visualizes subjects sorted decreasingly by y-values.
}
\examples{
data <- teal_data() |>
  within({
    subjects <- data.frame(
      subject_var = c("A", "B", "C"),
      AGE = sample(30:100, 3),
      ARM = c("Combination", "Combination", "Placebo")
    )

    waterfall_ds <- data.frame(
      subject_var = sample(c("A", "B", "C"), 10, replace = TRUE),
      value_var = sample(-20:90, 10, replace = TRUE),
      color_var = sample(c("CR", "PR", "SD", "PD"), 10, replace = TRUE)
    )
  })
join_keys(data) <- join_keys(
  join_key("subjects", "waterfall_ds", keys = c(subject_var = "subject_var"))
)

app <- init(
  data = data,
  modules = modules(
    tm_p_waterfall(
      plot_dataname = "waterfall_ds",
      subject_var = "subject_var",
      value_var = "value_var",
      sort_var = "value_var",
      color_var = "color_var",
      tooltip_vars = c("value_var", "subjects"),
      value_arbitrary_hlines = c(20, -30),
      bar_colors = c(
        CR = "#FF0000", PR = "#00FF00", SD = "#0000FF", PD = "#FFFF00"
      )
    )
  )
)

if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
