% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_markdown.R
\name{tm_rmarkdown}
\alias{tm_rmarkdown}
\title{\code{teal} module: Rmarkdown page}
\usage{
tm_rmarkdown(
  label = "App Info",
  text = character(0),
  params = list(title = "Document"),
  datanames = "all"
)
}
\arguments{
\item{text}{(\code{character}) arbitrary Rmd code}

\item{params}{A list of named parameters that override custom params
specified within the YAML front-matter (e.g. specifying a dataset to read or
a date range to confine output to). Pass \code{"ask"} to start an
application that helps guide parameter configuration.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Render arbitrary Rmarkdown code. \code{data} provided to teal application are available in the
rendered document.
}
\examples{
data <- teal_data() |>
  within({
    iris <- iris
    mtcars <- mtcars
  })
#

}
\section{Examples in Shinylive}{
\describe{
  \item{example-1}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqX8opKAeNdqAfQ8vG10AHwA+JV1dAHdaUgALFXYQKOjdWkZaUR8MrNE06JhSAhYco11i0sYCiGiAX0UIAGIlMHqAXSA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqX8opKAeNdqAfQ8vG10AHwA+JV1dAHdaUgALFXYQKOjdWkZaUR8MrNE06JhSAhYco11i0sYCiGiAX0UIAGIlMHqAXSA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
}
}

