% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_percentage_cross_table.R
\name{tm_t_percentage_cross_table}
\alias{tm_t_percentage_cross_table}
\title{Cross table based on rtables}
\usage{
tm_t_percentage_cross_table(label = "Cross Table", x, y,
  pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{(\code{character}) Label of the app in the teal menu}

\item{x}{(\code{choices_selected}) object with all available
choices with preselected option for variable X}

\item{y}{(\code{choices_selected}) object with all available
choices with preselected option for variable Y}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{...}{Arguments passed on to \code{teal.devel::standard_layout}
\describe{
  \item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
  \item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}}
}
\description{
Cross table based on rtables
}
\examples{
# Percentage cross table of variables from ADSL dataset
library(random.cdisc.data)

ADSL <- cadsl

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- cadsl",
    check = TRUE
  ),
  modules = root_modules(
    tm_t_percentage_cross_table(
      label = "Cross Table",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = c("COUNTRY", "STUDYID"),
          selected = "COUNTRY",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = c("SEX", "RACE"),
          selected = "SEX",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
}
\author{
wolfs25 waddella
}
