% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_table.R
\name{tm_table}
\alias{tm_table}
\title{Create a simple cross-table}
\usage{
tm_table(label, dataname, xvar, yvar, useNA = c("ifany", "no", "always"),
  pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{menu label}

\item{dataname}{name of dataset used to generate table}

\item{xvar}{(\code{choices_selected}) variable name of x variable. vector with variable names of possible x
variables. If
  missing or identincal to \code{xvar} then the table will be fixed to the
  \code{xvar}.}

\item{yvar}{(\code{choices_selected}) variable name of y variable. vector with variable names of possible y
variables. If
  missing or identincal to \code{xvar} then the table will be fixed to the
  \code{yvar}.}

\item{useNA}{optional pre-selected option indicating how to utilize NA in
table display. One of \code{'ifany'}, \code{'always'}, \code{'no'}. If
missing then \code{'ifany'} will be used. If vector then only the first
one will be used.}

\item{pre_output}{html tags appended below the output}

\item{post_output}{html tags appended after the output}
}
\description{
Create a table with the \code{\link{table}[base]} function
}
\examples{

library(random.cdisc.data)

asl <- radsl(seed = 1)

attr(asl, "source") <- "random.cdisc.data::radsl(seed = 1)"

x <- teal::init(
  data = list(ASL = asl),
  root_modules(
    tm_table("Table Choices",
      "ASL",
      xvar = choices_selected(c("SEX", "RACE", "STUDYID"), "SEX"),
      yvar = choices_selected(c("RACE", "SAFFL"), "RACE")
    ),
    tm_table("Table No Choices", "ASL",
      xvar = choices_selected("SEX"),
      yvar = choices_selected("RACE"),
      pre_output = helpText("Titles"),
      post_output = helpText("Footnotes")
    )
  )
)

\dontrun{
shinyApp(x$ui, x$server)
}
}
