% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_chunk_example.R
\name{tm_table_with_chunks}
\alias{tm_table_with_chunks}
\title{Create a simple cross-table}
\usage{
tm_table_with_chunks(label, dataname, xvar, yvar, pre_output = NULL,
  post_output = NULL)
}
\arguments{
\item{label}{menu label}

\item{dataname}{name of dataset used to generate table}

\item{xvar}{(\code{choices_selected}) variable name of x variable}

\item{yvar}{(\code{choices_selected}) variable name of y variable}

\item{pre_output}{html tags appended below the output}

\item{post_output}{html tags appended after the output}
}
\description{
Create a table with the \code{\link{table}[base]} function
}
\examples{

\dontrun{
library(random.cdisc.data)

ASL <- radsl(seed = 1)

attr(ASL, "source") <- "random.cdisc.data::radsl(seed = 1)"

x <- teal::init(
  data = list(ASL = ASL),
  root_modules(
    tm_data_table(),
    tm_variable_browser(),
    tm_table("Table Choices", "ASL",
      xvar = "SEX", yvar = "RACE",
      xvar_choices = c("SEX", "RACE", "STUDYID"),
      yvar_choices = c("RACE", "SAFFL")
    ),
    tm_table("Table No Choices", "ASL", "SEX", "RACE",
      pre_output = helpText("Titles"),
      post_output = helpText("Footnotes")
    )
  )
)

shinyApp(x$ui, x$server)
}
}
