% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_venn.R
\name{tm_venn2}
\alias{tm_venn2}
\title{venn2 teal module}
\usage{
tm_venn2(label, dataname, bm1, bm2, plot_height = c(600, 200, 2000),
  alpha = c(1, 0, 1), pre_output = tags$p("NAs get currently removed"),
  post_output = NULL)
}
\description{
venn2 teal module
}
