% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_html_table}
\alias{to_html_table}
\title{Create simple HTML table}
\usage{
to_html_table(x, colnames = TRUE, rownames = TRUE)
}
\arguments{
\item{x}{two dimensional object (data.frame, matrix, array)}

\item{colnames}{\code{logical} whether colnames should be displayed}

\item{rownames}{\code{logical} whether rownames should be displayed}
}
\value{
\code{shiny.tag} table object
}
\description{
Turns data.frame into simple non-styled HTML table
}
\examples{
teal.modules.general:::to_html_table(iris[1:5, ], rownames = FALSE)
}
