% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trend_line_stats}
\alias{trend_line_stats}
\title{returns a data frame that stores the trend line statistics of a polynomial line fitted to scatterplot}
\usage{
trend_line_stats(
  data_df,
  x_var,
  y_var,
  smoothing_degree,
  group_by_var = NULL,
  facet_by_var = NULL
)
}
\arguments{
\item{data_df}{(\code{data.frame}) the data where the x-axis and y-axis is host}

\item{x_var}{(\code{character}) the name of the column in `data_df` to be used as the x-axis}

\item{y_var}{(\code{character}) the name of the column in `data_df` to be used as the y-axis}

\item{smoothing_degree}{(\code{integer}) the degree of the polynomial line to be fitted}

\item{group_by_var}{optional, (\code{character}) the name of the column in `data_df` that is used for grouping.
This parameter is only used for determining whether a row in the output will have multiple groups or only one group.}

\item{facet_by_var}{optional, (\code{character}) a vector of the name(s) of the column(s) in `data_df`
to be used for filtering. Each row of the output will consists of trend line statistics for a single combination of
all the possible combinations of the matrix created by all of the unique values of columns in this parameter.}
}
\description{
returns a data frame that stores the trend line statistics of a polynomial line fitted to scatterplot
}
\examples{
\dontrun{
trend_line_stats(mpg, "displ", "hwy", 2)
trend_line_stats(mpg, "displ", "hwy", 2, group_by_var = "class", facet_by_var = "manufacturer")
}
}
