% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_data_table.R
\name{ui_page_data_table}
\alias{ui_page_data_table}
\title{ui function of Data Table Viewer}
\usage{
ui_page_data_table(id, datasets)
}
\arguments{
\item{id}{(\code{character}) UI id}

\item{datasets}{(\code{FilteredData}) object}
}
\description{
ui function of Data Table Viewer
}
