% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_basic_table_args}
\alias{validate_basic_table_args}
\title{Additional validation for \code{basic_table_args} argument}
\usage{
validate_basic_table_args(basic_table_args, table_names = NULL)
}
\description{
Additional validation for \code{basic_table_args} argument
}
