% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_ggplot2_args}
\alias{validate_ggplot2_args}
\title{Additional validation for \code{ggplot2_args} argument}
\usage{
validate_ggplot2_args(ggplot2_args, plot_names = NULL)
}
\arguments{
\item{ggplot2_args}{(`ggplot_args`) class list, generated by \code{ggplot2_args()} function.
Or list of such objects.}

\item{plot_names}{(`character`) vector of all plot names.
Defaults to NULL so not multi-plot in the module.}
}
\value{
nothing
}
\description{
A validation function.
}
