% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{validate_input}
\alias{validate_input}
\title{Validates the variable browser inputs}
\usage{
validate_input(input, plot_var, datasets)
}
\arguments{
\item{input}{(`session$input`) the shiny session input}

\item{plot_var}{(`list`) list of a data frame and an array of variable names}

\item{datasets}{(`FilteredData`) the datasets passed to the module}
}
\value{
`logical` TRUE if validations pass; a Shiny validation error otherwise
}
\description{
Validates the variable browser inputs
}
