% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varname_w_label}
\alias{varname_w_label}
\title{Get variable name with label}
\usage{
varname_w_label(var_name, dataset)
}
\arguments{
\item{var_name}{(\code{character}) Name of variable to extract labels from.}

\item{dataset}{(\code{dataset}) Name of analysis dataset.}
}
\value{
(\code{character}) String with variable name and label.
}
\description{
Get variable name with label
}
\examples{
\dontrun{
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

varname_w_label("AGE", ADSL)
}
}
