% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap string to specified width}
\usage{
wrap_text(text, wrap_width)
}
\arguments{
\item{text}{(\code{character}) Text to be wrapped.}

\item{wrap_width}{(\code{numeric}) Number of characters to wrap text to}
}
\value{
(\code{character}) Wrapped text.
}
\description{
Wrap string to specified width
}
\keyword{internal}
