% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{get_var_description}
\alias{get_var_description}
\title{Returns a short variable description.}
\usage{
get_var_description(data, dataset_name, var_name)
}
\arguments{
\item{data}{(\code{tdata}) the object containing the dataset}

\item{dataset_name}{(\code{character}) the name of the dataset containing the variable}

\item{var_name}{(\code{character}) the name of the variable}
}
\description{
The format of the variable description is:
\code{"<Long variable label> [<dataset name>.<variable name>]"}

Example: \code{"Study Identifier [ADSL.STUDYID]"}
}
\keyword{internal}
