% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tab_active_js}
\alias{is_tab_active_js}
\title{JavaScript expression to check if a tab is active}
\usage{
is_tab_active_js(id, name)
}
\arguments{
\item{id}{(\code{character(1)}) the id of the tab panel with tabs.}

\item{name}{(\code{character(1)}) the name of the tab.}
}
\value{
JavaScript expression to be used in \code{shiny::conditionalPanel()} to determine
if the specified tab is active.
}
\description{
JavaScript expression to check if a tab is active
}
\keyword{internal}
