% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{plot_var_summary}
\alias{plot_var_summary}
\title{Plot variable}
\usage{
plot_var_summary(
  var,
  var_lab,
  numeric_as_factor,
  display_density = is.numeric(var),
  remove_NA_hist = FALSE,
  outlier_definition,
  records_for_factor,
  ggplot2_args
)
}
\arguments{
\item{var}{vector of any type to be plotted. For numeric variables it produces histogram with
density line, for factors it creates frequency plot}

\item{var_lab}{text describing selected variable to be displayed on the plot}

\item{numeric_as_factor}{\code{logical} should the numeric variable be treated as a factor}

\item{display_density}{\code{logical} Should density estimation be displayed for numeric values?}

\item{remove_NA_hist}{\code{logical} Should \code{NA} values be removed for histogram of factor like variables.}

\item{outlier_definition}{If 0 no outliers are removed, otherwise
outliers (those more than outlier_definition*IQR below/above Q1/Q3 be removed)}

\item{records_for_factor}{\code{numeric} if the number of factor levels is >= than this value then
a graph of the factors isn't shown, only a list of values.}

\item{ggplot2_args}{(\code{ggplot2_args}) object created by \code{\link[teal.widgets:ggplot2_args]{teal.widgets::ggplot2_args()}}
with settings for the module plot.
The argument is merged with options variable \code{teal.ggplot2_args} and default module setup.

For more details see the vignette: \code{vignette("custom-ggplot2-arguments", package = "teal.widgets")}}
}
\value{
plot
}
\description{
Creates summary plot with statistics relevant to data type.
}
\keyword{internal}
