% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{plot_var_summary}
\alias{plot_var_summary}
\title{Plot variable}
\usage{
plot_var_summary(
  qenv,
  wrap_character = NULL,
  numeric_as_factor,
  display_density = FALSE,
  remove_NA_hist = FALSE,
  outlier_definition,
  records_for_factor,
  ggplot2_args
)
}
\arguments{
\item{qenv}{teal_data object were code should be evaluated.}

\item{wrap_character}{(\code{numeric}) number of characters at which to wrap text values of \code{var}}

\item{numeric_as_factor}{(\code{logical}) should the numeric variable be treated as a factor}

\item{display_density}{(\code{logical}) should density estimation be displayed for numeric values}

\item{remove_NA_hist}{(\code{logical}) should \code{NA} values be removed for histogram of factor like variables}

\item{outlier_definition}{if 0 no outliers are removed, otherwise
outliers (those more than outlier_definition*IQR below/above Q1/Q3 be removed)}

\item{records_for_factor}{(\code{numeric}) if the number of factor levels is >= than this value then
a graph of the factors isn't shown, only a list of values}

\item{ggplot2_args}{(\code{ggplot2_args}) object created by \code{\link[teal.widgets:ggplot2_args]{teal.widgets::ggplot2_args()}}
with settings for the module plot.
The argument is merged with options variable \code{teal.ggplot2_args} and default module setup.

For more details see the vignette: \code{vignette("custom-ggplot2-arguments", package = "teal.widgets")}}
}
\value{
plot
}
\description{
Creates summary plot with statistics relevant to data type.
}
\keyword{internal}
