% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADTTE}
\alias{rADTTE}
\title{Random Time to Event Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 67 columns.
}
\source{
internal
}
\usage{
rADTTE
}
\description{
Random Time to Event Analysis Dataset
}
\keyword{datasets}
\keyword{internal}
