% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{remove_outliers_from}
\alias{remove_outliers_from}
\title{Removes the outlier observation from an array}
\usage{
remove_outliers_from(var, outlier_definition)
}
\arguments{
\item{var}{(\code{numeric}) a numeric vector}

\item{outlier_definition}{(\code{numeric}) if \code{0} then no outliers are removed, otherwise
outliers (those more than \verb{outlier_definition*IQR below/above Q1/Q3}) are removed}
}
\value{
(\code{numeric}) vector without the outlier values
}
\description{
Removes the outlier observation from an array
}
\keyword{internal}
