% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{render_tab_table}
\alias{render_tab_table}
\title{Renders the table for a single dataset in the left-hand side tabset panel}
\usage{
render_tab_table(
  dataset_name,
  parent_dataname,
  output,
  data,
  input,
  columns_names,
  plot_var
)
}
\arguments{
\item{dataset_name}{(\code{character}) the name of the dataset}

\item{parent_dataname}{(\code{character}) the name of a parent \code{dataname} to filter out variables from}

\item{output}{(\code{session$output}) the shiny session output}

\item{data}{(\code{tdata}) the object containing all datasets}

\item{input}{(\code{session$input}) the shiny session input}

\item{columns_names}{(\code{environment}) the environment containing bindings for each dataset}

\item{plot_var}{(\code{list}) the list containing the currently selected dataset (tab) and its column names}
}
\description{
The table contains column names, column labels,
small summary about NA values and \code{sparkline} (if appropriate).
}
\keyword{internal}
