% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shared_params}
\alias{shared_params}
\title{Shared Parameters}
\arguments{
\item{plot_height}{optional, (\code{numeric}) A vector of length three with \code{c(value, min and max)}
for a slider encoding the plot height.}

\item{rotate_xaxis_labels}{optional, (\code{logical}) Whether to rotate plot X axis labels. Does not
rotate by default (\code{FALSE}).}
}
\description{
Contains arguments that are shared between multiple functions
  in the package to avoid repetition using \code{inheritParams}.
}
