% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shared_params}
\alias{shared_params}
\title{Shared Parameters}
\arguments{
\item{plot_height}{optional, (\code{numeric}) A vector of length three with \code{c(value, min and max)}
for a slider encoding the plot height.}

\item{plot_width}{optional, (\code{numeric}) A vector of length three with \code{c(value, min and max)}
for a slider encoding the plot width.}

\item{rotate_xaxis_labels}{optional, (\code{logical}) Whether to rotate plot X axis labels. Does not
rotate by default (\code{FALSE}).}

\item{ggtheme}{optional, (\code{character}) \code{ggplot} Theme to be used by default.
\code{gg_themes} is defined internally as
\code{c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void", "test")}
All themes can be chosen by the user. Defaults to \code{gray}.}

\item{ggplot2_args}{(`ggplot2_args`) object created by [`teal.devel::ggplot2_args()`]
with settings for the module table.
For more details see the help vignette:
`vignette("Custom ggplot2_args arguments module", package = "teal.devel")`
The argument is merged with options variable `teal.ggplot2_args` and default module setup.}
}
\description{
Contains arguments that are shared between multiple functions
  in the package to avoid repetition using \code{inheritParams}.
}
