% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.modules.general.R
\docType{package}
\name{teal.modules.general}
\alias{teal.modules.general}
\title{teal.modules.general: General modules to add to a teal application}
\description{
The modules in this package are generic modules that should work with any data set
(not necessarily for clinical trials data).
}
\keyword{internal}
