% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_a_regression.R
\name{tm_a_regression}
\alias{tm_a_regression}
\title{Scatterplot and Regression Model}
\usage{
tm_a_regression(
  label = "Regression Analysis",
  regressor,
  response,
  plot_height = c(600, 200, 2000),
  pre_output = NULL,
  post_output = NULL,
  default_plot_type = 1
)
}
\arguments{
\item{label}{(\code{character}) Label shown in the navigation item for the module.}

\item{regressor}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Regressor variables from an incoming dataset with filtering and selecting.}

\item{response}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Response variables from an incoming dataset with filtering and selecting.}

\item{plot_height}{optional, (\code{numeric}) A vector of length three with \code{c(value, min and max)}
for a slider encoding the plot height.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example a title.}

\item{post_output}{optional, \code{shiny.tag} with text placed after the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{default_plot_type}{optional, (\code{numeric}) Defaults to Response vs Regressor.
\itemize{
 \item{1 }{Response vs Regressor}
 \item{2 }{Residuals vs Fitted}
 \item{3 }{Normal Q-Q}
 \item{4 }{Scale-Location}
 \item{5 }{Cooks distance}
 \item{6 }{Residuals vs Leverage}
 \item{7 }{Cooks dist vs Leverage}
}}
}
\description{
Scatterplot and Regression Model
}
\note{
For more examples, please see the vignette "Using regression plots" via
  \code{vignette("using-regression-plots", package = "teal.modules.general")}.
}
\examples{
# Regression graphs from selected response variable (BMRKR1) and
# selected regressors (AGE)
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- radsl(cached = TRUE)",
    check = TRUE
  ),
  modules = root_modules(
    tm_a_regression(
      label = "Regression",
      response = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = "BMRKR1",
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = TRUE
        )
      ),
      regressor = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variables:",
          choices = variable_choices(ADSL, c("AGE", "SEX", "RACE")),
          selected = "AGE",
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
}
