% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_a_regression.R
\name{tm_a_regression}
\alias{tm_a_regression}
\title{\code{teal} module: Scatterplot and regression analysis}
\usage{
tm_a_regression(
  label = "Regression Analysis",
  regressor,
  response,
  plot_height = c(600, 200, 2000),
  plot_width = NULL,
  alpha = c(1, 0, 1),
  size = c(2, 1, 8),
  ggtheme = c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void"),
  ggplot2_args = teal.widgets::ggplot2_args(),
  pre_output = NULL,
  post_output = NULL,
  default_plot_type = 1,
  default_outlier_label = "USUBJID",
  label_segment_threshold = c(0.5, 0, 10),
  transformators = list(),
  decorators = list()
)
}
\arguments{
\item{regressor}{(\code{picks}) Specification for regressor variables selection.
Created using \code{\link[teal.transform:picks]{teal.transform::picks()}}, which allows selecting variables
to use as regressors in the regression model. \code{variables(multiple = TRUE)} allowed.}

\item{response}{(\code{picks}) Specification for response variable selection.
Created using \code{\link[teal.transform:picks]{teal.transform::picks()}}, which allows selecting a single numeric variable
to use as the response in the regression model. \code{variables(multiple = TRUE)} not allowed.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{plot_width}{(\code{numeric}) optional, specifies the plot width as a three-element vector of
\code{value}, \code{min}, and \code{max} for a slider encoding the plot width.}

\item{alpha}{(\code{integer(1)} or \code{integer(3)}) optional, specifies point opacity.
\itemize{
\item When the length of \code{alpha} is one: the plot points will have a fixed opacity.
\item When the length of \code{alpha} is three: the plot points opacity are dynamically adjusted based on
vector of \code{value}, \code{min}, and \code{max}.
}}

\item{size}{(\code{integer(1)} or \code{integer(3)}) optional, specifies point size.
\itemize{
\item When the length of \code{size} is one: the plot point sizes will have a fixed size.
\item When the length of \code{size} is three: the plot points size are dynamically adjusted based on
vector of \code{value}, \code{min}, and \code{max}.
}}

\item{ggtheme}{(\code{character}) optional, \code{ggplot2} theme to be used by default. Defaults to \code{"gray"}.}

\item{ggplot2_args}{(\code{ggplot2_args}) optional, object created by \code{\link[teal.widgets:ggplot2_args]{teal.widgets::ggplot2_args()}} with settings for all the plots or named list of \code{ggplot2_args} objects for plot-specific settings. The argument is merged with options variable \code{teal.ggplot2_args} and default module setup.

List names should match the following: \code{c("default", "Response vs Regressor", "Residuals vs Fitted", "Scale-Location", "Cook's distance", "Residuals vs Leverage", "Cook's dist vs Leverage")}.

For more details see the vignette: \code{vignette("custom-ggplot2-arguments", package = "teal.widgets")}.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{default_plot_type}{(\code{numeric}) optional, defaults to "Response vs Regressor".
\enumerate{
\item Response vs Regressor
\item Residuals vs Fitted
\item Normal Q-Q
\item Scale-Location
\item Cook's distance
\item Residuals vs Leverage
\item Cook's dist vs Leverage
}}

\item{label_segment_threshold}{(\code{numeric(1)} or \code{numeric(3)})
Minimum distance between label and point on the plot that triggers the creation of
a line segment between the two.
This may happen when the label cannot be placed next to the point as it overlaps another
label or point.
The value is used as the \code{min.segment.length} parameter to the \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}} function.

It can take the following forms:
\itemize{
\item \code{numeric(1)}: Fixed value used for the minimum distance and the slider is not presented in the UI.
\item \code{numeric(3)}: A slider is presented in the UI (under "Plot settings") to adjust the minimum distance dynamically.

It takes the form of \code{c(value, min, max)} and it is passed to the \code{value_min_max}
argument in \code{teal.widgets::optionalSliderInputValMinMax}.
}}

\item{decorators}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
(named \code{list} of lists of \code{teal_transform_module}) optional,
decorator for tables or plots included in the module output reported.
The decorators are applied to the respective output objects.}

\item{outlier}{(\code{picks}) Optional specification for outlier label variable selection.
Created using \code{\link[teal.transform:picks]{teal.transform::picks()}}, which allows selecting a factor or character variable
to label outlier points on the plots.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Module for visualizing regression analysis, including scatterplots and
various regression diagnostics plots.
It allows users to explore the relationship between a set of regressors and a response variable,
visualize residuals, and identify outliers.
}
\note{
For more examples, please see the vignette "Using regression plots" via
\code{vignette("using-regression-plots", package = "teal.modules.general")}.
}
\section{Decorating Module}{


This module generates the following objects, which can be modified in place using decorators:
\itemize{
\item \code{plot} (\code{ggplot})
}

A Decorator is applied to the specific output using a named list of \code{teal_transform_module} objects.
The name of this list corresponds to the name of the output to which the decorator is applied.
See code snippet below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tm_a_regression(
   ..., # arguments for module
   decorators = list(
     plot = teal_transform_module(...) # applied to the `plot` output
   )
)
}\if{html}{\out{</div>}}

For additional details and examples of decorators, refer to the vignette
\code{vignette("decorate-module-output", package = "teal.modules.general")}.

To learn more please refer to the vignette
\code{vignette("transform-module-output", package = "teal")} or the \code{\link[teal:teal_transform_module]{teal::teal_transform_module()}} documentation.
}

\examples{

# general data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  CO2 <- CO2
})

app <- init(
  data = data,
  modules = modules(
    tm_a_regression(
      label = "Regression",
      response = picks(
        datasets("CO2"),
        variables(choices = "uptake", selected = "uptake")
      ),
      regressor = picks(
        datasets("CO2"),
        variables(choices = c("conc", "Treatment"), selected = "conc", multiple = TRUE)
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# CDISC data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  ADSL <- teal.data::rADSL
})
join_keys(data) <- default_cdisc_join_keys[names(data)]

app <- init(
  data = data,
  modules = modules(
    tm_a_regression(
      label = "Regression",
      response = picks(
        datasets("ADSL"),
        variables(choices = "BMRKR1", selected = "BMRKR1")
      ),
      regressor = picks(
        datasets("ADSL"),
        variables(choices = c("AGE", "SEX", "RACE"), selected = "AGE", multiple = TRUE)
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
\section{Examples in Shinylive}{
\describe{
  \item{example-1}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokrcQAxLqkQ11XfxQpFC6cAAesKgiSoHBBsZc1AD6sVA2MUEhRroA7rSkABYq7Km4uiBKurqMcACOgrQ17L6ipMTURIyKEFUAwgDyAEzxugODSgC+3UpoqCMq+eyVAZm6ALwrwbjLfEIiouu6u8JiSz1VuqQwSVBJNRI1oqJWEGcXF9RQ9HD+GwpgWDgDzEzxI-2253ej1QJFEeg2qFoBAA1qI3u8Lqk4aQ0f8xv85BCMRctCxaF99uwCAUiEixId-oJUMFkXBwbo4SINHB+AywEyWWywN1iYTllCgY9RJ1DoiUbjIRisXAcUswPjhUTibpSYxyfRKdTaQR6RsCGriBZ2f87DUgvAyASypy4NzeX9CCQCOyYMJNFF4fYnM4RRjQ1VQ91urQTLp2CpyMxLDobLYKudREUIKwAILodizAAkDTKRbhjB0XUmSjAEwAukA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokrcQAxLqkQ11XfxQpFC6cAAesKgiSoHBBsZc1AD6sVA2MUEhRroA7rSkABYq7Km4uiBKurqMcACOgrQ17L6ipMTURIyKEFUAwgDyAEzxugODSgC+3UpoqCMq+eyVAZm6ALwrwbjLfEIiouu6u8JiSz1VuqQwSVBJNRI1oqJWEGcXF9RQ9HD+GwpgWDgDzEzxI-2253ej1QJFEeg2qFoBAA1qI3u8Lqk4aQ0f8xv85BCMRctCxaF99uwCAUiEixId-oJUMFkXBwbo4SINHB+AywEyWWywN1iYTllCgY9RJ1DoiUbjIRisXAcUswPjhUTibpSYxyfRKdTaQR6RsCGriBZ2f87DUgvAyASypy4NzeX9CCQCOyYMJNFF4fYnM4RRjQ1VQ91urQTLp2CpyMxLDobLYKudREUIKwAILodizAAkDTKRbhjB0XUmSjAEwAukA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
  \item{example-2}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXUY4AEdBWgb2CDFSYmoiRkUIOoBBHwCAGWTdVIxMxERGYbGlAF9+gCsiFTSAazhWUQrc23z+OBMoYVI0gn5aUQI09c2dveBoeH3MuQBdN2h0CZUxXYtRCuV0AF5QeFcCC+EIRKIIbo4cIxMCBnVJjA0lA0g0JA1RKIrBB0ZjMdQoPQ4NQkQowFg4ASxMSSPSYRjyYTUCRRHpIahaAQtvsQeS6pk+aRRWAFqN6XIOeLMVoWLQqQj2AQShsCGI6WAAEIAWSwAGksABGdm6PkiDRwfgGk3mq0KsWYxUeur4wmiXpIwXC0Wc8WSuDS4GykbysBe0Pk1WMdX0TXa3X6yEEKODADirjwunpAWcAA0bfSsIMvAXFbaaXAHU7IfS8wWqjALrQ4vz7E5nP1lYPPSD+v1aCZdOwVORmJYdDZbDUMaIyhBWIN0Ow0KgACQtKo73d8xg6PrLJRgJZfIA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXUY4AEdBWgb2CDFSYmoiRkUIOoBBHwCAGWTdVIxMxERGYbGlAF9+gCsiFTSAazhWUQrc23z+OBMoYVI0gn5aUQI09c2dveBoeH3MuQBdN2h0CZUxXYtRCuV0AF5QeFcCC+EIRKIIbo4cIxMCBnVJjA0lA0g0JA1RKIrBB0ZjMdQoPQ4NQkQowFg4ASxMSSPSYRjyYTUCRRHpIahaAQtvsQeS6pk+aRRWAFqN6XIOeLMVoWLQqQj2AQShsCGI6WAAEIAWSwAGksABGdm6PkiDRwfgGk3mq0KsWYxUeur4wmiXpIwXC0Wc8WSuDS4GykbysBe0Pk1WMdX0TXa3X6yEEKODADirjwunpAWcAA0bfSsIMvAXFbaaXAHU7IfS8wWqjALrQ4vz7E5nP1lYPPSD+v1aCZdOwVORmJYdDZbDUMaIyhBWIN0Ow0KgACQtKo73d8xg6PrLJRgJZfIA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
}
}

