% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_a_regression.R
\name{tm_a_regression}
\alias{tm_a_regression}
\title{\code{teal} module: Scatterplot and regression analysis}
\usage{
tm_a_regression(
  label = "Regression Analysis",
  regressor,
  response,
  plot_height = c(600, 200, 2000),
  plot_width = NULL,
  alpha = c(1, 0, 1),
  size = c(2, 1, 8),
  ggtheme = c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void"),
  ggplot2_args = teal.widgets::ggplot2_args(),
  pre_output = NULL,
  post_output = NULL,
  default_plot_type = 1,
  default_outlier_label = "USUBJID",
  label_segment_threshold = c(0.5, 0, 10)
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{regressor}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Regressor variables from an incoming dataset with filtering and selecting.}

\item{response}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Response variables from an incoming dataset with filtering and selecting.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{plot_width}{(\code{numeric}) optional, specifies the plot width as a three-element vector of
\code{value}, \code{min}, and \code{max} for a slider encoding the plot width.}

\item{alpha}{(\code{integer(1)} or \code{integer(3)}) optional, specifies point opacity.
\itemize{
\item When the length of \code{alpha} is one: the plot points will have a fixed opacity.
\item When the length of \code{alpha} is three: the plot points opacity are dynamically adjusted based on
vector of \code{value}, \code{min}, and \code{max}.
}}

\item{size}{(\code{integer(1)} or \code{integer(3)}) optional, specifies point size.
\itemize{
\item When the length of \code{size} is one: the plot point sizes will have a fixed size.
\item When the length of \code{size} is three: the plot points size are dynamically adjusted based on
vector of \code{value}, \code{min}, and \code{max}.
}}

\item{ggtheme}{(\code{character}) optional, \code{ggplot2} theme to be used by default. Defaults to \code{"gray"}.}

\item{ggplot2_args}{(\code{ggplot2_args}) optional, object created by \code{\link[teal.widgets:ggplot2_args]{teal.widgets::ggplot2_args()}}
with settings for all the plots or named list of \code{ggplot2_args} objects for plot-specific settings.
The argument is merged with options variable \code{teal.ggplot2_args} and default module setup.

List names should match the following: \verb{c("default", "Response vs Regressor", "Residuals vs Fitted", "Scale-Location", "Cook's distance", "Residuals vs Leverage"", "Cook's dist vs Leverage")}.

For more details see the vignette: \code{vignette("custom-ggplot2-arguments", package = "teal.widgets")}.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{default_plot_type}{(\code{numeric}) optional, defaults to "Response vs Regressor".
\enumerate{
\item Response vs Regressor
\item Residuals vs Fitted
\item Normal Q-Q
\item Scale-Location
\item Cook's distance
\item Residuals vs Leverage
\item Cook's dist vs Leverage
}}

\item{default_outlier_label}{(\code{character}) optional, default column selected to label outliers.}

\item{label_segment_threshold}{(\code{numeric(1)} or \code{numeric(3)})
Minimum distance between label and point on the plot that triggers the creation of
a line segment between the two.
This may happen when the label cannot be placed next to the point as it overlaps another
label or point.
The value is used as the \code{min.segment.length} parameter to the \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}} function.

It can take the following forms:
\itemize{
\item \code{numeric(1)}: Fixed value used for the minimum distance and the slider is not presented in the UI.
\item \code{numeric(3)}: A slider is presented in the UI (under "Plot settings") to adjust the minimum distance dynamically.

It takes the form of \code{c(value, min, max)} and it is passed to the \code{value_min_max}
argument in \code{teal.widgets::optionalSliderInputValMinMax}.
}}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Module for visualizing regression analysis, including scatterplots and
various regression diagnostics plots.
It allows users to explore the relationship between a set of regressors and a response variable,
visualize residuals, and identify outliers.
}
\note{
For more examples, please see the vignette "Using regression plots" via
\code{vignette("using-regression-plots", package = "teal.modules.general")}.
}
\examples{
# general data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  CO2 <- CO2
})
datanames(data) <- c("CO2")

app <- init(
  data = data,
  modules = modules(
    tm_a_regression(
      label = "Regression",
      response = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = "uptake",
          selected = "uptake",
          multiple = FALSE,
          fixed = TRUE
        )
      ),
      regressor = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variables:",
          choices = variable_choices(data[["CO2"]], c("conc", "Treatment")),
          selected = "conc",
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# CDISC data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  ADSL <- rADSL
})
datanames(data) <- "ADSL"
join_keys(data) <- default_cdisc_join_keys[datanames(data)]

app <- init(
  data = data,
  modules = modules(
    tm_a_regression(
      label = "Regression",
      response = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = "BMRKR1",
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = TRUE
        )
      ),
      regressor = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variables:",
          choices = variable_choices(data[["ADSL"]], c("AGE", "SEX", "RACE")),
          selected = "AGE",
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
\section{Examples in Shinylive}{
\itemize{
  \item example-1\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6uoxwAI6CtNXsPqKkxNREjIoQlQDCAPIATHG6-QNKAL5dKdDwosUZtlkE7Apgo6tdSmiowyp5K93+GboAvEdBuBW6fEIioqfXAsJiB5WVpDCJUInVEtWioisEFeb0q1Cg9DgfjOqywcD+YkBJFWl0OoP+qBIoj0ZxSiTCpGYGkSolQcGWV1BlWmsBxulW6zwlKp2JEGgerPJpBJZIpaKpYIhUIeqwAylCubotCxaBCRIgUcyBQR8kRaAQxCKwIJUEEANZwRX8gWcjRwfhanX6w1M41UmDCTSROkAMQAggAZUXOVEC0EmWihc0PBwuJVvLoCuS+qm-f6iDoPPEEonc0nkkECmnwLWMmMmiXss6mtO8zN+8GQ6H0sDitmkKUyuViBW2v1vFVqjX3M7Sxiy+giRKd9UvFLAYAMwarAC6M9KFMIJAIKJrdmqgXgZA20fDoJLwZhS4sRvblQd1CdIhDTh9e7eAaDFrO7q9rjtlUjVK-n6uXS6tAmLo7AqOQqbaHANi2OUhyiIUECsG66DsNsAAk9SlGh2KMDonQTEoYDjDOQA}{Open in Shinylive}\cr
    \if{html}{\out{<div style="height: 800px"><iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6uoxwAI6CtNXsPqKkxNREjIoQlQDCAPIATHG6-QNKAL5dKdDwosUZtlkE7Apgo6tdSmiowyp5K93+GboAvEdBuBW6fEIioqfXAsJiB5WVpDCJUInVEtWioisEFeb0q1Cg9DgfjOqywcD+YkBJFWl0OoP+qBIoj0ZxSiTCpGYGkSolQcGWV1BlWmsBxulW6zwlKp2JEGgerPJpBJZIpaKpYIhUIeqwAylCubotCxaBCRIgUcyBQR8kRaAQxCKwIJUEEANZwRX8gWcjRwfhanX6w1M41UmDCTSROkAMQAggAZUXOVEC0EmWihc0PBwuJVvLoCuS+qm-f6iDoPPEEonc0nkkECmnwLWMmMmiXss6mtO8zN+8GQ6H0sDitmkKUyuViBW2v1vFVqjX3M7Sxiy+giRKd9UvFLAYAMwarAC6M9KFMIJAIKJrdmqgXgZA20fDoJLwZhS4sRvblQd1CdIhDTh9e7eAaDFrO7q9rjtlUjVK-n6uXS6tAmLo7AqOQqbaHANi2OUhyiIUECsG66DsNsAAk9SlGh2KMDonQTEoYDjDOQA" height="125\%" width="125\%" allow="fullscreen" scrolling="auto" style="border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; -webkit-transform: scale(0.8); -webkit-transform-origin: 0 0"></iframe></div>}}
  \item example-2\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXUY4AEdBWgb2CDFSYmoiRkUIOoBBHwCAGWT64bGlAF9+zOh4UQrc23yFMCnRjaUAKyIVNIBrOFZlzLXjfjgTKGFSNIJ+WlECNP3Dk7PgBdgxFfCcgAum5oOgJipiuxaiFcroALyw8K4GF8IQiUQI3Ro4T-GF1UgwNJQNINCQNUSiKwQaEDOr06hQehwahYjZYODksRUkgbFF0+n1MSoEiiPSIzJpaKkZgaNKiVBwAi0wWC37wNmbEbbPD41VikQaLEGpUPBVKlWqwWM5msxEbAIs026LQsWhMkSIPl6q0EEoHAhiTUAIQAslgANJYACM3oFVt0Jo0cH4IfDUdjuvjVpg91ocXFugAYoNRgFnPyE-STLQoimsQ4XD76f0rXJK6qyRTRL0sZLpbKzYrlc26urCxstnGq0nSManXLzSPs6qbSzNY7DXPXYx3fQMV6s1W6n6A0HETu9yJHv7aIHzrlgMBJ9qNkCgVUR5sAOKuPC6B1nAADT5ACwCwQYvD-OR21HelZ3re0fz-Dsq1zahNALBsnArOC6hrOtU0REsy1cFc6lbVVKIomF+n6WgTF0dgVHIQdtDgGxbBqOlRDKCBWEGdB2DQVAABIWiqETRLFRgdD6WYlDAGYgSAA}{Open in Shinylive}\cr
    \if{html}{\out{<div style="height: 800px"><iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXUY4AEdBWgb2CDFSYmoiRkUIOoBBHwCAGWT64bGlAF9+zOh4UQrc23yFMCnRjaUAKyIVNIBrOFZlzLXjfjgTKGFSNIJ+WlECNP3Dk7PgBdgxFfCcgAum5oOgJipiuxaiFcroALyw8K4GF8IQiUQI3Ro4T-GF1UgwNJQNINCQNUSiKwQaEDOr06hQehwahYjZYODksRUkgbFF0+n1MSoEiiPSIzJpaKkZgaNKiVBwAi0wWC37wNmbEbbPD41VikQaLEGpUPBVKlWqwWM5msxEbAIs026LQsWhMkSIPl6q0EEoHAhiTUAIQAslgANJYACM3oFVt0Jo0cH4IfDUdjuvjVpg91ocXFugAYoNRgFnPyE-STLQoimsQ4XD76f0rXJK6qyRTRL0sZLpbKzYrlc26urCxstnGq0nSManXLzSPs6qbSzNY7DXPXYx3fQMV6s1W6n6A0HETu9yJHv7aIHzrlgMBJ9qNkCgVUR5sAOKuPC6B1nAADT5ACwCwQYvD-OR21HelZ3re0fz-Dsq1zahNALBsnArOC6hrOtU0REsy1cFc6lbVVKIomF+n6WgTF0dgVHIQdtDgGxbBqOlRDKCBWEGdB2DQVAABIWiqETRLFRgdD6WYlDAGYgSAA" height="125\%" width="125\%" allow="fullscreen" scrolling="auto" style="border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; -webkit-transform: scale(0.8); -webkit-transform-origin: 0 0"></iframe></div>}}
}
}

