% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_barplot.R
\name{tm_bep_safety0}
\alias{tm_bep_safety0}
\title{Bar plot for safety events}
\usage{
tm_bep_safety0(label, dataname, arm_var, arm_var_choices = arm_var,
  bep_var, bep_var_choices = bep_var, cov_var,
  cov_var_choices = cov_var, response_var,
  response_var_choices = response_var, responder_var,
  responder_var_choices = responder_var, non_responder_var,
  non_responder_var_choices = non_responder_var, exclude_var,
  exclude_var_choices = exclude_var, color_coding_var,
  color_coding_var_choices = color_coding_var, plot_height = c(1200,
  400, 5000), pre_output = NULL, post_output = NULL,
  code_data_processing = NULL)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{dataname}{(\code{character}) Dataset to be selected inside the app}

\item{arm_var}{(\code{character}) Name of the ARM var inside the dataset}

\item{arm_var_choices}{(\code{character}) Which variables can be used as the ARM var}

\item{bep_var}{(\code{character}) Which variables can be used as the BEP var}

\item{bep_var_choices}{(\code{character}) Which variables can be used as the BEP var}

\item{cov_var}{(\code{character}) Which variables can be used as the Covariance var}

\item{cov_var_choices}{(\code{character}) Which variables can be used as the Covariance var}

\item{response_var}{(\code{character}) Which variables can be used as the response var}

\item{response_var_choices}{(\code{character}) Which variables can be used as the response var}

\item{responder_var}{(\code{character}) Which variables can be used as the Responder var}

\item{responder_var_choices}{(\code{character}) Which variables can be used as the Responder var}

\item{non_responder_var}{(\code{character}) Which variables can be used as the non-Responder var}

\item{non_responder_var_choices}{(\code{character}) Which variables can be used as the non-Responder var}

\item{exclude_var}{(\code{character}) Which variables can be used as the Exclude var}

\item{exclude_var_choices}{(\code{character}) Which variables can be used as the Exclude var}

\item{color_coding_var}{(\code{character}) Which variables can be used as the Color Coding var}

\item{color_coding_var_choices}{(\code{character}) Which variables can be used as the Color Coding var}

\item{plot_height}{if scalar then the plot will have a fixed height. If a
slider should be presented to adjust the plot height dynamically then it
can be a vector of length three with vlaue, min and max.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{code_data_processing}{(\code{character}) Code that was used to pre-process the data}
}
\description{
Bar plot for safety events
}
\examples{
library(random.cdisc.data)

asl <- radsl(N = 600)
adte <- radaette(asl)

attr(asl, "source")  <- "random.cdisc.data::radsl(seed = 1)"
attr(adte, "source") <- "random.cdisc.data::radaette(asl, seed = 1)"

app <- teal::init(
  data = list(ASL = asl, ADTE = adte),
  modules = root_modules(
    tm_bep_safety0(
      label = "My label",
      dataname = "ADTE",
      arm_var=c("STUDYID"),
      bep_var = "",
      cov_var = "",
      response_var = c("AETTE1"),
      responder_var_choices = c("AETTE2","AETTE1"),
      responder_var = "",
      non_responder_var = "",
      exclude_var = "",
      color_coding_var = ""
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
