% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_data_table.R
\name{tm_data_table}
\alias{tm_data_table}
\title{Data Table Viewer Teal Module}
\usage{
tm_data_table(label = "Data table", variables_selected = list())
}
\arguments{
\item{label}{(\code{character})}

\item{variables_selected}{(\code{list}) a named list that says which variables should be
initially shown for particular dataset. Names in list should correspond with names provided in list `data()`.
If not specified for any dataset - first six variables from dataset will be shown.}
}
\description{
A data table viewer shows the data using a paginated table.
}
\examples{
library(random.cdisc.data)

ADSL <- cadsl

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- cadsl",
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table()
  )
)
\dontrun{
  shinyApp(app$ui, app$server)
}

# two-datasets example
library(random.cdisc.data)

ADSL <- cadsl
ADTTE <- cadaette


app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADTTE", ADTTE),
    code = "ADSL <- cadsl; ADTTE <- cadaette",
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table(
      variables_selected = list(ADSL  = c("SEX", "AGE","RACE"),
                                ADTTE = c("STUDYID","AGE")))
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: different subsets of long dataset

library(random.cdisc.data)

ADSL <- cadsl
ADLB <- cadlb

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADLB", ADLB),
    code = "ADSL <- cadsl; ADLB <- cadlb",
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table(
      variables_selected = list(ADSL  = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
                                ADLB = c("STUDYID", "USUBJID", "PARAM", "PARAMCD",
                                "AVISIT", "AVISITN", "AVAL", "CHG")))
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}

}
