% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_data_table.R
\name{tm_data_table}
\alias{tm_data_table}
\title{Data Table Viewer Teal Module}
\usage{
tm_data_table(label = "Data table", variables_selected = NULL)
}
\arguments{
\item{label}{(\code{character})}

\item{variables_selected}{(\code{list}) a named list that says which variables should be
initially  shown for which dataset}
}
\description{
A data table viewer shows the data using a paginated table.
}
\examples{


library(random.cdisc.data)

ASL <- radsl()
ADTE <- radaette(ASL)
attr(ASL, "source") <- "radsl(N = 600)"

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_data_table()
  )
)
\dontrun{
  shinyApp(x$ui, x$server)
}
}
