% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_data_table.R
\name{tm_data_table}
\alias{tm_data_table}
\title{Data Table Viewer Teal Module}
\usage{
tm_data_table(
  label = "Data table",
  variables_selected = list(),
  datasets_selected = character(0),
  dt_args = list(),
  dt_options = list(searching = FALSE, pageLength = 30, lengthMenu = c(5, 15, 30, 100),
    scrollX = TRUE)
)
}
\arguments{
\item{label}{(\code{character}) Label shown in the navigation item for the module.}

\item{variables_selected}{(\code{list}) A named list of character vectors of the variables (i.e. columns)
which should be initially shown for each dataset. Names of list elements should correspond to the names
of the datasets available in the app. If no entry is specified for a dataset, the first six variables from that
dataset will initially be shown.}

\item{datasets_selected}{(\code{character}) A vector of datasets which should be
shown and in what order. Names in the vector have to correspond with datasets names.
If vector of length zero (default) then all datasets are shown.}

\item{dt_args}{(named \code{list}) Additional arguments to be passed to \code{DT::datatable}
(must not include \code{data} or \code{options}).}

\item{dt_options}{(named \code{list}) The \code{options} argument to \code{DT::datatable}. By default
\code{list(searching = FALSE, pageLength = 30, lengthMenu = c(5, 15, 30, 100), scrollX = TRUE)}}
}
\description{
A data table viewer shows the data using a paginated table.
}
\examples{
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- radsl(cached = TRUE)"),
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table(
      variables_selected = list(ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX")),
      dt_args = list(caption = "ADSL Table Caption")
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# two-datasets example
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADTTE <- radtte(cached = TRUE)


app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- radsl(cached = TRUE)"),
    cdisc_dataset("ADTTE", ADTTE, code = "ADTTE <- radtte(cached = TRUE)"),
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table(
      variables_selected = list(
        ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
        ADTTE = c(
          "STUDYID", "USUBJID", "SUBJID", "SITEID",
          "PARAM", "PARAMCD", "ARM", "ARMCD", "AVAL", "CNSR"
        )
      ),
      dt_options = list(
        searching = FALSE,
        pageLength = 30,
        lengthMenu = c(5, 15, 30, 100),
        scrollX = TRUE
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: different subsets of long dataset
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADLB <- radlb(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- radsl(cached = TRUE)"),
    cdisc_dataset("ADLB", ADLB, code = "ADLB <- radlb(cached = TRUE)"),
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table(
      variables_selected = list(
        ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
        ADLB = c(
          "STUDYID", "USUBJID", "SUBJID", "SITEID",
          "PARAM", "PARAMCD", "AVISIT", "AVISITN", "AVAL", "CHG"
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
# datasets: subsetting or changing order of datasets inside tm_data_table
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADLB <- radlb(cached = TRUE)
ADTTE <- radtte(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- radsl(cached = TRUE)"),
    cdisc_dataset("ADLB", ADLB, code = "ADLB <- radlb(cached = TRUE)"),
    cdisc_dataset("ADTTE", ADTTE, code = "ADTTE <- radtte(cached = TRUE)"),
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table(
      variables_selected = list(
        ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
        ADLB = c(
          "STUDYID", "USUBJID", "SUBJID", "SITEID",
          "PARAM", "PARAMCD", "AVISIT", "AVISITN", "AVAL", "CHG"
        )
      ),
      datasets_selected = c("ADLB", "ADSL")
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

}
