% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_data_table.R
\name{tm_data_table}
\alias{tm_data_table}
\title{Data Table Viewer Teal Module}
\usage{
tm_data_table(
  label = "Data Table",
  variables_selected = list(),
  datasets_selected = character(0),
  dt_args = list(),
  dt_options = list(searching = FALSE, pageLength = 30, lengthMenu = c(5, 15, 30, 100),
    scrollX = TRUE),
  server_rendering = FALSE,
  pre_output = NULL,
  post_output = NULL
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module. Any label possible except
\code{"global_filters"} - read more in \code{mapping} argument of \link[teal]{teal_slices}.}

\item{variables_selected}{(\code{list}) A named list of character vectors of the variables (i.e. columns)
which should be initially shown for each dataset. Names of list elements should correspond to the names
of the datasets available in the app. If no entry is specified for a dataset, the first six variables from that
dataset will initially be shown.}

\item{datasets_selected}{(\code{character}) A vector of datasets which should be
shown and in what order. Names in the vector have to correspond with datasets names.
If vector of length zero (default) then all datasets are shown.}

\item{dt_args}{(named \code{list}) Additional arguments to be passed to \code{DT::datatable}
(must not include \code{data} or \code{options}).}

\item{dt_options}{(named \code{list}) The \code{options} argument to \code{DT::datatable}. By default
\code{list(searching = FALSE, pageLength = 30, lengthMenu = c(5, 15, 30, 100), scrollX = TRUE)}}

\item{server_rendering}{(\code{logical}) should the data table be rendered server side
(see \code{server} argument of \code{DT::renderDataTable()})}

\item{pre_output}{(\code{shiny.tag}, optional)\cr
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}, optional) with text placed after the output to put the output
into context. For example the \code{\link[shiny:helpText]{shiny::helpText()}} elements are useful.}
}
\description{
A data table viewer shows the data using a paginated table.
}
\details{
The \code{DT} package has an option \code{DT.TOJSON_ARGS} to show \code{Inf} and \code{NA} in data tables. If this is something
you require then set \code{options(DT.TOJSON_ARGS =  list(na = "string"))} before running the module.
Note though that sorting of numeric columns with \code{NA}/\code{Inf} will be lexicographic not numerical.
}
\examples{

ADSL <- teal.modules.general::rADSL

app <- teal::init(
  data = teal.data::cdisc_data(
    teal.data::cdisc_dataset("ADSL", ADSL, code = "ADSL <- teal.modules.general::rADSL"),
    check = TRUE
  ),
  modules = teal::modules(
    teal.modules.general::tm_data_table(
      variables_selected = list(ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX")),
      dt_args = list(caption = "ADSL Table Caption")
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
