% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_file_viewer.R
\name{tm_file_viewer}
\alias{tm_file_viewer}
\title{File Viewer Teal Module}
\usage{
tm_file_viewer(label = "File Viewer Module", input_path)
}
\arguments{
\item{label}{(\code{character}) Label shown in the navigation item for the module.}

\item{input_path}{(`list`) A list of the input paths to either: specific files of accepted formats,
a directory or a URL. The paths can be specified as absolute paths or relative to the running
directory of the application.}
}
\description{
The file viewer module provides a tool to upload and view static files.
Supported formats include text formats, \code{PDF}, \code{PNG}, \code{APNG},
\code{JPEG}, \code{SVG}, \code{WEBP}, \code{GIF} and \code{BMP}.
}
\examples{
data <- data.frame(1)
app <- init(
  data = teal_data(
    dataset("data", data)
  ),
  modules = root_modules(
    tm_file_viewer(
      input_path = list(
        folder = system.file("sample_files", package = "teal.modules.general"),
        pdf = system.file("sample_files/sample_file.pdf", package = "teal.modules.general"),
        png = system.file("sample_files/sample_file.png", package = "teal.modules.general"),
        txt = system.file("sample_files/sample_file.txt", package = "teal.modules.general"),
        url = "https://www.fda.gov/files/drugs/published/Portable-Document-Format-Specifications.pdf"
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

}
