% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_file_viewer.R
\name{tm_file_viewer}
\alias{tm_file_viewer}
\title{\code{teal} module: File viewer}
\usage{
tm_file_viewer(
  label = "File Viewer Module",
  input_path = list(`Current Working Directory` = ".")
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{input_path}{(\code{list}) of the input paths, optional. Each element can be:

Paths can be specified as absolute paths or relative to the running directory of the application.
Default to the current working directory if not supplied.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
The file viewer module provides a tool to view static files.
Supported formats include text formats, \code{PDF}, \code{PNG} \code{APNG},
\code{JPEG} \code{SVG}, \code{WEBP}, \code{GIF} and \code{BMP}.
}
\examples{
data <- teal_data()
data <- within(data, {
  data <- data.frame(1)
})
datanames(data) <- c("data")

app <- init(
  data = data,
  modules = modules(
    tm_file_viewer(
      input_path = list(
        folder = system.file("sample_files", package = "teal.modules.general"),
        png = system.file("sample_files/sample_file.png", package = "teal.modules.general"),
        txt = system.file("sample_files/sample_file.txt", package = "teal.modules.general"),
        url = "https://fda.gov/files/drugs/published/Portable-Document-Format-Specifications.pdf"
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
\section{Examples in Shinylive}{
\itemize{
\item example-1\cr
\href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqX8opKAeNdqAfQ8vG3dPbyNdAHdaUgALFXYgqFxdECVdXSSfTLCME2Z4dgBGRQgAX1Kk6HhRRLDbCIJ2BTAkltKlNFRslRjmiAysgF4cr1x03T4hEVFdEanhMX6MjNIYfxNaEX8tWjhImWWVjJVUQVJ-VE9Yud06UVIj44yTImp+GVvRVge4GDytnBmmBRLBUNtNjMWikrgQANZQKS3Fp+XgCRbiKQQNTUdrjAbPDKoCASL4-cj-SFAlqgmDguAbQGiAD0tPpjJEGGJEmhulhCKRIxRcG4aOmYkk0lkeImhNIAA9SGTfpTAcC2RCmaywZrOQrSLz+Yi9EKwKiFjNJdjpWA5PjCRlBIxqMiwLFSKRUKJEMzmSYPJIiFo-Vr+IxBBIWWd6PdYnB+MyAApERheGNwQwAESIBEE8DIhgAYimYJ5DABlVBwAi0TYETxWCDiVD8EwtWUrUrHLu6UqlWu6dgqcjMSw6Gy2NIE0TxCCsACC6HYXQAJIJaClV6IZDpGKUykowGUALpAA}{Open in Shinylive}\cr
\if{html}{
\out{<div class = "iframe-wrapper" style="height: 800px; width: 100\%"><script type="text/javascript">$(function() {  var if_pkgdown = [...document.scripts].filter(x => x.src.includes("pkgdown.js")).length > 0;  if (if_pkgdown) {    $(".iframe-wrapper").css("width", "140\%");  }});</script><iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqX8opKAeNdqAfQ8vG3dPbyNdAHdaUgALFXYgqFxdECVdXSSfTLCME2Z4dgBGRQgAX1Kk6HhRRLDbCIJ2BTAkltKlNFRslRjmiAysgF4cr1x03T4hEVFdEanhMX6MjNIYfxNaEX8tWjhImWWVjJVUQVJ-VE9Yud06UVIj44yTImp+GVvRVge4GDytnBmmBRLBUNtNjMWikrgQANZQKS3Fp+XgCRbiKQQNTUdrjAbPDKoCASL4-cj-SFAlqgmDguAbQGiAD0tPpjJEGGJEmhulhCKRIxRcG4aOmYkk0lkeImhNIAA9SGTfpTAcC2RCmaywZrOQrSLz+Yi9EKwKiFjNJdjpWA5PjCRlBIxqMiwLFSKRUKJEMzmSYPJIiFo-Vr+IxBBIWWd6PdYnB+MyAApERheGNwQwAESIBEE8DIhgAYimYJ5DABlVBwAi0TYETxWCDiVD8EwtWUrUrHLu6UqlWu6dgqcjMSw6Gy2NIE0TxCCsACC6HYXQAJIJaClV6IZDpGKUykowGUALpAA" scrolling="auto" style="height: 100\%; width: 100\%; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem"></iframe></div>}
}\cr
  }
}

