% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_front_page.R
\name{tm_front_page}
\alias{tm_front_page}
\title{Front page module}
\usage{
tm_front_page(
  label = "Front page",
  header_text = character(0),
  tables = list(),
  additional_tags = tagList(),
  footnotes = character(0),
  show_metadata = FALSE
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module. Any label possible except
\code{"global_filters"} - read more in \code{mapping} argument of \link[teal]{teal_slices}.}

\item{header_text}{\verb{character vector} text to be shown at the top of the module, for each
element, if named the name is shown first in bold as a header followed by the value. The first
element's header is displayed larger than the others}

\item{tables}{\verb{named list of dataframes} tables to be shown in the module}

\item{additional_tags}{\code{shiny.tag.list} or \code{html} additional shiny tags or \code{html} to be included after the table,
for example to include an image, \code{tagList(tags$img(src = "image.png"))} or to include further \code{html},
\code{HTML("html text here")}}

\item{footnotes}{\verb{character vector} text to be shown at the bottom of the module, for each
element, if named the name is shown first in bold, followed by the value}

\item{show_metadata}{\code{logical} should the metadata of the datasets be available on the module?}
}
\value{
A \code{teal} module to be used in \code{teal} applications
}
\description{
This \code{teal} module creates a simple front page for \code{teal} applications
}
\examples{

table_1 <- data.frame(Info = c("A", "B"), Text = c("A", "B"))
table_2 <- data.frame(`Column 1` = c("C", "D"), `Column 2` = c(5.5, 6.6), `Column 3` = c("A", "B"))
table_3 <- data.frame(Info = c("E", "F"), Text = c("G", "H"))

table_input <- list(
  "Table 1" = table_1,
  "Table 2" = table_2,
  "Table 3" = table_3
)

ADSL <- teal.modules.general::rADSL
app <- teal::init(
  data = teal.data::cdisc_data(
    teal.data::cdisc_dataset("ADSL", ADSL,
      code = "ADSL <- teal.modules.general::rADSL",
      metadata = list("Author" = "NEST team", "data_source" = "synthetic data")
    ),
    check = TRUE
  ),
  modules = teal::modules(
    teal.modules.general::tm_front_page(
      header_text = c(
        "Important information" = "It can go here.",
        "Other information" = "Can go here."
      ),
      tables = table_input,
      additional_tags = HTML("Additional HTML or shiny tags go here <br>"),
      footnotes = c("X" = "is the first footnote", "Y is the second footnote"),
      show_metadata = TRUE
    )
  ),
  header = tags$h1("Sample Application"),
  footer = tags$p("Application footer"),
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
