% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_front_page.R
\name{tm_front_page}
\alias{tm_front_page}
\title{\code{teal} module: Front page}
\usage{
tm_front_page(
  label = "Front page",
  header_text = character(0),
  tables = list(),
  additional_tags = tagList(),
  footnotes = character(0),
  show_metadata = FALSE
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{header_text}{(\code{character} vector) text to be shown at the top of the module, for each
element, if named the name is shown first in bold as a header followed by the value. The first
element's header is displayed larger than the others.}

\item{tables}{(\verb{named list} of \code{data.frame}s) tables to be shown in the module.}

\item{additional_tags}{(\code{shiny.tag.list} or \code{html}) additional \code{shiny} tags or \code{html} to be included after the table,
for example to include an image, \code{tagList(tags$img(src = "image.png"))} or to include further \code{html},
\code{HTML("html text here")}.}

\item{footnotes}{(\code{character} vector) of text to be shown at the bottom of the module, for each
element, if named the name is shown first in bold, followed by the value.}

\item{show_metadata}{(\code{logical}) indicating whether the metadata of the datasets be available on the module.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Creates a simple front page for \code{teal} applications, displaying
introductory text, tables, additional \code{html} or \code{shiny} tags, and footnotes.
}
\examples{
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  ADSL <- rADSL
  attr(ADSL, "metadata") <- list("Author" = "NEST team", "data_source" = "synthetic data")
})
datanames(data) <- "ADSL"
join_keys(data) <- default_cdisc_join_keys[datanames(data)]

table_1 <- data.frame(Info = c("A", "B"), Text = c("A", "B"))
table_2 <- data.frame(`Column 1` = c("C", "D"), `Column 2` = c(5.5, 6.6), `Column 3` = c("A", "B"))
table_3 <- data.frame(Info = c("E", "F"), Text = c("G", "H"))

table_input <- list(
  "Table 1" = table_1,
  "Table 2" = table_2,
  "Table 3" = table_3
)

app <- init(
  data = data,
  modules = modules(
    tm_front_page(
      header_text = c(
        "Important information" = "It can go here.",
        "Other information" = "Can go here."
      ),
      tables = table_input,
      additional_tags = HTML("Additional HTML or shiny tags go here <br>"),
      footnotes = c("X" = "is the first footnote", "Y is the second footnote"),
      show_metadata = TRUE
    )
  ),
  header = tags$h1("Sample Application"),
  footer = tags$p("Application footer"),
)

if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
\section{Examples in Shinylive}{
\itemize{
  \item example-1\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqX8opKAeNdqAfQ8vG3dPbyNdAHdaUgALFXYgqFxdECVdXUY4AEdBWiz2CDFSYmoiRkUIDIBBABEAZQAZH0y6pvTdT1JGdjbGlIUweC8kwdsIulFSdkHqwTjywd0AXl1BgDlnertdLhhBgbAk-1EiQUYCOCXVwdFWMli4TQJdUbBKgF9KpOh4UUSwuNjLMGo1BkoAFZEFT+ADWcFY-ySQNecBMUGEpH8BH4tFEBH8UJh8MRwB+sDEAK8cgAum4yFB6CJ-ABGFpJDAmZjwdgASQgJiIK10BBmYGqBzWYAAQmMUnY4AAPUjC0WzSWDWXvSpeJlwfwAJnZYU53Lg7AABgBhIjUQQwKosi2qsVWjVgWpy3TW232qoG52rUUAVgwwZSADYMBG5CkfXaHboAMyBkViiV4KVauQ6xnMpPGrymil8gVCoNi1yZwYAMS9CuVLsGAHF3QAJMaVJS65kqVDzFqTaYdQZ2PN6FnXXbj1m4EdgMd63QGqc9-UGudVKWLkTJ1czpNKLvQdAtFQxGZbpLCpKbjJ8IQiUTCh-CSkdDKkGD+LkkLGoKApEvDIQN0R4oH4GR-HIRsKw-UCpV5GBUHKLwyF0FRBUYGBPCsCAp0GXkVQIKAqgkIVHiyDADng0DBgAeTiGQMLLbDcJIAiwCtUjdHIsCZDgaiwFo3RYxEtdn1WNd-D7eY7wQiDcU0EhuGgwDJN0Ns7AAWUadN+CUvDuE0nTmnKXRRHiCBWGnCRnz4yi9H0JgAD45REwUiFICAvLEJswAADU4vFdkeUx8imUwiC8nzyHdABNDDnyYiy4GICB+CimLfPcrdQMsohIn8YYILCYUHBceDKgyMSt3AyDGGFLw7IAEliFkxXqWBUF3ap0DoEjlPw955M88hGqk9SWtQdN+toQa8Ky8bcuPWgTF0dgVHG9RNB0GxbDSLdLJUVg+pmtBUBavIUgulrRBkHQKiUD4lDAD4aSAA}{Open in Shinylive}\cr
    \if{html}{
      \out{
        <div class = "iframe-wrapper" style="height: 800px; width: 100\%">
          <script type="text/javascript">
$(function() {
  var if_pkgdown = [...document.scripts].filter(x => x.src.includes('pkgdown.js')).length > 0;
  if (if_pkgdown) {
    $('.iframe-wrapper').css('width', '140%');
  }
});
          </script>
          <iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqX8opKAeNdqAfQ8vG3dPbyNdAHdaUgALFXYgqFxdECVdXUY4AEdBWiz2CDFSYmoiRkUIDIBBABEAZQAZH0y6pvTdT1JGdjbGlIUweC8kwdsIulFSdkHqwTjywd0AXl1BgDlnertdLhhBgbAk-1EiQUYCOCXVwdFWMli4TQJdUbBKgF9KpOh4UUSwuNjLMGo1BkoAFZEFT+ADWcFY-ySQNecBMUGEpH8BH4tFEBH8UJh8MRwB+sDEAK8cgAum4yFB6CJ-ABGFpJDAmZjwdgASQgJiIK10BBmYGqBzWYAAQmMUnY4AAPUjC0WzSWDWXvSpeJlwfwAJnZYU53Lg7AABgBhIjUQQwKosi2qsVWjVgWpy3TW232qoG52rUUAVgwwZSADYMBG5CkfXaHboAMyBkViiV4KVauQ6xnMpPGrymil8gVCoNi1yZwYAMS9CuVLsGAHF3QAJMaVJS65kqVDzFqTaYdQZ2PN6FnXXbj1m4EdgMd63QGqc9-UGudVKWLkTJ1czpNKLvQdAtFQxGZbpLCpKbjJ8IQiUTCh-CSkdDKkGD+LkkLGoKApEvDIQN0R4oH4GR-HIRsKw-UCpV5GBUHKLwyF0FRBUYGBPCsCAp0GXkVQIKAqgkIVHiyDADng0DBgAeTiGQMLLbDcJIAiwCtUjdHIsCZDgaiwFo3RYxEtdn1WNd-D7eY7wQiDcU0EhuGgwDJN0Ns7AAWUadN+CUvDuE0nTmnKXRRHiCBWGnCRnz4yi9H0JgAD45REwUiFICAvLEJswAADU4vFdkeUx8imUwiC8nzyHdABNDDnyYiy4GICB+CimLfPcrdQMsohIn8YYILCYUHBceDKgyMSt3AyDGGFLw7IAEliFkxXqWBUF3ap0DoEjlPw955M88hGqk9SWtQdN+toQa8Ky8bcuPWgTF0dgVHG9RNB0GxbDSLdLJUVg+pmtBUBavIUgulrRBkHQKiUD4lDAD4aSAA" allow="fullscreen" scrolling="auto" style="height: 100\%; width: 100\%; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem"></iframe>
        </div>
      }
    }
}
}

