% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_association.R
\name{tm_g_association}
\alias{tm_g_association}
\title{Stack Plots of variables and show association with reference variable}
\usage{
tm_g_association(label = "Association", dataname, var,
  show_association = TRUE, plot_height = c(600, 400, 5000),
  pre_output = NULL, post_output = NULL, with_show_r_code = TRUE)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{dataname}{(\code{character}) data set name to analyze}

\item{var}{(\code{choices_selected}) TODO}

\item{show_association}{(\code{logical}) TODO}

\item{plot_height}{(\code{numeric}) vector with three elements defining selected, min and max plot height}

\item{with_show_r_code}{(\code{logical}) Whether show R Code button shall be enabled}
}
\description{
Stack Plots of variables and show association with reference variable
}
\examples{

library(random.cdisc.data)
asl <- radsl()

x <- teal::init(
  data = list(ASL = asl),
  modules = root_modules(
    tm_g_association(
      dataname = "ASL",
      var = choices_selected(names(asl), "AGE")
    )
  )
)

\dontrun{
shinyApp(x$ui, x$server)
}
}
