% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_association.R
\name{tm_g_association}
\alias{tm_g_association}
\title{Stack Plots of variables and show association with reference variable}
\usage{
tm_g_association(label = "Association", dataname, ref, vars,
  show_association = TRUE, plot_height = c(600, 400, 5000),
  pre_output = NULL, post_output = NULL, with_show_r_code = TRUE)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{dataname}{(\code{character}) data set name to analyze}

\item{ref}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
reference variable, must set \code{multiple = FALSE}}

\item{vars}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
associated variables}

\item{show_association}{(\code{logical}) wheater show association of \code{vars} with refference variable}

\item{plot_height}{(\code{numeric}) vector with three elements defining selected, min and max plot height}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{with_show_r_code}{(\code{logical}) Whether show R Code button shall be enabled}
}
\description{
Stack Plots of variables and show association with reference variable
}
\examples{

# datasets: single wide
library(random.cdisc.data)
ASL <- cadsl

app <- init(
  data = cdisc_data(
    ASL = ASL,
    code = "ASL <- cadsl",
    check = FALSE
  ),
  modules = root_modules(
    tm_g_association(
      dataname = "ASL",
      ref = data_extract_spec(
        dataname = "ASL",
        columns = columns_spec(label = "Reference variable",
                               choices = names(ASL),
                               selected = "AGE",
                               fixed = FALSE)
      ),
      vars = data_extract_spec(
        dataname = "ASL",
        columns = columns_spec(label = "Associated variables",
                               choices = names(ASL),
                               selected = "SEX",
                               multiple = TRUE,
                               fixed = FALSE)
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: different wide

library(random.cdisc.data)
library(tern)
library(dplyr)

ASL <- cadsl
ASL <- mutate_at(ASL,
                 .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
                 .funs = funs(as.factor(.))) \%>\% select("ARM", "ACTARM", "ACTARMCD",
 "SEX", "STRATA1", "AGE", "USUBJID", "STUDYID", "STRATA2")
keys(ASL) <- c("STUDYID", "USUBJID")


ADSL_2 <- mutate_at(cadsl,
                 .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
                 .funs = funs(as.factor(.))) \%>\% select("ACTARM", "AGE", "STRATA2", "COUNTRY", "USUBJID", "STUDYID")
keys(ADSL_2) <- c("STUDYID", "USUBJID")


app <- init(
  data = cdisc_data(
    ASL = ASL,
    ADSL_2 = ADSL_2,
    code = 'ASL <- cadsl
            ASL <- mutate_at(ASL,
                 .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
                 .funs = funs(as.factor(.))) \%>\% select("ARM", "ACTARM", "ACTARMCD",
                     "SEX", "STRATA1", "AGE", "USUBJID", "STUDYID", "STRATA2")
            ADSL_2 <- mutate_at(cadsl,
                 .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
                 .funs = funs(as.factor(.))) \%>\% select("ACTARM", "AGE", "STRATA2", "COUNTRY", "USUBJID", "STUDYID")
            keys(ASL) <- keys(ADSL_2) <- c("STUDYID", "USUBJID")',
    check = FALSE),
  modules = root_modules(
    tm_g_association(
      label = "Association plots",
      dataname = c("ASL", "ADSL_2"),
      ref = data_extract_spec(
        dataname = "ASL",
        columns = columns_spec(
         label = "Select variable",
          choices = c("AGE", "SEX", "STRATA1", "RACE"),
          selected = c("AGE"),
          multiple = FALSE
        )),
      vars = data_extract_spec(
        dataname = "ADSL_2",
        columns = columns_spec(
          label = "Select variables",
          choices = c("COUNTRY", "AGE", "RACE"),
          selected = c("AGE", "COUNTRY", "RACE"),
          multiple = TRUE
        ))
    ) #tm_g_regression
  )# root_modules
)# init
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: wide and long

library(random.cdisc.data)
library(tern)

ASL <- cadsl
ADRS <- cadrs

keys(ASL) <- c("STUDYID", "USUBJID")
keys(ADRS) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")

app <- init(
 data = cdisc_data(
   ASL = ASL,
   ADRS = ADRS,
   code = 'ASL <- cadsl
           ADRS <- cadrs
           keys(ASL) <- c("STUDYID", "USUBJID")
           keys(ADRS) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")',
   check = FALSE),
 modules = root_modules(
   tm_g_association(
     label = "Association Plots",
     dataname = c("ASL", "ADRS"),
     ref = data_extract_spec(
       dataname = "ADRS",
       filter = list(
         filter_spec(
           vars = "PARAM",
          choices = unique(ADRS$PARAM),
           selected = unique(ADRS$PARAM)[1],
           multiple = FALSE,
           label = "Choose response"
         ),
         filter_spec(
           vars = "AVISIT",
           choices = levels(ADRS$AVISIT),
           selected = levels(ADRS$AVISIT)[1],
           multiple = FALSE,
           label = "Choose visit"
         )
       ),
       columns = columns_spec(
         choices = "AVAL",
         selected = "AVAL",
         multiple = FALSE,
         label = "variable"
       )
    ),
     vars = data_extract_spec(
       dataname = "ASL",
       columns = columns_spec(
         choices = c("SEX", "AGE", "RACE", "COUNTRY"),
         selected = c("SEX", "AGE"),
         multiple = TRUE,
         fixed = FALSE
      )
     )
    ) #tm_g_association
  )# root_modules
)# init

\dontrun{
shinyApp(app$ui, app$server)
}

}
