% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_association.R
\name{tm_g_association}
\alias{tm_g_association}
\title{Stack Plots of variables and show association with reference variable}
\usage{
tm_g_association(label = "Association", dataname, ref, vars,
  show_association = TRUE, plot_height = c(600, 400, 5000),
  pre_output = NULL, post_output = NULL, with_show_r_code = TRUE)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{dataname}{(\code{character}) data set name to analyze}

\item{ref}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
reference variable, must set \code{multiple = FALSE}}

\item{vars}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
associated variables}

\item{show_association}{(\code{logical}) wheater show association of \code{vars} with refference variable}

\item{plot_height}{(\code{numeric}) vector with three elements defining selected, min and max plot height}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{with_show_r_code}{(\code{logical}) Whether show R Code button shall be enabled}
}
\description{
Stack Plots of variables and show association with reference variable
}
\examples{

# datasets: single wide
library(random.cdisc.data)
ASL <- cadsl

app <- init(
  data = cdisc_data(
    ASL = ASL,
    code = "ASL <- cadsl",
    check = FALSE
  ),
  modules = root_modules(
    tm_g_association(
      dataname = "ASL",
      ref = data_extract_spec(
        dataname = "ASL",
        columns = columns_spec(
          label = "Reference variable",
          choices = names(ASL),
          selected = "AGE",
          fixed = FALSE
        )
      ),
      vars = data_extract_spec(
        dataname = "ASL",
        columns = columns_spec(
          label = "Associated variables",
          choices = names(ASL),
          selected = "BMRKR1",
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: different wide

library(random.cdisc.data)
library(tern)
library(dplyr)

ASL <- cadsl
ASL <- mutate_at(ASL,
                 .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
                 .funs = funs(as.factor(.))) \%>\% select("ARM", "ACTARM", "ACTARMCD",
 "SEX", "STRATA1", "AGE", "USUBJID", "STUDYID", "STRATA2")
keys(ASL) <- c("STUDYID", "USUBJID")


ADSL_2 <- mutate_at(cadsl,
                 .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
                 .funs = funs(as.factor(.))) \%>\% select("ACTARM", "AGE", "STRATA2", "COUNTRY", "USUBJID", "STUDYID")
keys(ADSL_2) <- c("STUDYID", "USUBJID")

app <- init(
  data = cdisc_data(
    ASL = ASL,
    ADSL_2 = ADSL_2,
    code = 'ASL <- cadsl
            keys(ASL) <- c("STUDYID", "USUBJID")',
    check = FALSE),
  modules = root_modules(
    tm_g_association(
      label = "Regression",
      dataname = c("ASL", "ADSL_2"),
      ref = data_extract_spec(
        dataname = "ASL",
        columns = columns_spec(
          label = "Select variable",
          choices = c("AGE", "SEX", "STRATA1", "RACE"),
          selected = c("STRATA1"),
          multiple = FALSE
        )),
      vars = data_extract_spec(
        dataname = "ADSL_2",
        columns = columns_spec(
          label = "Select variables",
          choices = c("COUNTRY", "AGE", "RACE"),
          selected = c("AGE", "COUNTRY", "RACE"),
          multiple = TRUE
        ))
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: multiple long datasets
library(random.cdisc.data)

ASL <- cadsl
ADRS <- cadrs
ADTTE <- cadtte

app <- init(
  data = cdisc_data(
    ASL = ASL,
    ADRS = ADRS,
    ADTTE = ADTTE,
    code = "ASL <- cadsl; ADRS <- cadrs; ADTTE <- cadtte",
    check = FALSE
  ),
  modules = root_modules(
    tm_g_association(
      dataname = c("ASL", "ADRS", "ADTTE"),
      ref = list(
        data_extract_spec(
          dataname = "ADRS",
          columns = columns_spec(
            label = "Reference variable",
            choices = c("AGE", "SEX"),
            selected = c("AGE"),
            fixed = FALSE
          )
        ),
        data_extract_spec(
          dataname = "ADTTE",
          columns = columns_spec(
            label = "Reference variable",
            choices = c("AVAL", "CNSR"),
            selected = "AVAL",
            fixed = FALSE
          )
        )
      ),
      vars = list(
        data_extract_spec(
          dataname = "ADRS",
          columns = columns_spec(
            label = "Associated variables",
            choices = names(ADRS),
            selected = c("AVAL", "AVALC"),
            fixed = FALSE
          ),
          filter = filter_spec(
            vars = c("PARAMCD"),
            choices = unique(ADTTE$PARAMCD),
            selected = "OS",
            multiple = FALSE,
            label = "ADTTE filter"
          )
        ),
        data_extract_spec(
          dataname = "ADTTE",
          filter = filter_spec(
            vars = c("PARAMCD", "AVISIT"),
            choices = apply(expand.grid(levels(ADRS$PARAMCD), levels(ADRS$AVISIT)), 1, paste, collapse = " - "),
            selected = "OVRINV - Screening",
            multiple = TRUE,
            label = "ADRS filter"
          ),
          columns = columns_spec(
            label = "Associated variables",
            choices = names(ADTTE),
            selected = NULL,
            fixed = FALSE
          )
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: wide and long

library(random.cdisc.data)
library(tern)

ASL <- cadsl
ADRS <- cadrs

keys(ASL) <- c("STUDYID", "USUBJID")
keys(ADRS) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")

app <- init(
 data = cdisc_data(
   ASL = ASL,
   ADRS = ADRS,
   code = 'ASL <- cadsl
           ADRS <- cadrs
           keys(ASL) <- c("STUDYID", "USUBJID")
           keys(ADRS) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")',
   check = FALSE),
 modules = root_modules(
   tm_g_association(
     label = "Association Plots",
     dataname = c("ASL", "ADRS"),
     ref = data_extract_spec(
       dataname = "ASL",
       columns = columns_spec(
         choices = c("SEX", "AGE", "RACE", "COUNTRY"),
         selected = c("AGE"),
         multiple = FALSE,
         fixed = FALSE
      )
     ),
     vars = data_extract_spec(
       dataname = "ADRS",
       filter = list(
         filter_spec(
           vars = "PARAM",
           choices = levels(ADRS$PARAM),
           selected = levels(ADRS$PARAM)[1],
           multiple = FALSE,
           label = "Choose response"
         ),
         filter_spec(
           vars = "AVISIT",
           choices = levels(ADRS$AVISIT),
           selected = levels(ADRS$AVISIT)[1],
           multiple = FALSE,
           label = "Choose visit"
         )
       ),
       columns = columns_spec(
         choices = "AVAL",
         selected = "AVAL",
         multiple = FALSE,
         label = "variable"
       )
      )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: same long

library(random.cdisc.data)
library(tern)

ASL <- cadsl
ADRS <- cadrs

keys(ASL) <- c("STUDYID", "USUBJID")
keys(ADRS) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")

app <- init(
 data = cdisc_data(
   ASL = ASL,
   ADRS = ADRS,
   code = 'ASL <- cadsl
           ADRS <- cadrs
           keys(ASL) <- c("STUDYID", "USUBJID")
           keys(ADRS) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")',
   check = FALSE),
 modules = root_modules(
   tm_g_association(
     label = "Association Plots",
     dataname = "ADRS",
     ref = data_extract_spec(
       dataname = "ADRS",
       columns = columns_spec(
         choices = names(ADRS),
         selected = "AVAL",
         multiple = FALSE,
         fixed = FALSE
      )
     ),
     vars = data_extract_spec(
       dataname = "ADRS",
       columns = columns_spec(
         choices = names(ADRS),
         selected = "PARAMCD",
         multiple = FALSE,
         label = "variable"
       )
      )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}


# datasets: different subsets of long dataset

library(random.cdisc.data)
library(tern)

ASL <- cadsl
ADLB <- cadlb

keys(ASL) <- c("STUDYID", "USUBJID")
keys(ADLB) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")

app <- init(
  data = cdisc_data(
    ASL = ASL,
    ADLB = ADLB,
    code = 'ASL <- cadsl
           ADLB <- cadlb
           keys(ASL) <- c("STUDYID", "USUBJID")
           keys(ADLB) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")',
    check = FALSE
  ),
  modules = root_modules(
    tm_g_association(
      dataname = "ADLB",
      ref = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[1],
            multiple = FALSE,
            label = "Choose measurement"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Choose visit"
          )
        ),
        columns = columns_spec(
          choices = names(ADLB),
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE,
          label = "Reference variable"
        )
      ),
      vars = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[1],
            multiple = FALSE,
            label = "Choose measurement"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Choose visit"
          )
        ),
        columns = columns_spec(
          choices = names(ADLB),
          selected = "ARMCD",
          multiple = TRUE,
          fixed = FALSE,
          label = "Associated variables"
        )
      )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
