% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_association.R
\name{tm_g_association}
\alias{tm_g_association}
\title{\code{teal} module: Stack plots of variables and show association with reference variable}
\usage{
tm_g_association(
  label = "Association",
  ref = picks(datasets(), variables(choices = tidyselect::where(is.numeric) |
    teal.transform::is_categorical(min.len = 2, max.len = 10), selected = 1),
    values(selected = tidyselect::everything(), multiple = TRUE)),
  vars = picks(datasets(), variables(choices = tidyselect::where(is.numeric) |
    teal.transform::is_categorical(min.len = 2, max.len = 10), selected = 2, multiple =
    TRUE), values()),
  show_association = TRUE,
  plot_height = c(600, 400, 5000),
  plot_width = NULL,
  distribution_theme = c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic",
    "void"),
  association_theme = c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic",
    "void"),
  pre_output = NULL,
  post_output = NULL,
  ggplot2_args = teal.widgets::ggplot2_args(),
  transformators = list(),
  decorators = list()
)
}
\arguments{
\item{ref}{(\code{picks})
Reference variable specification created using \code{picks()}.}

\item{vars}{(\code{picks})
Variables to be associated with the reference variable, specified using \code{picks()}.}

\item{show_association}{(\code{logical}) optional, whether show association of \code{vars}
with reference variable. Defaults to \code{TRUE}.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{plot_width}{(\code{numeric}) optional, specifies the plot width as a three-element vector of
\code{value}, \code{min}, and \code{max} for a slider encoding the plot width.}

\item{distribution_theme, association_theme}{(\code{character}) optional, \code{ggplot2} themes to be used by default.
Default to \code{"gray"}.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{ggplot2_args}{(\code{ggplot2_args}) optional, object created by \code{\link[teal.widgets:ggplot2_args]{teal.widgets::ggplot2_args()}} with settings for all the plots or named list of \code{ggplot2_args} objects for plot-specific settings. The argument is merged with options variable \code{teal.ggplot2_args} and default module setup.

List names should match the following: \code{c("default", "Bivariate1", "Bivariate2")}.

For more details see the vignette: \code{vignette("custom-ggplot2-arguments", package = "teal.widgets")}.}

\item{decorators}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
(named \code{list} of lists of \code{teal_transform_module}) optional,
decorator for tables or plots included in the module output reported.
The decorators are applied to the respective output objects.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Module provides functionality for visualizing the distribution of variables and
their association with a reference variable.
It supports configuring the appearance of the plots, including themes and whether to show associations.
}
\note{
For more examples, please see the vignette "Using association plot" via
\code{vignette("using-association-plot", package = "teal.modules.general")}.
}
\section{Decorating Module}{


This module generates the following objects, which can be modified in place using decorators:
\itemize{
\item \code{plot} (\code{grob} created with \code{\link[ggplot2:ggplotGrob]{ggplot2::ggplotGrob()}})
}

A Decorator is applied to the specific output using a named list of \code{teal_transform_module} objects.
The name of this list corresponds to the name of the output to which the decorator is applied.
See code snippet below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tm_g_association(
   ..., # arguments for module
   decorators = list(
     plot = teal_transform_module(...) # applied to the `plot` output
   )
)
}\if{html}{\out{</div>}}

For additional details and examples of decorators, refer to the vignette
\code{vignette("decorate-module-output", package = "teal.modules.general")}.

To learn more please refer to the vignette
\code{vignette("transform-module-output", package = "teal")} or the \code{\link[teal:teal_transform_module]{teal::teal_transform_module()}} documentation.
}

\examples{
# general data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  CO2 <- CO2
  factors <- names(Filter(isTRUE, vapply(CO2, is.factor, logical(1L))))
  CO2[factors] <- lapply(CO2[factors], as.character)
})

app <- init(
  data = data,
  modules = modules(
    tm_g_association(
      ref = picks(
        datasets("CO2"),
        variables(
          choices = c("Plant", "Type", "Treatment"),
          selected = "Plant"
        )
      ),
      vars = picks(
        datasets("CO2"),
        variables(
          choices = c("Plant", "Type", "Treatment"),
          selected = "Treatment",
          multiple = TRUE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# CDISC data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  ADSL <- teal.data::rADSL
})
join_keys(data) <- default_cdisc_join_keys[names(data)]

app <- init(
  data = data,
  modules = modules(
    tm_g_association(
      ref = picks(
        datasets("ADSL"),
        variables(
          choices = c("SEX", "RACE", "COUNTRY", "ARM", "STRATA1", "STRATA2", "ITTFL", "BMRKR2"),
          selected = "RACE"
        )
      ),
      vars = picks(
        datasets("ADSL"),
        variables(
          choices = c("SEX", "RACE", "COUNTRY", "ARM", "STRATA1", "STRATA2", "ITTFL", "BMRKR2"),
          selected = "BMRKR2",
          multiple = TRUE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
\section{Examples in Shinylive}{
\describe{
  \item{example-1}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6uoxwAI6CtNXsPqKkxNREjIoQlQDCAPIATHG6-QMVpuqkHaLD0PCi7ABitNTkjOy0og4upVpokRyjpZsYJpMdpe0StATc7ACMADJyL129g8BnGtMAusPU+2ohw+XymjFEP1KUHEBHyLEmMi6AF8ukp9sMVHl2OMUroALz+DK4cZ8IQiGYE0nCMTY7qVXSkGCJCSJaGiIgEWiBKwQWn0+nVEz43SoG4AawW435lRSojgpElYFGCjAcmJdOluj2jC59HJfM19NhRBuYmFBGxYAACgCyCrSiq7KxUHB7bpHdVAvA7ar1Yb6XKRBo4PxhSqbVAfVLpW8Y37pdqKSLxZKNdLZfLFcrfdH+drdfrc9Ljaakxbw7bSG7Hc7XXh3WA7J7GRQqzm05rA3Bg6GCR64F7W-ai-yYMJNJE9ATtq4O-TY-yF7pY10urQhRtVMxLDobLZynTRIUIKwAILodj7AAk9Sh6CvcsYOk6SiRSjASJ+QA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6uoxwAI6CtNXsPqKkxNREjIoQlQDCAPIATHG6-QMVpuqkHaLD0PCi7ABitNTkjOy0og4upVpokRyjpZsYJpMdpe0StATc7ACMADJyL129g8BnGtMAusPU+2ohw+XymjFEP1KUHEBHyLEmMi6AF8ukp9sMVHl2OMUroALz+DK4cZ8IQiGYE0nCMTY7qVXSkGCJCSJaGiIgEWiBKwQWn0+nVEz43SoG4AawW435lRSojgpElYFGCjAcmJdOluj2jC59HJfM19NhRBuYmFBGxYAACgCyCrSiq7KxUHB7bpHdVAvA7ar1Yb6XKRBo4PxhSqbVAfVLpW8Y37pdqKSLxZKNdLZfLFcrfdH+drdfrc9Ljaakxbw7bSG7Hc7XXh3WA7J7GRQqzm05rA3Bg6GCR64F7W-ai-yYMJNJE9ATtq4O-TY-yF7pY10urQhRtVMxLDobLZynTRIUIKwAILodj7AAk9Sh6CvcsYOk6SiRSjASJ+QA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
  \item{example-2}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXUY4AEdBWgb2CDFSYmoiRkUIOoBBHwCAGWTdVIxMxERGYbGlAF9+gCsiFTSAazhWUQrc23z+OBMoYVI0gn5aUQI09c2dveBoeH3MuQBdN2h0CZUxXYtRCuV0AF5QeFcCC+EIRKIIbo4cIxMCBnVJjA0hI0lBRKIiARaGErBB0ZjMQ0TEjULQCFt9iDKXVMqI4KQmWAFqMFGA5DCMSzdFoWCT6AiKcLMQQShsCGIkQRgWAAs4ABp8qp8rCDLyuPC6PleADyjgAcg4AJpao3crAAWVtfICDkGdkGAEZnaq3R6AEw+vx2OwAMV5hr5ACEHVgANJYQP8wXSzHskQaOD8JE6vUG5ks-rCgUFuqixiIyF0hlMoUstkcrk8vkluuU8viyWllmy+WKyHKl0an26-U+00W60+waOn2u3Ue72R30LwZJ7VgYNhiMbmPxxMtlOp3TpuCZ7OQ6OxhPr7uUmAXWhxPSQhwuO+6IuFgtf-r9Wg0uwKjkMwlg6DYtg1BiohlBArCDOg7BoKgAAkLRVMhKHsowOh9MsShgEsXxAA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXUY4AEdBWgb2CDFSYmoiRkUIOoBBHwCAGWTdVIxMxERGYbGlAF9+gCsiFTSAazhWUQrc23z+OBMoYVI0gn5aUQI09c2dveBoeH3MuQBdN2h0CZUxXYtRCuV0AF5QeFcCC+EIRKIIbo4cIxMCBnVJjA0hI0lBRKIiARaGErBB0ZjMQ0TEjULQCFt9iDKXVMqI4KQmWAFqMFGA5DCMSzdFoWCT6AiKcLMQQShsCGIkQRgWAAs4ABp8qp8rCDLyuPC6PleADyjgAcg4AJpao3crAAWVtfICDkGdkGAEZnaq3R6AEw+vx2OwAMV5hr5ACEHVgANJYQP8wXSzHskQaOD8JE6vUG5ks-rCgUFuqixiIyF0hlMoUstkcrk8vkluuU8viyWllmy+WKyHKl0an26-U+00W60+waOn2u3Ue72R30LwZJ7VgYNhiMbmPxxMtlOp3TpuCZ7OQ6OxhPr7uUmAXWhxPSQhwuO+6IuFgtf-r9Wg0uwKjkMwlg6DYtg1BiohlBArCDOg7BoKgAAkLRVMhKHsowOh9MsShgEsXxAA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
}
}

