% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_bivariate.R
\name{tm_g_bivariate}
\alias{tm_g_bivariate}
\title{Cross table based on rtables}
\usage{
tm_g_bivariate(label = "Bivariate Plots", dataname, x_var,
  x_var_choices = x_var, y_var, y_var_choices = y_var)
}
\arguments{
\item{x_var}{variable name selected variable when shiny app starts}

\item{x_var_choices}{vector with variable names available as choices}

\item{y_var}{variable name selected variable when shiny app starts}

\item{y_var_choices}{vector with variable names available as choices}
}
\description{
Cross table based on rtables
}
\examples{

\dontrun{

N <- 100
ASL <- data.frame(
  USUBJID = paste("id", seq_len(N), sep ="-"),
  STUDYID = "study1",
  cont = rnorm(N),
  disc = factor(sample(letters[1:5], N, TRUE)),
  cont2 = runif(N),
  disc2 = factor(sample(LETTERS[1:5], N, TRUE))
)

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_g_bivariate(
      dataname = "ASL",
      x_var = "cont",
      x_var_choices = names(ASL),
      y_var = "cont2",
      y_var_choices = names(ASL)
    )
  )
)

shinyApp(x$ui, x$server)
}

}
