% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_bivariate.R
\name{tm_g_bivariate}
\alias{tm_g_bivariate}
\title{Univariate and bivariate visualizations.}
\usage{
tm_g_bivariate(label = "Bivariate Plots", dataname, x_var,
  x_var_choices = x_var, y_var, y_var_choices = y_var,
  use_density = FALSE, color_by_var, color_by_var_choices, row_facet_var,
  row_facet_var_choices, col_facet_var, col_facet_var_choices,
  free_x_scales = FALSE, free_y_scales = FALSE, plot_height = c(600,
  200, 2000), pre_output = NULL, post_output = NULL,
  code_data_processing = NULL)
}
\arguments{
\item{x_var}{variable name selected to plot along the x-axis by default. Variable can be numeric, factor or character.}

\item{x_var_choices}{character vector with variable names available as choices to plot along x-axis}

\item{y_var}{variable name selected to plot along the y-axis by default. Variable can be numeric, factor or character.}

\item{y_var_choices}{character vector with variable names available as choices to plot along y-axis}

\item{use_density}{boolean value for whether density is plotted}

\item{row_facet_var}{variable for x facets}

\item{row_facet_var_choices}{vector with \code{row_facet_var} choices}

\item{col_facet_var}{variable for y facets}

\item{col_facet_var_choices}{vector with \code{col_facet_var} choices}

\item{plot_height}{range of plot height}
}
\description{
Univariate and bivariate visualizations.
}
\details{
This is a general module to visualize 1 & 2 dimensional data.
}
\examples{

\dontrun{

N <- 100
ASL <- data.frame(
  USUBJID = paste("id", seq_len(N), sep = "-"),
  STUDYID = "study1",
  F1 = factor(sample(paste0("facet_1_", c("A", "B")), N, TRUE)),
  F2 = factor(sample(paste0("facet_2_", c("a", "b", "c")), N, TRUE)),
  cont = rnorm(N),
  disc = factor(sample(letters[1:5], N, TRUE)),
  cont2 = runif(N),
  disc2 = factor(sample(LETTERS[1:5], N, TRUE))
)

attr(ASL, "source") <- "# ASL is random data"

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_g_bivariate(
      dataname = "ASL",
      x_var = "cont",
      x_var_choices = names(ASL),
      y_var = "cont2",
      y_var_choices = names(ASL),
      use_density = FALSE,
      color_by_var = "STUDYID",
      color_by_var_choices = "STUDYID",
      row_facet_var = NULL,
      row_facet_var_choices = c("F1", "F2"),
      col_facet_var = NULL,
      col_facet_var_choices = c("F1", "F2"),
      plot_height = c(600, 200, 2000)
    )
  ))

shinyApp(x$ui, x$server)

}

}
