% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_bivariate.R
\name{tm_g_bivariate}
\alias{tm_g_bivariate}
\title{Univariate and bivariate visualizations.}
\usage{
tm_g_bivariate(label = "Bivariate Plots", x, y, row_facet, col_facet,
  colour = NULL, fill = NULL, size = NULL, use_density = FALSE,
  expert_settings = TRUE, free_x_scales = FALSE,
  free_y_scales = FALSE, plot_height = c(600, 200, 2000),
  ggtheme = "minimal", with_show_r_code = TRUE, pre_output = NULL,
  post_output = NULL)
}
\arguments{
\item{label}{(\code{character}) Label of the module}

\item{x}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Variable name selected to plot along the x-axis by default. Variable can be numeric, factor or character.
No empty selections are allowed!}

\item{y}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Variable name selected to plot along the y-axis by default. Variable can be numeric, factor or character.}

\item{row_facet}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Variable for row facetting}

\item{col_facet}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Variable for col facetting}

\item{colour}{optional, (\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Variable selection for the colouring inside the expert settings}

\item{fill}{optional, (\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Variable selection for the filling inside the expert settings}

\item{size}{optional, (\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Variable selection for the size of \code{geom_point} plots inside the expert settings}

\item{use_density}{(\code{logical}) value for whether density (\code{TRUE}) is plotted or frequency (\code{FALSE})}

\item{expert_settings}{(\code{logical}) Whether coloring, filling and size should be chosen
by the user}

\item{free_x_scales}{(\code{logical}) If X scaling shall be changeable}

\item{free_y_scales}{(\code{logical}) If Y scaling shall be changeable}

\item{plot_height}{(\code{numeric}) \code{c(value, min and max)} of plot height slider}

\item{ggtheme}{(\code{character}) ggplot theme to be used by default. All themes can be chosen by the user.}

\item{with_show_r_code}{(\code{logical}) Whether show R code button shall be shown}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
Univariate and bivariate visualizations.
}
\details{
This is a general module to visualize 1 & 2 dimensional data.
}
\examples{
# datasets: single wide
# Bivariate plot of selected variable (AGE) against selected (SEX)
library(random.cdisc.data)
ADSL <- cadsl

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- cadsl",
    check = TRUE
  ),
  modules = root_modules(
    tm_g_bivariate(
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = names(ADSL),
          selected = "AGE",
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = names(ADSL),
          selected = "SEX",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      row_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variables:",
          choices = names(ADSL),
          selected = NULL,
          multiple = TRUE,
          fixed = FALSE
        )
      ),
      col_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variables:",
          choices = names(ADSL),
          selected = NULL,
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: different wide
# Bivariate plot with RACE (factor) plotted over AGE (numeric)
library(random.cdisc.data)
library(dplyr)

ADSL <- cadsl
ADSL_2 <- ADSL

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    dataset("ADSL_2", ADSL_2, keys = get_cdisc_keys("ADSL")),
    code = 'ADSL <- cadsl
            ADSL_2 <- ADSL',
    check = FALSE #TODO
  ),
  modules = root_modules(
    tm_g_bivariate(
      label = "Association plots",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = c("AGE", "SEX", "STRATA1", "RACE"),
          selected = c("AGE"),
          multiple = FALSE
        )),
      y = data_extract_spec(
        dataname = "ADSL_2",
        select = select_spec(
          label = "Select variables:",
          choices = c("COUNTRY", "AGE", "RACE"),
          selected = "RACE",
          multiple = FALSE
        )),
      # Row facetting by first data set
      row_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
         label = "Select variable:",
         choices = names(ADSL),
         selected = NULL,
         multiple = FALSE,
         fixed = FALSE
        )
      ),
      # Col facetting by second data set
      col_facet = data_extract_spec(
        dataname = "ADSL_2",
        select = select_spec(
          label = "Select variables:",
          choices = names(ADSL),
          selected = NULL,
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: multiple long datasets
# Bivariate plot of different parameters from ADRS or ADTTE datasets
library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs
ADTTE <- cadtte

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    cdisc_dataset("ADTTE", ADTTE),
    code = "ADSL <- cadsl; ADRS <- cadrs; ADTTE <- cadtte",
    check = TRUE
  ),
  modules = root_modules(
    tm_g_bivariate(
      label = "Bivariate Plots of two long datasets",
      x = data_extract_spec(
        dataname = "ADRS",
        filter = filter_spec(
          label = "Select endpoints:",
          vars = c("PARAMCD", "AVISIT"),
          choices = apply(as.matrix(expand.grid(levels(ADRS$PARAMCD), levels(ADRS$AVISIT))),
                          1, paste, collapse = " - "),
          selected = "OVRINV - Screening",
          multiple = TRUE
        ),
        select = select_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = TRUE
        )
      ),
      y = data_extract_spec(
        dataname = "ADTTE",
        select = select_spec(
          label = "Select variable:",
          choices = c("AVAL", "CNSR"),
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE
        ),
        filter = filter_spec(
          label = "Select endpoint:",
          vars = c("PARAMCD"),
          choices = unique(ADTTE$PARAMCD),
          selected = "OS",
          multiple = FALSE
        )
      ),
      row_facet = data_extract_spec(
        dataname = "ADRS",
        filter = filter_spec(
          label = "Select endpoints:",
          vars = c("PARAMCD", "AVISIT"),
          choices = apply(as.matrix(expand.grid(levels(ADRS$PARAMCD), levels(ADRS$AVISIT))),
                          1, paste, collapse = " - "),
          selected = "OVRINV - Screening",
          multiple = TRUE
        ),
        select = select_spec(
          label = "Select variable:",
          choices = c("AVAL", "AVISIT"),
          selected = "AVISIT",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      col_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variables:",
          choices = c("SEX", "RACE"),
          selected = NULL,
          multiple = TRUE,
          fixed = FALSE
        )
      ),
      expert_settings = TRUE,
      plot_height = c(600, 200, 2000),
      ggtheme = "grey"
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: wide, long
# Bivariate plot with biomarker measurement (ADSL$BMRKR1) plotted over
# date of response (ADRS$AVAL)
library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs

app <- init(
 data = cdisc_data(
   cdisc_dataset("ADSL", ADSL),
   cdisc_dataset("ADRS", ADRS),
   code = "ADSL <- cadsl; ADRS <- cadrs",
   check = TRUE
 ),
 modules = root_modules(
   tm_g_bivariate(
     label = "Association Plots",
     x = data_extract_spec(
       dataname = "ADRS",
       filter = list(
         filter_spec(
           vars = "PARAM",
           choices = levels(ADRS$PARAM),
           selected = levels(ADRS$PARAM)[1],
           multiple = FALSE,
           label = "Select response:"
         ),
         filter_spec(
           vars = "AVISIT",
           choices = levels(ADRS$AVISIT),
           selected = levels(ADRS$AVISIT)[1],
           multiple = FALSE,
           label = "Select visit:"
         )
       ),
       select = select_spec(
         choices = "AVAL",
         selected = "AVAL",
         multiple = FALSE,
         label = "Select variable:",
         fixed = TRUE
       )
    ),
     y = data_extract_spec(
       dataname = "ADSL",
       select = select_spec(
         choices = c("BMRKR1", "SEX", "AGE", "RACE", "COUNTRY"),
         selected = "BMRKR1",
         multiple = FALSE,
         label = "Select variable:",
         fixed = FALSE
       )
     ),
     row_facet = data_extract_spec(
       dataname = "ADRS",
       select = select_spec(
          choices = "PARAMCD",
          selected = "PARAMCD",
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
     ),
     col_facet = data_extract_spec(
       dataname = "ADRS",
       select = select_spec(
          choices = "AVISIT",
          selected = "AVISIT",
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: wide, long, long

library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs
ADLB <- cadlb

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    cdisc_dataset("ADLB", ADLB),
    code = "ADSL <- cadsl; ADRS <- cadrs; ADLB <- cadlb",
    check = TRUE
  ),
  modules = root_modules(
    tm_g_bivariate(
      label = "Association Plots",
      x = data_extract_spec(
        dataname = "ADRS",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADRS$PARAMCD),
            selected = levels(ADRS$PARAMCD)[1],
            multiple = FALSE,
            label = "Select response:"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADRS$AVISIT),
            selected = levels(ADRS$AVISIT)[1],
            multiple = FALSE,
            label = "Select visit:"
          )
        ),
        select = select_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          label = "Select variable:"
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          choices = c("BMRKR1", "SEX", "AGE", "RACE", "COUNTRY"),
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      row_facet = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAM",
            choices = levels(ADLB$PARAM),
            selected = levels(ADLB$PARAM)[1],
            multiple = FALSE,
            label = "Select measurement:"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Select visit:"
          )
        ),
        select = select_spec(
          choices = "ARMCD",
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      ),
      col_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          choices = c("SEX", "AGE", "RACE", "COUNTRY"),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      ),
      expert_settings = TRUE,
      plot_height = c(600, 200, 2000),
      ggtheme = "grey"
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: same dataset long
# Bivariate plot of AVAL vs biomarker value split by PARAMCD and AVISIT

library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    code = "ADSL <- cadsl; ADRS <- cadrs",
    check = TRUE
  ),
  modules = root_modules(
    tm_g_bivariate(
      x = data_extract_spec(
        dataname = "ADRS",
        select = select_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      ),
      y = data_extract_spec(
        dataname = "ADRS",
        select = select_spec(
          choices = c("BMRKR1", "BMRKR2"),
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      ),
      row_facet = data_extract_spec(
        dataname = "ADRS",
        select = select_spec(
          choices = "PARAMCD",
          selected = "PARAMCD",
          multiple = TRUE,
          fixed = FALSE,
          label = "Select variables:"
        )
      ),
      col_facet = data_extract_spec(
        dataname = "ADRS",
        select = select_spec(
          choices = "AVISIT",
          selected = "AVISIT",
          multiple = TRUE,
          fixed = FALSE,
          label = "Select variables:"
        )
      )
    )
  )
)
\dontrun{
  shinyApp(app$ui, app$server)
}

# datasets: different subsets of long dataset
# Bivariate plot of one lab measurement versus another, optionally split by additional variables

library(random.cdisc.data)

ADSL <- cadsl
ADLB <- cadlb

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADLB", ADLB),
    code = "ADSL <- cadsl; ADLB <- cadlb",
    check = TRUE
  ),
  modules = root_modules(
    tm_g_bivariate(
      x = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[1],
            multiple = FALSE,
            label = "Select lab:"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[2],
            multiple = FALSE,
            label = "Select visit:"
          )
        ),
        select = select_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = TRUE
        )
      ),
      y = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[1],
            multiple = FALSE,
            label = "Select lab:"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Select visit:"
          )
        ),
        select = select_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = TRUE
        )
      ),
      use_density = FALSE,
      row_facet = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[1],
            multiple = FALSE,
            label = "Select lab:"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Select category:"
          )
        ),
        select = select_spec(
          choices = c("RACE", "SEX", "ARMCD", "ACTARMCD"),
          selected = NULL,
          multiple = TRUE,
          fixed = FALSE,
          label = "Select variable:"
        )
      ),
      col_facet = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[1],
            multiple = FALSE,
            label = "Select lab:"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[3],
            multiple = FALSE,
            label = "Select category:"
          )
        ),
        select = select_spec(
          choices = c("RACE", "SEX", "ARMCD", "ACTARMCD"),
          selected = NULL,
          multiple = TRUE,
          fixed = FALSE,
          label = "Select variables:"
        )
      ),
      expert_settings = TRUE,
      plot_height = c(600, 200, 2000),
      ggtheme = "grey"
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
}
