% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_bivariate.R
\name{tm_g_bivariate}
\alias{tm_g_bivariate}
\title{Univariate and bivariate visualizations.}
\usage{
tm_g_bivariate(label = "Uni- and Bivariate Plots", dataname, x_var,
  x_var_choices = x_var, y_var, y_var_choices = y_var,
  use_density = FALSE, color_by_var, color_by_var_choices, x_facet_var,
  x_facet_var_choices, y_facet_var, y_facet_var_choices,
  plot_height = c(600, 200, 2000))
}
\arguments{
\item{x_var}{variable name selected to plot along the x-axis by default. Variable can be numeric, factor or character.}

\item{x_var_choices}{character vector with variable names available as choices to plot along x-axis}

\item{y_var}{variable name selected to plot along the y-axis by default. Variable can be numeric, factor or character.}

\item{y_var_choices}{character vector with variable names available as choices to plot along y-axis}

\item{use_density}{boolean value for whether density is plotted}

\item{x_facet_var}{variable for x facets}

\item{x_facet_var_choices}{vector with \code{x_facet_var} choices}

\item{y_facet_var}{variable for y facets}

\item{y_facet_var_choices}{vector with \code{y_facet_var} choices}

\item{plot_height}{range of plot height}
}
\description{
Univariate and bivariate visualizations.
}
\details{
This is a general module to visualize 1 & 2 dimensional data.
<<<<<<< HEAD


 1d

 cont  abs  hist
 cat.  abs  barplot


 2d
 x         y     interest    plot
 cont.   cont.   abs       scatterplot
 cont.   cat.    abs       boxplot
 cat.    cat.    abs       mosaic plot
 cont.   cont.   density
 cont.   cat.    density
 cat.    cat.    density


 color_by
 facet_by
=======
>>>>>>> devel-tm_g_bivariate
}
\examples{

## TODO:
# 1) Review bivariate plots -- output is not correct when 2 discrete vars are selected, also need to add density
# 2) add line_colorby_var and line_colorby_var_choices
# 3) Add x- y- faceting for 1D and 2D plots

\dontrun{

N <- 100
ASL <- data.frame(
  USUBJID = paste("id", seq_len(N), sep = "-"),
  STUDYID = "study1",
  cont = rnorm(N),
  disc = factor(sample(letters[1:5], N, TRUE)),
  cont2 = runif(N),
  disc2 = factor(sample(LETTERS[1:5], N, TRUE))
)

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_g_bivariate(
      dataname = "ASL",
      x_var = "cont",
      x_var_choices = names(ASL),
      y_var = "cont2",
      y_var_choices = names(ASL),
      use_density = FALSE,
      x_facet_var = "SEX",
      x_facet_var_choices = c("SEX", "ARM"),
      y_facet_var = "ARM",
      y_facet_var_choices = c("ARM", "SEX"),
      plot_height = c(600, 200, 2000)
    )
  ))

shinyApp(x$ui, x$server)

}

}
