% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_regression.R
\name{tm_g_regression}
\alias{tm_g_regression}
\title{Scatterplot and Regression Model}
\usage{
tm_g_regression(label = "Regression Analysis", dataname, regressor,
  response, plot_height = c(600, 200, 2000), pre_output = NULL,
  post_output = NULL)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{dataname}{name of datasets used to generate the regression plot (just used for labeling)}

\item{regressor}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
regressor variable from an incoming dataset with filtering and selecting.}

\item{response}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
response variable from an incoming dataset with filtering and selecting.}

\item{plot_height}{(\code{numeric}) a vector of length three with \code{c(value, min and max)} for a slider
encoding the plot height.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
Scatterplot and Regression Model
}
\examples{
library(random.cdisc.data)
library(tern)


# datasets: same wide

library(random.cdisc.data)
library(tern)

ADSL <- cadsl
keys(ADSL) <- c("STUDYID", "USUBJID")

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = 'ADSL <- cadsl
            keys(ADSL) <- c("STUDYID", "USUBJID")',
    check = FALSE),
  modules = root_modules(
    tm_g_regression(
      label = "Regression",
      dataname = "ADSL",
      response = data_extract_spec(
        dataname = "ADSL",
        columns = columns_spec(
          choices = c("BMRKR1", "BMRKR2"),
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = FALSE,
          label = "variable"
        )
      ),
      regressor = data_extract_spec(
        dataname = "ADSL",
        columns = columns_spec(
          choices = c("AGE", "SEX", "RACE"),
          selected = c("AGE"),
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: different wide

library(random.cdisc.data)
library(tern)
library(dplyr)

ADSL <- cadsl
ADSL <- mutate_at(ADSL,
                 .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
                 .funs = list(~as.factor(.))) \%>\% select("ARM", "ACTARM", "ACTARMCD",
 "SEX", "AGE", "USUBJID", "STUDYID", "BMRKR1", "BMRKR2")
keys(ADSL) <- c("STUDYID", "USUBJID")


ADSL_2 <- mutate_at(cadsl,
                 .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
                 .funs = list(~as.factor(.))) \%>\% select("ACTARM", "AGE", "STRATA2", "COUNTRY",
                 "USUBJID", "STUDYID")
keys(ADSL_2) <- c("STUDYID", "USUBJID")


app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    dataset("ADSL_2", ADSL_2),
    code = 'ADSL <- cadsl
            ADSL <- mutate_at(ADSL,
                 .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
                 .funs = list(~as.factor(.))) \%>\% select("ARM", "ACTARM", "ACTARMCD",
                     "SEX", "STRATA1", "AGE", "USUBJID", "STUDYID", "STRATA2")
            ADSL_2 <- mutate_at(cadsl,
                 .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
                 .funs = list(~as.factor(.))) \%>\% select("ACTARM", "AGE", "STRATA2",
                 "COUNTRY", "USUBJID", "STUDYID")
            keys(ADSL) <- keys(ADSL_2) <- c("STUDYID", "USUBJID")',
    check = FALSE),
  modules = root_modules(
    tm_g_regression(
      label = "Regression",
      dataname = c("ADSL", "ADSL_2"),
      response = data_extract_spec(
        dataname = "ADSL",
        columns = columns_spec(
         label = "Select variable",
          choices = c("BMRKR1", "BMRKR2"),
          selected = c("BMRKR1"),
          multiple = FALSE
        )),
      regressor = data_extract_spec(
        dataname = "ADSL_2",
        columns = columns_spec(
          label = "Select variables",
          choices = c("AGE", "RACE", "COUNTRY"),
          selected = c("AGE", "RACE"),
          multiple = TRUE
        ))
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: same long
# Examine linear relationship between responses of different parameters

library(random.cdisc.data)
library(tern)

ADSL <- cadsl
ADRS <- cadrs

keys(ADSL) <-  c("STUDYID", "USUBJID")
keys(ADRS) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    code = 'ADSL <- cadsl
            ADRS <- cadrs
            keys(ADSL) <-  c("STUDYID", "USUBJID")
            keys(ADRS) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")',
    check = FALSE),
  modules = root_modules(
    tm_g_regression(
      label = "Regression",
      dataname = c("ADSL", "ADRS"),
      response = data_extract_spec(
        dataname = "ADRS",
        filter = list(
          filter_spec(
            vars = "PARAM",
            choices = levels(ADRS$PARAM),
            selected = levels(ADRS$PARAM)[1],
            multiple = FALSE,
            label = "Choose measurement"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADRS$AVISIT),
            selected = levels(ADRS$AVISIT)[1],
            multiple = FALSE,
            label = "Choose visit"
          )
        ),
        columns = columns_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE,
          label = "Response variable"
        )
      ),
      regressor = data_extract_spec(
        dataname = "ADRS",
        filter = list(
          filter_spec(
            vars = "PARAM",
            choices = levels(ADRS$PARAM),
            selected = levels(ADRS$PARAM)[1:2],
            multiple = TRUE,
            label = "Choose measurement"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADRS$AVISIT),
            selected = levels(ADRS$AVISIT)[2],
            multiple = TRUE,
            label = "Choose visit"
          )
        ),
        columns = columns_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE,
          label = "Regressor variable"
        )
      )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: multiple long datasets
library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs
ADTTE <- cadtte

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    cdisc_dataset("ADTTE", ADTTE),
    code = "ADSL <- cadsl; ADRS <- cadrs; ADTTE <- cadtte",
    check = FALSE
  ),
  modules = root_modules(
    tm_g_regression(
      label = "Regression Analysis on two long datasets",
      dataname = c("ADSL", "ADRS", "ADTTE"),
      response = data_extract_spec(
        dataname = "ADTTE",
        columns = columns_spec(
          choices = c("AVAL", "CNSR"),
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE
        ),
        filter = filter_spec(
          vars = c("PARAMCD"),
          choices = unique(ADTTE$PARAMCD),
          selected = "OS",
          multiple = FALSE,
          label = "ADTTE filter"
        )
      ),
      regressor = list(
        data_extract_spec(
          dataname = "ADRS",
          columns = columns_spec(
            choices = names(ADRS),
            selected = "AVAL",
            multiple = TRUE,
            fixed = FALSE
          ),
          filter = filter_spec(
            vars = c("PARAMCD", "AVISIT"),
            choices = apply(expand.grid(
            levels(ADRS$PARAMCD), levels(ADRS$AVISIT)), 1, paste, collapse = " - "),
            selected = "OVRINV - Screening",
            multiple = TRUE,
            label = "ADRS filter"
          )
        ),
        data_extract_spec(
          dataname = "ADSL",
          columns = columns_spec(
            choices = c("BMRKR1", "BMRKR2"),
            selected = c("BMRKR1"),
            multiple = TRUE,
            fixed = FALSE
          )
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: wide and long

library(random.cdisc.data)
library(tern)

ADSL <- cadsl
ADLB <- cadlb

keys(ADSL) <- c("STUDYID", "USUBJID")
keys(ADLB) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")

app <- init(
 data = cdisc_data(
   cdisc_dataset("ADSL", ADSL),
   cdisc_dataset("ADLB", ADLB),
   code = 'ADSL <- cadsl
           ADLB <- cadlb
           keys(ADSL) <- c("STUDYID", "USUBJID")
           keys(ADLB) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")',
   check = FALSE),
 modules = root_modules(
   tm_g_regression(
     label = "Regression",
     dataname = c("ADSL", "ADLB"),
     response = data_extract_spec(
       dataname = "ADLB",
       filter = list(
         filter_spec(
           vars = "PARAM",
          choices = levels(ADLB$PARAM),
           selected = levels(ADLB$PARAM)[1],
           multiple = FALSE,
           label = "Choose measurement"
         ),
         filter_spec(
           vars = "AVISIT",
           choices = levels(ADLB$AVISIT),
           selected = levels(ADLB$AVISIT)[1],
           multiple = FALSE,
           label = "Choose visit"
         )
       ),
       columns = columns_spec(
         choices = "AVAL",
         selected = "AVAL",
         multiple = FALSE,
         fixed = FALSE,
         label = "variable"
       )
    ),
     regressor = data_extract_spec(
       dataname = "ADSL",
       columns = columns_spec(
         choices = c("BMRKR1", "BMRKR2", "AGE"),
         selected = c("AGE"),
         multiple = TRUE,
         fixed = FALSE
      )
     )
   )
)
)

\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: wide, long, long

library(random.cdisc.data)
library(tern)

ADSL <- cadsl
ADLB <- cadlb
ADRS <- cadrs

keys(ADSL) <- c("STUDYID", "USUBJID")
keys(ADLB) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")
keys(ADRS) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADLB", ADLB),
    cdisc_dataset("ADRS", ADRS),
    code = 'ADSL <- cadsl
            ADLB <- cadlb
            keys(ADSL) <- c("STUDYID", "USUBJID")
            keys(ADLB) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")
            keys(ADRS) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")',
    check = FALSE),
  modules = root_modules(
    tm_g_regression(
      label = "Regression",
      dataname = c("ADSL", "ADLB", "ADRS"),
      response = list(data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAM",
            choices = levels(ADLB$PARAM),
            selected = levels(ADLB$PARAM)[1],
            multiple = FALSE,
            label = "Choose measurement"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Choose visit"
          )
        ),
        columns = columns_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE,
          label = "variable"
        )
      ),
      data_extract_spec(
        dataname = "ADRS",
        filter = list(
          filter_spec(
            vars = "ARMCD",
            choices = levels(ADRS$ARMCD),
            selected = levels(ADRS$ARMCD)[1],
            multiple = FALSE,
            label = "Choose ARM"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADRS$AVISIT),
            selected = levels(ADRS$AVISIT)[1],
            multiple = FALSE,
            label = "Choose visit"
          )
        ),
        columns = columns_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE
        )
      )),
      regressor = data_extract_spec(
        dataname = "ADSL",
        columns = columns_spec(
          choices = c("BMRKR1", "BMRKR2"),
          selected = c("BMRKR1"),
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: different subsets of long dataset
# Examine linear relationship between different lab measurements

library(random.cdisc.data)
library(tern)

ADSL <- cadsl
ADLB <- cadlb

keys(ADSL) <- c("STUDYID", "USUBJID")
keys(ADLB) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADLB", ADLB),
    code = 'ADSL <- cadsl
            ADLB <- cadlb
            keys(ADSL) <- c("STUDYID", "USUBJID")
            keys(ADLB) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")',
    check = FALSE),
  modules = root_modules(
    tm_g_regression(
      label = "Regression",
      dataname = "ADLB",
      response = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[1],
            multiple = FALSE,
            label = "Lab"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Visit"
          )
        ),
        columns = columns_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = TRUE,
          label = "Response variable"
        )
      ),
      regressor = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[2:3],
            multiple = TRUE,
            label = "Choose measurement"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Choose visit"
          )
        ),
        columns = columns_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = TRUE,
          label = "Regressor variable"
        )
      )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
