% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_regression.R
\name{tm_g_regression}
\alias{tm_g_regression}
\title{Scatterplot and Regression Model}
\usage{
tm_g_regression(label = "Regression Analysis", dataname, response_var,
  regressor_var, plot_height = c(600, 200, 2000), pre_output = NULL,
  post_output = NULL)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{dataname}{name of dataset used to generate table}

\item{response_var}{(\code{choices_selected}) Output of \code{teal::choices_selected} to define the
response variable}

\item{regressor_var}{(\code{choices_selected}) Output of \code{teal::choices_selected} to define the
regressor variable}

\item{plot_height}{if scalar then the plot will have a fixed height. If a
slider should be presented to adjust the plot height dynamically then it
can be a vector of length three with \code{c(value, min and max)}.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
This module is designed for horizontal data
}
\examples{

N <- 100
ASL <- data.frame(
  USUBJID = paste("id", seq_len(N), sep = "-"),
  STUDYID = "study1",
  F1 = factor(sample(paste0("facet_1_", c("A", "B")), N, TRUE)),
  F2 = factor(sample(paste0("facet_2_", c("a", "b", "c")), N, TRUE)),
  cont = rnorm(N),
  disc = factor(sample(letters[1:5], N, TRUE)),
  cont2 = runif(N),
  disc2 = factor(sample(LETTERS[1:5], N, TRUE))
)

attr(ASL, "source") <- "# ASL is random data"

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_g_regression(
      dataname = "ASL",
      response_var = choices_selected(c("cont", "cont2"), "cont"),
      regressor_var = choices_selected(names(ASL), "cont2"),
      plot_height = c(600, 200, 2000)
    )
  )
)

\dontrun{
shinyApp(x$ui, x$server)
}
}
