% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplot.R
\name{tm_g_scatterplot}
\alias{tm_g_scatterplot}
\title{\code{teal} module: Scatterplot}
\usage{
tm_g_scatterplot(
  label = "Scatterplot",
  x,
  y,
  color_by = NULL,
  size_by = NULL,
  row_facet = NULL,
  col_facet = NULL,
  plot_height = c(600, 200, 2000),
  plot_width = NULL,
  alpha = c(1, 0, 1),
  shape = shape_names,
  size = c(5, 1, 15),
  max_deg = 5L,
  rotate_xaxis_labels = FALSE,
  ggtheme = c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void"),
  pre_output = NULL,
  post_output = NULL,
  table_dec = 4,
  ggplot2_args = teal.widgets::ggplot2_args()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{x}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) Specifies
variable names selected to plot along the x-axis by default.}

\item{y}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) Specifies
variable names selected to plot along the y-axis by default.}

\item{color_by}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) optional,
defines the color encoding. If \code{NULL} then no color encoding option will be displayed.}

\item{size_by}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) optional,
defines the point size encoding. If \code{NULL} then no size encoding option will be displayed.}

\item{row_facet}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) optional,
specifies the variable(s) for faceting rows.}

\item{col_facet}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) optional,
specifies the variable(s) for faceting columns.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{plot_width}{(\code{numeric}) optional, specifies the plot width as a three-element vector of
\code{value}, \code{min}, and \code{max} for a slider encoding the plot width.}

\item{alpha}{(\code{integer(1)} or \code{integer(3)}) optional, specifies point opacity.
\itemize{
\item When the length of \code{alpha} is one: the plot points will have a fixed opacity.
\item When the length of \code{alpha} is three: the plot points opacity are dynamically adjusted based on
vector of \code{value}, \code{min}, and \code{max}.
}}

\item{shape}{(\code{character}) optional, character vector with the names of the
shape, e.g. \code{c("triangle", "square", "circle")}. It defaults to \code{shape_names}. This is a complete list from
\code{vignette("ggplot2-specs", package="ggplot2")}.}

\item{size}{(\code{integer(1)} or \code{integer(3)}) optional, specifies point size.
\itemize{
\item When the length of \code{size} is one: the plot point sizes will have a fixed size.
\item When the length of \code{size} is three: the plot points size are dynamically adjusted based on
vector of \code{value}, \code{min}, and \code{max}.
}}

\item{max_deg}{(\code{integer}) optional, maximum degree for the polynomial trend line. Must not be less than 1.}

\item{rotate_xaxis_labels}{(\code{logical}) optional, whether to rotate plot X axis labels. Does not
rotate by default (\code{FALSE}).}

\item{ggtheme}{(\code{character}) optional, \code{ggplot2} theme to be used by default. Defaults to \code{"gray"}.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{table_dec}{(\code{integer}) optional, number of decimal places used to round numeric values in the table.}

\item{ggplot2_args}{(\code{ggplot2_args}) object created by \code{\link[teal.widgets:ggplot2_args]{teal.widgets::ggplot2_args()}}
with settings for the module plot.
The argument is merged with options variable \code{teal.ggplot2_args} and default module setup.

For more details see the vignette: \code{vignette("custom-ggplot2-arguments", package = "teal.widgets")}}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Generates a customizable scatterplot using \code{ggplot2}.
This module allows users to select variables for the x and y axes,
color and size encodings, faceting options, and more. It supports log transformations,
trend line additions, and dynamic adjustments of point opacity and size through UI controls.
}
\note{
For more examples, please see the vignette "Using scatterplot" via
\code{vignette("using-scatterplot", package = "teal.modules.general")}.
}
\examples{
\dontshow{if (requireNamespace("ggpmisc", quietly = TRUE) && requireNamespace("ggExtra", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# general data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  CO2 <- CO2
})
datanames(data) <- "CO2"

app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplot(
      label = "Scatterplot Choices",
      x = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["CO2"]], c("conc", "uptake")),
          selected = "conc",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["CO2"]], c("conc", "uptake")),
          selected = "uptake",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      color_by = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(
            data[["CO2"]],
            c("Plant", "Type", "Treatment", "conc", "uptake")
          ),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      size_by = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["CO2"]], c("conc", "uptake")),
          selected = "uptake",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      row_facet = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["CO2"]], c("Plant", "Type", "Treatment")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      col_facet = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["CO2"]], c("Plant", "Type", "Treatment")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("ggpmisc", quietly = TRUE) && requireNamespace("ggExtra", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# CDISC data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  ADSL <- rADSL
})
datanames(data) <- c("ADSL")
join_keys(data) <- default_cdisc_join_keys[datanames(data)]

app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplot(
      label = "Scatterplot Choices",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], c("AGE", "BMRKR1", "BMRKR2")),
          selected = "AGE",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], c("AGE", "BMRKR1", "BMRKR2")),
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      color_by = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(
            data[["ADSL"]],
            c("AGE", "BMRKR1", "BMRKR2", "RACE", "REGION1")
          ),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      size_by = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], c("AGE", "BMRKR1")),
          selected = "AGE",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      row_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], c("BMRKR2", "RACE", "REGION1")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      col_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], c("BMRKR2", "RACE", "REGION1")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
\dontshow{\}) # examplesIf}
}
\section{Examples in Shinylive}{
\itemize{
  \item example-1\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNruAcrDFUdTh2BTAJCVQYWlECMNxdL1o4UmpWXQBeeydnWwAyPN13JL8A0SCCELCI5wAPUmZ4xO8UtMzsl0UIAGJdKQg1al1+KFIoXTha2FQRJRGxg2MuagB9eagbOdHxo10Ad1pSAAsVdnWEkCVdIs9vd3YB0VJiaiJGLuuAYQB5ACZF3Q-X5KAC+XXW0HgojO21suzCQLCSiUaFQAJUh1CEGu63a5yuuj4QhEonaROEYix12upBgKwkK1io3IjBmRFIVOp12oUHocCGWTCAGUCMyZGzSICjkRaJVRPECVzanjtitJg11KRGag4AROVycdtIXpBWBEXhFQbRPzdZKstaRBptbr9Qbubz+e1hTaNLotCxaLyRIgFdi3dSCNLZWJ2v7GIH6CIVpGZXKYWNgMAEX8wgBdXMJPVhYgWJphQSoMYAazgYTkclwlrdDttcH4XsIJDiFrD4cJwk0MxNugAYgBBAAyQucjd74ZMtFqbfa46nrjnXI+BobTd06Sy6zV9WYTvKLt3hrGxo75tnfZbvvtPq1Z6LG4NPL5At03sdkrjCbBqGfbXCm0aklkAFBnAyZRmm6yZtmwJgPmhahJ2pZ4D+YAVtWtZgPWd4gQ+5DtqauFQDWwEgTAA60EOK6TtORF9guS5kaOTHrn2W6bixEZEK8jArPQ+7DKq6oni+Opvn2EIBDeOY9vez7tCRzqySBn6eqaQqqVBiZwCGykgWBcqxgG0GwamlIXgaCFZmaSmoXZXJFmAAAKPJkGWYB2KwOq+XY7ijPAPlYcWXa+RRVEEa5ug7u+XIkcuWS+I4E4Tvxbq0dQg4iIxa7ZQabGpZxa52bx1KJc2tAAF4waJKpjEeGqnjJrpcvJ8CKchxXXCRanPhpnUfh636-rafqWYZxn9RGcExpBM1JmZlIOUheYFro7klt2CTlpWlH4YR8UpRxh14dRfa5flw6rsx8WlRxD3ceGVXXDVBqMEQewrCYwR2uJLWSZqI12d1w6bSZzaqU+f7g0l1LaRNYB6X+03xtBc3xWtEGY4BMF4+mUCIU5yGoTt6FeVA4UHX5AX4fTwVwKFFCkHWX0qX+ZXpZl83XLd9EFVkr0C6Yi5la9lW7lzAmrADlRA4eoPteeSOQ7111WnDujqa+o1cijHbo1NBlATD4Z4xZWOGdZ4Ek2T5qU+5NN09h-mBRFfkhbS7Oc+L53tHzWXxULDGi1x4vPYV04y++VVbl0XS0CYujsCoLKatoIT1rolxhqIJwQKwY7oOwqIACTeAkVfWowOjvKCShgCCuZAA}{Open in Shinylive}\cr
    \if{html}{\out{<iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNruAcrDFUdTh2BTAJCVQYWlECMNxdL1o4UmpWXQBeeydnWwAyPN13JL8A0SCCELCI5wAPUmZ4xO8UtMzsl0UIAGJdKQg1al1+KFIoXTha2FQRJRGxg2MuagB9eagbOdHxo10Ad1pSAAsVdnWEkCVdIs9vd3YB0VJiaiJGLuuAYQB5ACZF3Q-X5KAC+XXW0HgojO21suzCQLCSiUaFQAJUh1CEGu63a5yuuj4QhEonaROEYix12upBgKwkK1io3IjBmRFIVOp12oUHocCGWTCAGUCMyZGzSICjkRaJVRPECVzanjtitJg11KRGag4AROVycdtIXpBWBEXhFQbRPzdZKstaRBptbr9Qbubz+e1hTaNLotCxaLyRIgFdi3dSCNLZWJ2v7GIH6CIVpGZXKYWNgMAEX8wgBdXMJPVhYgWJphQSoMYAazgYTkclwlrdDttcH4XsIJDiFrD4cJwk0MxNugAYgBBAAyQucjd74ZMtFqbfa46nrjnXI+BobTd06Sy6zV9WYTvKLt3hrGxo75tnfZbvvtPq1Z6LG4NPL5At03sdkrjCbBqGfbXCm0aklkAFBnAyZRmm6yZtmwJgPmhahJ2pZ4D+YAVtWtZgPWd4gQ+5DtqauFQDWwEgTAA60EOK6TtORF9guS5kaOTHrn2W6bixEZEK8jArPQ+7DKq6oni+Opvn2EIBDeOY9vez7tCRzqySBn6eqaQqqVBiZwCGykgWBcqxgG0GwamlIXgaCFZmaSmoXZXJFmAAAKPJkGWYB2KwOq+XY7ijPAPlYcWXa+RRVEEa5ug7u+XIkcuWS+I4E4Tvxbq0dQg4iIxa7ZQabGpZxa52bx1KJc2tAAF4waJKpjEeGqnjJrpcvJ8CKchxXXCRanPhpnUfh636-rafqWYZxn9RGcExpBM1JmZlIOUheYFro7klt2CTlpWlH4YR8UpRxh14dRfa5flw6rsx8WlRxD3ceGVXXDVBqMEQewrCYwR2uJLWSZqI12d1w6bSZzaqU+f7g0l1LaRNYB6X+03xtBc3xWtEGY4BMF4+mUCIU5yGoTt6FeVA4UHX5AX4fTwVwKFFCkHWX0qX+ZXpZl83XLd9EFVkr0C6Yi5la9lW7lzAmrADlRA4eoPteeSOQ7111WnDujqa+o1cijHbo1NBlATD4Z4xZWOGdZ4Ek2T5qU+5NN09h-mBRFfkhbS7Oc+L53tHzWXxULDGi1x4vPYV04y++VVbl0XS0CYujsCoLKatoIT1rolxhqIJwQKwY7oOwqIACTeAkVfWowOjvKCShgCCuZAA" height="800" width="150\%" allow="fullscreen" scrolling="auto" style="border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem;"></iframe>}}
  \item example-2\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNruAcrDFUdTh2BTAJCVQYWlECMNxdL1o4UmpWXQBeeydnWwAyPN13JL8A0SCCELCI5wAPUmZ4xO8UtMzsl0UIAGJdAGEAEQBJAGU+3X4oUihdOFrYVBElSemDYy5qAH0VqBtlqZmjXQB3WlIACxV2HYSQJV0iz293dggxUmJqIkYuh4BBAYjAAyayKAOBSgAvl0dtB4KJrgdbEcCKEwOCgWEugArIgqTYAazgrARO2Rxn4cBMUGEpE2BH4MQIm1x+KJJOAsLKiOmcgAukolGhUKCVGdQhAHjt2jd7ro+EIRKJ2grhGIJQ8HqQYJsJJtYlNyIxFkRSBrNQ9qFB6HBqO0wiMCIaZCbSP1znjKqJ4nKLbUZQdNnMGuo6eU4KjfRaJgc4XosmEMT7JdGHqJbRG3Vl0yINPrUBHzanLdbbfawCMMxpdFoWLRrSJEMni5qCB7aF72rXGPX6CJ6e2vTyoMBgInAZiwHy+QlI+iAOKuPC6MIAIQAslgANJYACMTTXm53ACYsXJcFHUznM3B+OW-ovmy35bTaIt47oAGJ-IEjZwXlMWxMWhalvdpv1-VxAOjX4YIA1N0iyHYg3qZg83DSNoItLl4HvCcnyvKss10a90ILTDn10K0bTtBMKyIms6wbOAmzwS9ozbT0xC7Ji+zgAcuNJA5R3HCEpxnXQ5wfJcEkPbc9wPMAN3k08wDkc92ItUjyDvOjlJ3fc2Kw6MYFfd9wJ-P94OfYDQN0r9LKgltYItDTjM+b5NnoRCY2mFCQzIwtNN8qA4zwsTrMI3NiO0-MguMi1qLLOjK2ixie2Y1jIuLTiO24rJu17ftcqHYLsOEsd0Xw8TspbKTH2XOSDMU-SsFU2SwCwP4+hkldOucechgAeV8QyXOLNzKO0sCsl8RwgSBWqTLMkQLMgpaLVsmaHMg4LxoeSbo1EWgAC9+O8gM-ODNCw3Iotoxwj9RMnDaSIY7MiLiijnyS2i+tSzN0qKliCJywd8qB5iBLyoTphEqqxOnWc0WklqjwUtTDpbab7MTBrXoeUzqE0cysggqyytMEDtvJpyJvYrHNUYIhjk2alKmI5DrtDL77vK6YwropMjOx963ui3nKd+8sAerQrMtB1MSoh+W+Ohocdnh4WkcktFWvavqup6xSsAG4bRrPAnxZvey5oWq2iZJ1ayccq2tvs2m9oZpbPjZ4JOcDbnAu+1NHvCl7gu09pYowvno2llKGNVxtFY48HlQK3jivT4cteqnW531k3ut6sJTcGkaxsZqKbfaO3Fspx232dnaKYSzV3bWv8veM8bYK6LpaBMXR2BUI1Q20EJ1N0O4U1ES4IFYP50HYYUABJvASdf00YHQfihJQwEhPkgA}{Open in Shinylive}\cr
    \if{html}{\out{<iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNruAcrDFUdTh2BTAJCVQYWlECMNxdL1o4UmpWXQBeeydnWwAyPN13JL8A0SCCELCI5wAPUmZ4xO8UtMzsl0UIAGJdAGEAEQBJAGU+3X4oUihdOFrYVBElSemDYy5qAH0VqBtlqZmjXQB3WlIACxV2HYSQJV0iz293dggxUmJqIkYuh4BBAYjAAyayKAOBSgAvl0dtB4KJrgdbEcCKEwOCgWEugArIgqTYAazgrARO2Rxn4cBMUGEpE2BH4MQIm1x+KJJOAsLKiOmcgAukolGhUKCVGdQhAHjt2jd7ro+EIRKJ2grhGIJQ8HqQYJsJJtYlNyIxFkRSBrNQ9qFB6HBqO0wiMCIaZCbSP1znjKqJ4nKLbUZQdNnMGuo6eU4KjfRaJgc4XosmEMT7JdGHqJbRG3Vl0yINPrUBHzanLdbbfawCMMxpdFoWLRrSJEMni5qCB7aF72rXGPX6CJ6e2vTyoMBgInAZiwHy+QlI+iAOKuPC6MIAIQAslgANJYACMTTXm53ACYsXJcFHUznM3B+OW-ovmy35bTaIt47oAGJ-IEjZwXlMWxMWhalvdpv1-VxAOjX4YIA1N0iyHYg3qZg83DSNoItLl4HvCcnyvKss10a90ILTDn10K0bTtBMKyIms6wbOAmzwS9ozbT0xC7Ji+zgAcuNJA5R3HCEpxnXQ5wfJcEkPbc9wPMAN3k08wDkc92ItUjyDvOjlJ3fc2Kw6MYFfd9wJ-P94OfYDQN0r9LKgltYItDTjM+b5NnoRCY2mFCQzIwtNN8qA4zwsTrMI3NiO0-MguMi1qLLOjK2ixie2Y1jIuLTiO24rJu17ftcqHYLsOEsd0Xw8TspbKTH2XOSDMU-SsFU2SwCwP4+hkldOucechgAeV8QyXOLNzKO0sCsl8RwgSBWqTLMkQLMgpaLVsmaHMg4LxoeSbo1EWgAC9+O8gM-ODNCw3Iotoxwj9RMnDaSIY7MiLiijnyS2i+tSzN0qKliCJywd8qB5iBLyoTphEqqxOnWc0WklqjwUtTDpbab7MTBrXoeUzqE0cysggqyytMEDtvJpyJvYrHNUYIhjk2alKmI5DrtDL77vK6YwropMjOx963ui3nKd+8sAerQrMtB1MSoh+W+Ohocdnh4WkcktFWvavqup6xSsAG4bRrPAnxZvey5oWq2iZJ1ayccq2tvs2m9oZpbPjZ4JOcDbnAu+1NHvCl7gu09pYowvno2llKGNVxtFY48HlQK3jivT4cteqnW531k3ut6sJTcGkaxsZqKbfaO3Fspx232dnaKYSzV3bWv8veM8bYK6LpaBMXR2BUI1Q20EJ1N0O4U1ES4IFYP50HYYUABJvASdf00YHQfihJQwEhPkgA" height="800" width="150\%" allow="fullscreen" scrolling="auto" style="border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem;"></iframe>}}
}
}

