% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplot.R
\name{tm_g_scatterplot}
\alias{tm_g_scatterplot}
\title{\code{teal} module: Scatterplot}
\usage{
tm_g_scatterplot(
  label = "Scatterplot",
  x,
  y,
  color_by = NULL,
  size_by = NULL,
  row_facet = NULL,
  col_facet = NULL,
  plot_height = c(600, 200, 2000),
  plot_width = NULL,
  alpha = c(1, 0, 1),
  shape = shape_names,
  size = c(5, 1, 15),
  max_deg = 5L,
  rotate_xaxis_labels = FALSE,
  ggtheme = c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void"),
  pre_output = NULL,
  post_output = NULL,
  table_dec = 4,
  ggplot2_args = teal.widgets::ggplot2_args()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{x}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) Specifies
variable names selected to plot along the x-axis by default.}

\item{y}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) Specifies
variable names selected to plot along the y-axis by default.}

\item{color_by}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) optional,
defines the color encoding. If \code{NULL} then no color encoding option will be displayed.}

\item{size_by}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) optional,
defines the point size encoding. If \code{NULL} then no size encoding option will be displayed.}

\item{row_facet}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) optional,
specifies the variable(s) for faceting rows.}

\item{col_facet}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec}) optional,
specifies the variable(s) for faceting columns.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{plot_width}{(\code{numeric}) optional, specifies the plot width as a three-element vector of
\code{value}, \code{min}, and \code{max} for a slider encoding the plot width.}

\item{alpha}{(\code{integer(1)} or \code{integer(3)}) optional, specifies point opacity.
\itemize{
\item When the length of \code{alpha} is one: the plot points will have a fixed opacity.
\item When the length of \code{alpha} is three: the plot points opacity are dynamically adjusted based on
vector of \code{value}, \code{min}, and \code{max}.
}}

\item{shape}{(\code{character}) optional, character vector with the names of the
shape, e.g. \code{c("triangle", "square", "circle")}. It defaults to \code{shape_names}. This is a complete list from
\code{vignette("ggplot2-specs", package="ggplot2")}.}

\item{size}{(\code{integer(1)} or \code{integer(3)}) optional, specifies point size.
\itemize{
\item When the length of \code{size} is one: the plot point sizes will have a fixed size.
\item When the length of \code{size} is three: the plot points size are dynamically adjusted based on
vector of \code{value}, \code{min}, and \code{max}.
}}

\item{max_deg}{(\code{integer}) optional, maximum degree for the polynomial trend line. Must not be less than 1.}

\item{rotate_xaxis_labels}{(\code{logical}) optional, whether to rotate plot X axis labels. Does not
rotate by default (\code{FALSE}).}

\item{ggtheme}{(\code{character}) optional, \code{ggplot2} theme to be used by default. Defaults to \code{"gray"}.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{table_dec}{(\code{integer}) optional, number of decimal places used to round numeric values in the table.}

\item{ggplot2_args}{(\code{ggplot2_args}) object created by \code{\link[teal.widgets:ggplot2_args]{teal.widgets::ggplot2_args()}}
with settings for the module plot.
The argument is merged with options variable \code{teal.ggplot2_args} and default module setup.

For more details see the vignette: \code{vignette("custom-ggplot2-arguments", package = "teal.widgets")}}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Generates a customizable scatterplot using \code{ggplot2}.
This module allows users to select variables for the x and y axes,
color and size encodings, faceting options, and more. It supports log transformations,
trend line additions, and dynamic adjustments of point opacity and size through UI controls.
}
\note{
For more examples, please see the vignette "Using scatterplot" via
\code{vignette("using-scatterplot", package = "teal.modules.general")}.
}
\examples{
\dontshow{if (require("ggpmisc", quietly = TRUE) && require("ggExtra", quietly = TRUE) && require("colourpicker", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# general data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  CO2 <- CO2
})
datanames(data) <- "CO2"

app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplot(
      label = "Scatterplot Choices",
      x = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["CO2"]], c("conc", "uptake")),
          selected = "conc",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["CO2"]], c("conc", "uptake")),
          selected = "uptake",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      color_by = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(
            data[["CO2"]],
            c("Plant", "Type", "Treatment", "conc", "uptake")
          ),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      size_by = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["CO2"]], c("conc", "uptake")),
          selected = "uptake",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      row_facet = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["CO2"]], c("Plant", "Type", "Treatment")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      col_facet = data_extract_spec(
        dataname = "CO2",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["CO2"]], c("Plant", "Type", "Treatment")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
\dontshow{\}) # examplesIf}
\dontshow{if (require("ggpmisc", quietly = TRUE) && require("ggExtra", quietly = TRUE) && require("colourpicker", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# CDISC data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  ADSL <- rADSL
})
datanames(data) <- c("ADSL")
join_keys(data) <- default_cdisc_join_keys[datanames(data)]

app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplot(
      label = "Scatterplot Choices",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], c("AGE", "BMRKR1", "BMRKR2")),
          selected = "AGE",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], c("AGE", "BMRKR1", "BMRKR2")),
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      color_by = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(
            data[["ADSL"]],
            c("AGE", "BMRKR1", "BMRKR2", "RACE", "REGION1")
          ),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      size_by = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], c("AGE", "BMRKR1")),
          selected = "AGE",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      row_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], c("BMRKR2", "RACE", "REGION1")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      col_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], c("BMRKR2", "RACE", "REGION1")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
\dontshow{\}) # examplesIf}
}
\section{Examples in Shinylive}{
  \itemize{
    \item{example-1}{\cr
      \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNrv2CsBISqDC0ogT+uLpetHCk1Ky6ALz2Ts62AGTpuu7Rvv6BzgAepMwRUd6x8UkpLhlZOd55hETURIKMqLQEANYyZdGVCckOtUoAxLpSEGrUuvxQpFC6cIWwqCJK84sGxlzUAPpbUDabC0tGugDutKQAFirsR5EgSrrZno1w7NOipMStjEUEDeAGEAPIAJh2unBEKUAF8gUdoPBRI8zrYLv5Yf4lEo0KhoSobn5gXMztUnq9dHwhCJRNVacIxKS3m9SDB9hJ9mEFuQOq1SKy2W9qFB6HBZsl-ABlAh8mTrIikGG3IhdMQRakiwqUs77FYldSkHmoOAEYUit7I2B6aVgHF4bVW0SS80q5KukQaU3my1W0XiyXVWVujS6LQsWjikSILVkgNvAhqjUM5KRxjR+gifbJ9UEFlHYDAbGQ-wAXXLkQt-mIFjK-kEqEWvX8cjkuGdAa97rg-BDzXrToTiZgwk06ztugAYgBBAAyMucnZHAZMtEKfeqc8XrlXbKBAY7Xd0Q3JiwNxWYPtEZpr++tZxRU9LcOHid0PfDnrDJtvfpPEUxQlKVdFDb0VQzLNY3jD8kxTAs0wjKMYzgXMEMLM5i1fCsq10GtB3CPAwLAJsWzgNtjwfNkv3Ift7TIqBW3fOCx2oCcRG3BclxXODTA3Ldkh3JdAN0Q8rSogN-iIRh9noM8jkvI0bzvf0RRteAB0dXjE1o6paN9e8+OA4N7RlX9kMzVC4xYj881TaooNQ9D8xZUT1KwksHTLMBKx0viCIABTFMgGzAOxWDNMK7HcBZ4FC4jaxIIjIkbZsmIosBxI-SS4NowTdAAOUced5380dx1oScuN3cq1wE+iZ24vccpPXKRVEWgAC80PkvUL0Na8-1U0SNJfby3zqmiLJ-CDDLUq0TNA8D3Us6C4BsqaRXsxDHJQ7M0J2zDFmwibcOrPxCLCxjmPbLbpoggq0vI2DWMq6qhOa+63nXTdGuElrE2yg8psYIhLn2Ex1FifqoCUob5tGp9bS0nz7r02b3UR6jAxAgdzIgtbrNeuyMKQpyDpc1N0ROrzHT8-DLuCqAEtS8LIsytmYrgOKKFISjvs-X8CuK0rBbYjipwBwXfoKgHROBt52rZf5Iehj1zzhwbjWxj8xtRybRIxoW5v-Iy4KW-GLIpmDbMTI7yf2nMHZpqBTvpvCgpC-nEvZqLfe53mEru9yTd7RrRbK0OJaqzjPtq0PZf+5qFba51xKBIFaBMXR2BUfljW0L5210F4yVEe4IFYWd0HYAkABJvEiBvXUYHRAQRJQwHhcsgA}{Open in Shinylive}\cr
      \if{html}{
        \out{          <div class = "iframe-wrapper" style="height: 800px; width: 100\%">            <script type="text/javascript">$(function() {  var if_pkgdown = [...document.scripts].filter(x => x.src.includes("pkgdown.js")).length > 0;  if (if_pkgdown) {    $(".iframe-wrapper").css("width", "140\%");  }});            </script>            <iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNrv2CsBISqDC0ogT+uLpetHCk1Ky6ALz2Ts62AGTpuu7Rvv6BzgAepMwRUd6x8UkpLhlZOd55hETURIKMqLQEANYyZdGVCckOtUoAxLpSEGrUuvxQpFC6cIWwqCJK84sGxlzUAPpbUDabC0tGugDutKQAFirsR5EgSrrZno1w7NOipMStjEUEDeAGEAPIAJh2unBEKUAF8gUdoPBRI8zrYLv5Yf4lEo0KhoSobn5gXMztUnq9dHwhCJRNVacIxKS3m9SDB9hJ9mEFuQOq1SKy2W9qFB6HBZsl-ABlAh8mTrIikGG3IhdMQRakiwqUs77FYldSkHmoOAEYUit7I2B6aVgHF4bVW0SS80q5KukQaU3my1W0XiyXVWVujS6LQsWjikSILVkgNvAhqjUM5KRxjR+gifbJ9UEFlHYDAbGQ-wAXXLkQt-mIFjK-kEqEWvX8cjkuGdAa97rg-BDzXrToTiZgwk06ztugAYgBBAAyMucnZHAZMtEKfeqc8XrlXbKBAY7Xd0Q3JiwNxWYPtEZpr++tZxRU9LcOHid0PfDnrDJtvfpPEUxQlKVdFDb0VQzLNY3jD8kxTAs0wjKMYzgXMEMLM5i1fCsq10GtB3CPAwLAJsWzgNtjwfNkv3Ift7TIqBW3fOCx2oCcRG3BclxXODTA3Ldkh3JdAN0Q8rSogN-iIRh9noM8jkvI0bzvf0RRteAB0dXjE1o6paN9e8+OA4N7RlX9kMzVC4xYj881TaooNQ9D8xZUT1KwksHTLMBKx0viCIABTFMgGzAOxWDNMK7HcBZ4FC4jaxIIjIkbZsmIosBxI-SS4NowTdAAOUced5380dx1oScuN3cq1wE+iZ24vccpPXKRVEWgAC80PkvUL0Na8-1U0SNJfby3zqmiLJ-CDDLUq0TNA8D3Us6C4BsqaRXsxDHJQ7M0J2zDFmwibcOrPxCLCxjmPbLbpoggq0vI2DWMq6qhOa+63nXTdGuElrE2yg8psYIhLn2Ex1FifqoCUob5tGp9bS0nz7r02b3UR6jAxAgdzIgtbrNeuyMKQpyDpc1N0ROrzHT8-DLuCqAEtS8LIsytmYrgOKKFISjvs-X8CuK0rBbYjipwBwXfoKgHROBt52rZf5Iehj1zzhwbjWxj8xtRybRIxoW5v-Iy4KW-GLIpmDbMTI7yf2nMHZpqBTvpvCgpC-nEvZqLfe53mEru9yTd7RrRbK0OJaqzjPtq0PZf+5qFba51xKBIFaBMXR2BUfljW0L5210F4yVEe4IFYWd0HYAkABJvEiBvXUYHRAQRJQwHhcsgA" allow="fullscreen" scrolling="auto" style="height: 100\%; width: 100\%; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem"></iframe>          </div>        }
      }
    }
    \item{example-2}{\cr
      \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNrv2CsBISqDC0ogT+uLpetHCk1Ky6ALz2Ts62AGTpuu7Rvv6BzgAepMwRUd6x8UkpLhlZOd55hETURIKMqLQEANYyZdGVCckOtUoAxLoAwgAiAJIAypO6-FCkULpwhbCoIkorawbGXNQA+vtQNnur60a6AO60pAAWKuznkSBKutmejXDsEDEpGIrUYigg3wAgtN5gAZQ7ZaFwpQAX3B52g8FEb2utluBD8YCRsP84IAVkQVCdeqxsec8cZ+HATFBhKQTgR+KECCcKVSaaJgBjYGIcWs5ABdJRKNCoBEqR5+CHLa7Vd5fXR8IQiUTVLXCUUa76kGAnCQnMKrcgdVqkJXfB26ahQehwajVfzzAhWmQ7IikKZPSkEMQRI0Owpq64nTYldTs0SoOAE8OO4XwD1EmEkvCph2iN3JgPJAsiDQWpMp5WOx3O13u5KewsaXRaFi0F0iRBh6s174EINdMTVNuMDv0EQcwchunXYDAfzE-wSiWRFNEgDirjwun8ACEALJYADSWAAjGV90fTwAmUlyXB5x2lotwfiZyFbnt9x0wNm0HY9GSAAxSFYXmZxH17PsTFoQo32qUDwNcaCHXBPsHzzIYVTWGNimYctE2Te0f3TIDdyzZFc1Q74XxbEtmwTSsSJ-J0XTdTN5kY1t207OBu2o1j+2nYdklHcdJwHYNRXOedF2zZdV10ddP23SIrxPc9LzAQ9NLvMA5Ewmj80YhDGx068tMEoS-2oTRAMQsCIKgoTTDgszdCQiCn10dCayMvsQSIRgTnobDzjwuNCOYnyyI-BTrL7Oji10ZKK2InzvjrDjzK4ssA3EviBJcoSpKHXUxN4ic4CnaTsUyh1ZIXSicxXErXJUr8dw008L26iy9O0rBIUmNSKKwZwN1mAB5AA5Pq-NYgKhOSjzZscWFYXan9bPskRHOQ7aYPc98QKclClrzZbn1oAAvGqwqjXDYwIpiMuMuLzKXRKa2S6o0qIqtXOyhsKLyoseLHIrv1KkSKshiSarKmcxSgOSWsUtdCVU7TdN6+8jt+0zToonGfr7XaAP2s7Doa2D4JJryLp-Ra0MJxgiDuE4WRDFKIpe+N0qB0jrkxcj5KownaO4hj8qFlifxBzjuMK6rioa5HRIRvjavK1H0e+trlMJPGsH09SwGG0ahsmmb5oJhrVpJ9bNqlh1KYcmnnLpk6Du84zWe+a6HRBbn1FiJ6oEi175di0WRXiyWfL+2Wi1j4ysvY0Gm3y7W1Zh1jNfh1XJLh-XmsNpT11N83xpGsb-Amqa5oW4OkuJ6oXa2hqPepzzzrd756Y8pmfMD3zwz88FwVoExdHYFRrXjbR-kM3RPmVUQXggVhIXQdhZQAEm8SIj4LRgdDBVElDAFEJSAA}{Open in Shinylive}\cr
      \if{html}{
        \out{          <div class = "iframe-wrapper" style="height: 800px; width: 100\%">            <script type="text/javascript">$(function() {  var if_pkgdown = [...document.scripts].filter(x => x.src.includes("pkgdown.js")).length > 0;  if (if_pkgdown) {    $(".iframe-wrapper").css("width", "140\%");  }});            </script>            <iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNrv2CsBISqDC0ogT+uLpetHCk1Ky6ALz2Ts62AGTpuu7Rvv6BzgAepMwRUd6x8UkpLhlZOd55hETURIKMqLQEANYyZdGVCckOtUoAxLoAwgAiAJIAypO6-FCkULpwhbCoIkorawbGXNQA+vtQNnur60a6AO60pAAWKuznkSBKutmejXDsEDEpGIrUYigg3wAgtN5gAZQ7ZaFwpQAX3B52g8FEb2utluBD8YCRsP84IAVkQVCdeqxsec8cZ+HATFBhKQTgR+KECCcKVSaaJgBjYGIcWs5ABdJRKNCoBEqR5+CHLa7Vd5fXR8IQiUTVLXCUUa76kGAnCQnMKrcgdVqkJXfB26ahQehwajVfzzAhWmQ7IikKZPSkEMQRI0Owpq64nTYldTs0SoOAE8OO4XwD1EmEkvCph2iN3JgPJAsiDQWpMp5WOx3O13u5KewsaXRaFi0F0iRBh6s174EINdMTVNuMDv0EQcwchunXYDAfzE-wSiWRFNEgDirjwun8ACEALJYADSWAAjGV90fTwAmUlyXB5x2lotwfiZyFbnt9x0wNm0HY9GSAAxSFYXmZxH17PsTFoQo32qUDwNcaCHXBPsHzzIYVTWGNimYctE2Te0f3TIDdyzZFc1Q74XxbEtmwTSsSJ-J0XTdTN5kY1t207OBu2o1j+2nYdklHcdJwHYNRXOedF2zZdV10ddP23SIrxPc9LzAQ9NLvMA5Ewmj80YhDGx068tMEoS-2oTRAMQsCIKgoTTDgszdCQiCn10dCayMvsQSIRgTnobDzjwuNCOYnyyI-BTrL7Oji10ZKK2InzvjrDjzK4ssA3EviBJcoSpKHXUxN4ic4CnaTsUyh1ZIXSicxXErXJUr8dw008L26iy9O0rBIUmNSKKwZwN1mAB5AA5Pq-NYgKhOSjzZscWFYXan9bPskRHOQ7aYPc98QKclClrzZbn1oAAvGqwqjXDYwIpiMuMuLzKXRKa2S6o0qIqtXOyhsKLyoseLHIrv1KkSKshiSarKmcxSgOSWsUtdCVU7TdN6+8jt+0zToonGfr7XaAP2s7Doa2D4JJryLp-Ra0MJxgiDuE4WRDFKIpe+N0qB0jrkxcj5KownaO4hj8qFlifxBzjuMK6rioa5HRIRvjavK1H0e+trlMJPGsH09SwGG0ahsmmb5oJhrVpJ9bNqlh1KYcmnnLpk6Du84zWe+a6HRBbn1FiJ6oEi175di0WRXiyWfL+2Wi1j4ysvY0Gm3y7W1Zh1jNfh1XJLh-XmsNpT11N83xpGsb-Amqa5oW4OkuJ6oXa2hqPepzzzrd756Y8pmfMD3zwz88FwVoExdHYFRrXjbR-kM3RPmVUQXggVhIXQdhZQAEm8SIj4LRgdDBVElDAFEJSAA" allow="fullscreen" scrolling="auto" style="height: 100\%; width: 100\%; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem"></iframe>          </div>        }
      }
    }
  }
}

