% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplot.R
\name{tm_g_scatterplot}
\alias{tm_g_scatterplot}
\title{\code{teal} module: Scatterplot}
\usage{
tm_g_scatterplot(
  label = "Scatterplot",
  x,
  y,
  color_by = NULL,
  size_by = NULL,
  row_facet = NULL,
  col_facet = NULL,
  plot_height = c(600, 200, 2000),
  plot_width = NULL,
  alpha = c(1, 0, 1),
  shape = shape_names,
  size = c(5, 1, 15),
  max_deg = 5L,
  rotate_xaxis_labels = FALSE,
  ggtheme = c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void"),
  pre_output = NULL,
  post_output = NULL,
  table_dec = 4,
  ggplot2_args = teal.widgets::ggplot2_args(),
  transformators = list(),
  decorators = list()
)
}
\arguments{
\item{x}{(\code{picks} or \code{list} of multiple \code{picks}) Specifies
variable names selected to plot along the x-axis by default.}

\item{y}{(\code{picks} or \code{list} of multiple \code{picks}) Specifies
variable names selected to plot along the y-axis by default.}

\item{color_by}{(\code{picks} or \code{list} of multiple \code{picks}) optional,
defines the color encoding. If \code{NULL} then no color encoding option will be displayed.}

\item{size_by}{(\code{picks} or \code{list} of multiple \code{picks}) optional,
defines the point size encoding. If \code{NULL} then no size encoding option will be displayed.}

\item{row_facet}{(\code{picks} or \code{list} of multiple \code{picks}) optional,
specifies the variable(s) for faceting rows.}

\item{col_facet}{(\code{picks} or \code{list} of multiple \code{picks}) optional,
specifies the variable(s) for faceting columns.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{plot_width}{(\code{numeric}) optional, specifies the plot width as a three-element vector of
\code{value}, \code{min}, and \code{max} for a slider encoding the plot width.}

\item{alpha}{(\code{integer(1)} or \code{integer(3)}) optional, specifies point opacity.
\itemize{
\item When the length of \code{alpha} is one: the plot points will have a fixed opacity.
\item When the length of \code{alpha} is three: the plot points opacity are dynamically adjusted based on
vector of \code{value}, \code{min}, and \code{max}.
}}

\item{shape}{(\code{character}) optional, character vector with the names of the
shape, e.g. \code{c("triangle", "square", "circle")}. It defaults to \code{shape_names}. This is a complete list from
\code{vignette("ggplot2-specs", package="ggplot2")}.}

\item{size}{(\code{integer(1)} or \code{integer(3)}) optional, specifies point size.
\itemize{
\item When the length of \code{size} is one: the plot point sizes will have a fixed size.
\item When the length of \code{size} is three: the plot points size are dynamically adjusted based on
vector of \code{value}, \code{min}, and \code{max}.
}}

\item{max_deg}{(\code{integer}) optional, maximum degree for the polynomial trend line. Must not be less than 1.}

\item{rotate_xaxis_labels}{(\code{logical}) optional, whether to rotate plot X axis labels. Does not
rotate by default (\code{FALSE}).}

\item{ggtheme}{(\code{character}) optional, \code{ggplot2} theme to be used by default. Defaults to \code{"gray"}.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{table_dec}{(\code{integer}) optional, number of decimal places used to round numeric values in the table.}

\item{ggplot2_args}{(\code{ggplot2_args}) object created by \code{\link[teal.widgets:ggplot2_args]{teal.widgets::ggplot2_args()}}
with settings for the module plot.
The argument is merged with options variable \code{teal.ggplot2_args} and default module setup.

For more details see the vignette: \code{vignette("custom-ggplot2-arguments", package = "teal.widgets")}}

\item{decorators}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
(named \code{list} of lists of \code{teal_transform_module}) optional,
decorator for tables or plots included in the module output reported.
The decorators are applied to the respective output objects.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Generates a customizable scatterplot using \code{ggplot2}.
This module allows users to select variables for the x and y axes,
color and size encodings, faceting options, and more. It supports log transformations,
trend line additions, and dynamic adjustments of point opacity and size through UI controls.
}
\note{
For more examples, please see the vignette "Using scatterplot" via
\code{vignette("using-scatterplot", package = "teal.modules.general")}.
}
\section{Decorating Module}{


This module generates the following objects, which can be modified in place using decorators:
\itemize{
\item \code{plot} (\code{ggplot})
}

A Decorator is applied to the specific output using a named list of \code{teal_transform_module} objects.
The name of this list corresponds to the name of the output to which the decorator is applied.
See code snippet below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tm_g_scatterplot(
   ..., # arguments for module
   decorators = list(
     plot = teal_transform_module(...) # applied to the `plot` output
   )
)
}\if{html}{\out{</div>}}

For additional details and examples of decorators, refer to the vignette
\code{vignette("decorate-module-output", package = "teal.modules.general")}.

To learn more please refer to the vignette
\code{vignette("transform-module-output", package = "teal")} or the \code{\link[teal:teal_transform_module]{teal::teal_transform_module()}} documentation.
}

\examples{
# general data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  CO2 <- CO2
})

app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplot(
      label = "Scatterplot Choices",
      x = picks(
        datasets("CO2"),
        variables(
          choices = c("conc", "uptake"),
          selected = "conc"
        ),
        values()
      ),
      y = picks(
        datasets("CO2"),
        variables(
          choices = c("conc", "uptake"),
          selected = "uptake"
        ),
        values()
      ),
      color_by = picks(
        datasets("CO2"),
        variables(
          choices = c("Plant", "Type", "Treatment", "conc", "uptake"),
          selected = NULL
        ),
        values()
      ),
      size_by = picks(
        datasets("CO2"),
        variables(choices = c("conc", "uptake"), selected = "uptake"),
        values()
      ),
      row_facet = picks(
        datasets("CO2"),
        variables(
          choices = c("Plant", "Type", "Treatment"),
          selected = NULL
        ),
        values()
      ),
      col_facet = picks(
        datasets("CO2"),
        variables(choices = c("Plant", "Type", "Treatment"), selected = NULL),
        values()
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# CDISC data example
data <- teal_data()
data <- within(data, {
  require(nestcolor)
  ADSL <- teal.data::rADSL
})
join_keys(data) <- default_cdisc_join_keys[names(data)]

app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplot(
      label = "Scatterplot Choices",
      x = picks(
        datasets("ADSL"),
        variables(choices = c("AGE", "BMRKR1", "BMRKR2"), selected = "AGE"),
        values()
      ),
      y = picks(
        datasets("ADSL"),
        variables(choices = c("AGE", "BMRKR1", "BMRKR2"), selected = "BMRKR1"),
        values()
      ),
      color_by = picks(
        datasets("ADSL"),
        variables(c("AGE", "BMRKR1", "BMRKR2", "RACE", "REGION1"), selected = NULL),
        values()
      ),
      size_by = picks(
        datasets("ADSL"),
        variables(choices = c("AGE", "BMRKR1"), selected = "AGE"),
        values()
      ),
      row_facet = picks(
        datasets("ADSL"),
        variables(choices = c("BMRKR2", "RACE", "REGION1"), selected = NULL),
        values()
      ),
      col_facet = picks(
        datasets("ADSL"),
        variables(choices = c("BMRKR2", "RACE", "REGION1"), selected = NULL),
        values()
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
\section{Examples in Shinylive}{
\describe{
  \item{example-1}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6uoxwAI6CtNXsPqKkxNREjIoQlQDCAPIATHG6-QNKAL5dSmiowyp57BX+GboAvMtBuEt8QiKia7o7wmKL3ZW6pDCJEomiBIHkjJFEpKfn59RQ9HB+6wpgAGV7qRHs9SCN8kRaAQxP8tmd3qEDqhoQBrURvd7nFKiOCkDH-Ub-OTwrHnLQsWhfPaYsmVAiQ6FiA4ERaEEgEOG6f6CVBBVFwYmkumVXEiDRwfgHf7ECz-JZkkkKrEU6iCE5dRXC96sZFogkIrE4vEEsBEsBKw3vCmMKn0GnKskMqEw-brVkyjlcnl8qACoWOrFiuASqV-MC8-mCsCByqWkWq9UYzVY+NYtodRL0XXrFEEdG0o0ZXH4tnmtNkm12h1W9OM10stkABU+ZG9YDsrFQ0dK-zs1UC8DbeG57LlI59UYDtfewdDBwAco4ADLL2O6Csq7hJtK1zei2gALzgWZzujzBfXxtLhMG04TlOpJ2dTLdug9Y85E4jvv9FtKc7kGGo6Rn60b7uS24aoGEGMEQ2SJCY6h4nq+YGiK16muW2pbraT7oSK771sy7rNq2pDtp23aUQOlwUBR-7rqKPwhkBi4rmuM4bjh1pQcmME8e+RBJEhMLgrm+qFu8mFlnejFcVW+HsC+Dakf8LZQMOvYdl2Pajv2cCDvRQq6IBkrsauEGVIm0F7sqKZdF0tAmLo7AqI86iaDoNi2OUZyiIUECsAAgug7AzAAJPUpSRbijA6J0ExKGA4wALpAA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6uoxwAI6CtNXsPqKkxNREjIoQlQDCAPIATHG6-QNKAL5dSmiowyp57BX+GboAvMtBuEt8QiKia7o7wmKL3ZW6pDCJEomiBIHkjJFEpKfn59RQ9HB+6wpgAGV7qRHs9SCN8kRaAQxP8tmd3qEDqhoQBrURvd7nFKiOCkDH-Ub-OTwrHnLQsWhfPaYsmVAiQ6FiA4ERaEEgEOG6f6CVBBVFwYmkumVXEiDRwfgHf7ECz-JZkkkKrEU6iCE5dRXC96sZFogkIrE4vEEsBEsBKw3vCmMKn0GnKskMqEw-brVkyjlcnl8qACoWOrFiuASqV-MC8-mCsCByqWkWq9UYzVY+NYtodRL0XXrFEEdG0o0ZXH4tnmtNkm12h1W9OM10stkABU+ZG9YDsrFQ0dK-zs1UC8DbeG57LlI59UYDtfewdDBwAco4ADLL2O6Csq7hJtK1zei2gALzgWZzujzBfXxtLhMG04TlOpJ2dTLdug9Y85E4jvv9FtKc7kGGo6Rn60b7uS24aoGEGMEQ2SJCY6h4nq+YGiK16muW2pbraT7oSK771sy7rNq2pDtp23aUQOlwUBR-7rqKPwhkBi4rmuM4bjh1pQcmME8e+RBJEhMLgrm+qFu8mFlnejFcVW+HsC+Dakf8LZQMOvYdl2Pajv2cCDvRQq6IBkrsauEGVIm0F7sqKZdF0tAmLo7AqI86iaDoNi2OUZyiIUECsAAgug7AzAAJPUpSRbijA6J0ExKGA4wALpAA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
  \item{example-2}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXUY4AEdBWgb2CDFSYmoiRkUIOoBBHwCAGWTdVIxMxERGYbGlAF9+gCsiFTSAazhWUQrc23z+OBMoYVI0gn5aUQI09c2dveBoeH3MuQBdN2h0CZUxXYtRCuV0AF5QeFcCC+EIRKIIbo4cIxMCBnVJjA0hI0ncwuRGHEiKR0ZjMdQoPQ4NQkQowAECASZMTSN4ShsCGJ6TCMeSokjULQCFt9iDyXVMqI4KQxWAFqN6XJeRLMVoWLQqQj2AQOcKxEiCMD5QBxVx4XT0gBCAFksABpLAARh5lrAtodWAATEqqtKRBo4Pw6abzcrxRL1dRBGj+qrw3zMaxBcLRWTVVCoNLZcaFb6I+T1YxNfRtbrOQbIUb6YMza7rXbHS6LQ3PT6wMrdP64IHg5DW0384nC9wY-s4xKE6rur00vRk5ChSKxcPMVKZXK8x2VRmiyWy7m6y33Y3nfWT23z1hBl5zVV6VhnCa-AB5AByzc73d7SLfjlGoxTruo6xgWuhAeSoi0AAXnAc4LroS5pmBkq5Nmm4jIq24oboe5ami5b6oiVaHneboeoO25djSPbkH2bq1mGO6qlGY7ZKuEGYowRAFGkZxcmyi6piuGaZuhuaYUOol4aWBF6lyxG6NWF6Ou295gNet5Xk+r4fr61EBnRv7-oBzGRiB45gZxdTdHx6gyimy7pqq645jWknYaudQyWW8mVkpxoUd6V43mRD46e+n5+jRP6Qn+AHWWqFnsfGEYTv0-S0CYujsCohLqJoOg2LYNQYqIZQQKwgzoOwaCoAAJC0VR1fV0qMDofTLEoYBLF8QA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXUY4AEdBWgb2CDFSYmoiRkUIOoBBHwCAGWTdVIxMxERGYbGlAF9+gCsiFTSAazhWUQrc23z+OBMoYVI0gn5aUQI09c2dveBoeH3MuQBdN2h0CZUxXYtRCuV0AF5QeFcCC+EIRKIIbo4cIxMCBnVJjA0hI0ncwuRGHEiKR0ZjMdQoPQ4NQkQowAECASZMTSN4ShsCGJ6TCMeSokjULQCFt9iDyXVMqI4KQxWAFqN6XJeRLMVoWLQqQj2AQOcKxEiCMD5QBxVx4XT0gBCAFksABpLAARh5lrAtodWAATEqqtKRBo4Pw6abzcrxRL1dRBGj+qrw3zMaxBcLRWTVVCoNLZcaFb6I+T1YxNfRtbrOQbIUb6YMza7rXbHS6LQ3PT6wMrdP64IHg5DW0384nC9wY-s4xKE6rur00vRk5ChSKxcPMVKZXK8x2VRmiyWy7m6y33Y3nfWT23z1hBl5zVV6VhnCa-AB5AByzc73d7SLfjlGoxTruo6xgWuhAeSoi0AAXnAc4LroS5pmBkq5Nmm4jIq24oboe5ami5b6oiVaHneboeoO25djSPbkH2bq1mGO6qlGY7ZKuEGYowRAFGkZxcmyi6piuGaZuhuaYUOol4aWBF6lyxG6NWF6Ou295gNet5Xk+r4fr61EBnRv7-oBzGRiB45gZxdTdHx6gyimy7pqq645jWknYaudQyWW8mVkpxoUd6V43mRD46e+n5+jRP6Qn+AHWWqFnsfGEYTv0-S0CYujsCohLqJoOg2LYNQYqIZQQKwgzoOwaCoAAJC0VR1fV0qMDofTLEoYBLF8QA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
}
}

