% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{tm_g_scatterplotmatrix}
\alias{tm_g_scatterplotmatrix}
\title{\code{teal} module: Scatterplot matrix}
\usage{
tm_g_scatterplotmatrix(
  label = "Scatterplot Matrix",
  variables = list(teal.picks::picks(teal.picks::datasets(),
    teal.picks::variables(selected = seq(1L, 5L), multiple = TRUE))),
  min_n_variables = 2L,
  max_n_variables = 5L,
  plot_height = c(600, 200, 2000),
  plot_width = NULL,
  pre_output = NULL,
  post_output = NULL,
  transformators = list(),
  decorators = list()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{variables}{(\code{picks} or \code{list} of \code{picks})
Specifies plotting variables from an incoming dataset with filtering and selecting. In case of
\code{picks} use \code{teal.picks::variables(..., ordered = TRUE)} if plot elements should be
rendered according to selection order.}

\item{min_n_variables}{(\code{integer(1)})
Minimum number of variables that must be selected.}

\item{max_n_variables}{(\code{integer(1)})
Maximum number of variables that can be selected.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{plot_width}{(\code{numeric}) optional, specifies the plot width as a three-element vector of
\code{value}, \code{min}, and \code{max} for a slider encoding the plot width.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{transformators}{(\code{list} of \code{teal_transform_module}) that will be applied to transform module's data input.
To learn more check \code{vignette("transform-input-data", package = "teal")}.}

\item{decorators}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
(named \code{list} of lists of \code{teal_transform_module}) optional,
decorator for tables or plots included in the module output reported.
The decorators are applied to the respective output objects.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Generates a scatterplot matrix from selected \code{variables} from datasets.
Each plot within the matrix represents the relationship between two variables,
providing the overview of correlations and distributions across selected data.
}
\note{
For more examples, please see the vignette "Using scatterplot matrix" via
\code{vignette("using-scatterplot-matrix", package = "teal.modules.general")}.

When \emph{Add Correlation} is enabled, a simple \strong{Omit NAs} checkbox
controls NA handling (checked = \code{"pairwise.complete.obs"}, matching the
historical default).  Unchecking it reveals a dropdown with all five
\code{stats::cor()} \code{use} options for advanced control.
}
\section{Decorating Module}{


This module generates the following objects, which can be modified in place using decorators:
\itemize{
\item \code{plot} (\code{ggplot} - a \code{patchwork} assembled from individual \code{ggplot} panels)
}

A Decorator is applied to the specific output using a named list of \code{teal_transform_module} objects.
The name of this list corresponds to the name of the output to which the decorator is applied.
See code snippet below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tm_g_scatterplotmatrix(
   ..., # arguments for module
   decorators = list(
     plot = teal_transform_module(...) # applied to the `plot` output
   )
)
}\if{html}{\out{</div>}}

For additional details and examples of decorators, refer to the vignette
\code{vignette("decorate-module-output", package = "teal.modules.general")}.

To learn more please refer to the vignette
\code{vignette("transform-module-output", package = "teal")} or the \code{\link[teal:teal_transform_module]{teal::teal_transform_module()}} documentation.
}

\section{Reporting}{



This module returns an object of class \code{teal_module}, that contains a \code{server} function.
Since the server function returns a \code{teal_report} object, this makes this module reportable, which means that
the reporting functionality will be turned on automatically by the \code{teal} framework.

For more information on reporting in \code{teal}, see the vignettes:
\itemize{
\item \code{vignette("reportable-shiny-application", package = "teal.reporter")}
\item \code{vignette("adding-support-for-reporting-to-custom-modules", package = "teal")}
}

}

\examples{
# general data example
data <- teal_data()
data <- within(data, {
  countries <- data.frame(
    id = c("DE", "FR", "IT", "ES", "PT", "GR", "NL", "BE", "LU", "AT"),
    government = factor(
      c(2, 2, 2, 1, 2, 2, 1, 1, 1, 2),
      labels = c("Monarchy", "Republic")
    ),
    language_family = factor(
      c(1, 3, 3, 3, 3, 2, 1, 1, 3, 1),
      labels = c("Germanic", "Hellenic", "Romance")
    ),
    population = c(83, 67, 60, 47, 10, 11, 17, 11, 0.6, 9),
    area = c(357, 551, 301, 505, 92, 132, 41, 30, 2.6, 83),
    gdp = c(3.4, 2.7, 2.1, 1.4, 0.3, 0.2, 0.7, 0.5, 0.1, 0.4),
    debt = c(2.1, 2.3, 2.4, 2.6, 2.3, 2.4, 2.3, 2.4, 2.3, 2.4)
  )
  sales <- data.frame(
    id = 1:50,
    country_id = sample(
      c("DE", "FR", "IT", "ES", "PT", "GR", "NL", "BE", "LU", "AT"),
      size = 50,
      replace = TRUE
    ),
    year = sort(sample(2010:2020, 50, replace = TRUE)),
    venue = sample(c("small", "medium", "large", "online"), 50, replace = TRUE),
    cancelled = sample(c(TRUE, FALSE), 50, replace = TRUE),
    quantity = rnorm(50, 100, 20),
    costs = rnorm(50, 80, 20),
    profit = rnorm(50, 20, 10)
  )
})
join_keys(data) <- join_keys(
  join_key("countries", "countries", "id"),
  join_key("sales", "sales", "id"),
  join_key("countries", "sales", c("id" = "country_id"))
)

app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplotmatrix(
      label = "Scatterplot matrix",
      variables = list(
        teal.picks::picks(
          datasets("countries"),
          teal.picks::variables(
            choices = tidyselect::everything(),
            selected = c("area", "gdp", "debt"),
            multiple = TRUE,
            ordered = TRUE
          ),
          teal.picks::values()
        ),
        teal.picks::picks(
          datasets("sales"),
          teal.picks::variables(
            choices = c("quantity", "costs", "profit"),
            selected = c("quantity", "costs", "profit"),
            multiple = TRUE,
            ordered = TRUE
          )
        )
      ),
      transformators = list(
        teal_transform_filter(
          teal.picks::picks(
            datasets("sales"),
            teal.picks::variables("country_id"),
            teal.picks::values()
          )
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# CDISC data example
data <- teal_data()
data <- within(data, {
  ADSL <- teal.data::rADSL
  ADRS <- teal.data::rADRS
})
join_keys(data) <- default_cdisc_join_keys[names(data)]

app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplotmatrix(
      label = "Scatterplot matrix",
      variables = list(
        teal.picks::picks(
          datasets("ADSL"),
          teal.picks::variables(
            choices = tidyselect::everything(),
            selected = c("AGE", "RACE", "SEX"),
            multiple = TRUE,
            ordered = TRUE,
            fixed = FALSE
          ),
          teal.picks::values()
        ),
        teal.picks::picks(
          datasets("ADRS"),
          teal.picks::variables(
            choices = tidyselect::everything(),
            selected = c("AGE", "AVAL", "ADY"),
            multiple = TRUE,
            ordered = TRUE,
            fixed = FALSE
          )
        )
      ),
      transformators = list(
        teal_transform_filter(
          teal.picks::picks(
            teal.picks::datasets("ADRS"),
            teal.picks::variables("PARAMCD"),
            teal.picks::values(selected = "BESRSPI")
          )
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
\section{Examples in Shinylive}{
\describe{
  \item{example-1}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6usTmpIy0YnH+GRgmzPDsFZW6tPy6ALxVHWAAIq54ugpgAGJYk6WTAJJ2cxNgzgDKK5MACsvjkwDis-tgAHIAMltgAEJj82DnjlcAgntyuJ2VEkQ6jBDwZH6pnUpCIjA6EC6XQI7AATKV4bpEQBGBFo3SojGlTGw96fLrUKD0ODUURAmGTACyJBYBHyrCuWDgqEE9DoBEmikhXTx3MqhIgEkEUCkiRMsFo1FYQPFGjBEKhlRhmIAzKU1boNRqUdj1djeYrdITiaTyUMDjIYFBlByTgAJEkiG2MohWixwTn4g1dVBEFmEzQkM0ADg1ADYAOylMMABlKABYoxi4xjMcik8jMTGMGHSgBOb2VFhwYIDGEqgCsSYrFdVMcxFZjFfzKJViPjdYROdKocLXn4qDNKow8a7SdhGDTI9K2Y12cR2aT2ebumzWZHff4cHopDNE5xGG106R3ZPR9HZ67F4n565lTvulE3AaWRSLTacAVXR6QORiEbHx8lURA1GwiQ-gMT4wJEn74kqQyjFcMxXEsVwbFcuxXEcVwXFctxXI8LxvIBhqiLQABeegDABcG6IwzKEgQVH2E4rhAX2rAlowQKiGCpDsFBMFwvWMaILCMbiaUAF0Qx6jMQ4LhyH2OgQIIzGCSI7AUmAohWtQ1BXPA-C0IIMBXISjBSFcJB0D4nJSSm9GRHJQIKc4fYENaTH6XAvSQREmkwm5pRTM85zrO5DmlE5jHyaxfYAI7CmQeTSgMfxgjA7DSciMYpuJHlEKIpBkulECZdlKbBvlMZ9qgjBECYeRAhljBZdJknJg+XIAL5cgAVkQKiJAA1nArCiMUGS2Fkg3DWNE1fnNECjeNQzVGQdRiFcG21PUohXD09mdMtq0cJMT4iAdJyXdtJxHWA3qnQt60gZt+1XLd12lNpD1ApMu1gQ9SlKFyShoIOWQqHkX4pECJSdHwQhXUCSPCGIX6VKQMCJBIiSiJ5pDkIwkRENjgR1KEmMEkSJL-WA6yE8TpO7lae2hHMtFaCwtBEijAx0MV1NQgkGCoLQBAjaIiCIOLkuTbRUIpKIcAla9oEfY9JGGl0oty1LMvc3UfMY4rip0kNTGlbomj8BNJJwBoMtwL8rAFCoEg2NrOtdCrIgaL5ZqTMWUBXBIA5XFuO7HUBPswMImgwa5rHez7uhglu9F+SxLhm-eqeGnrEsG4g3PUGpk0PoafYiyWPD69LsvFwrseKsrqsKzpz7XTXipF-Lhs8ybLdp9C+SWw0ZZDEl1qaKQDInMQxXfas9WNXkMej5UfuO+Q2faTPKXzztRUlVca9NaQm9b-H1CJyIycuAXOsZzIgcDG5ee6FXio--ntG1GtKIEwmVAhgmtoLfiZsEiJEARAYBmUxSSmJsLPudcxbNxlg3VBhp25qwut3a+o9+4lyNrzNkptCBvVqKwcC-AiFpxIY3MuFc0ityhH-HktEf7dVBkoWgJhdDsBUMTEE2hPxKTKJ0UQhQICsGeOgdgEMAAkghaClBUSrRgvxepKDAD1AAukAA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6usTmpIy0YnH+GRgmzPDsFZW6tPy6ALxVHWAAIq54ugpgAGJYk6WTAJJ2cxNgzgDKK5MACsvjkwDis-tgAHIAMltgAEJj82DnjlcAgntyuJ2VEkQ6jBDwZH6pnUpCIjA6EC6XQI7AATKV4bpEQBGBFo3SojGlTGw96fLrUKD0ODUURAmGTACyJBYBHyrCuWDgqEE9DoBEmikhXTx3MqhIgEkEUCkiRMsFo1FYQPFGjBEKhlRhmIAzKU1boNRqUdj1djeYrdITiaTyUMDjIYFBlByTgAJEkiG2MohWixwTn4g1dVBEFmEzQkM0ADg1ADYAOylMMABlKABYoxi4xjMcik8jMTGMGHSgBOb2VFhwYIDGEqgCsSYrFdVMcxFZjFfzKJViPjdYROdKocLXn4qDNKow8a7SdhGDTI9K2Y12cR2aT2ebumzWZHff4cHopDNE5xGG106R3ZPR9HZ67F4n565lTvulE3AaWRSLTacAVXR6QORiEbHx8lURA1GwiQ-gMT4wJEn74kqQyjFcMxXEsVwbFcuxXEcVwXFctxXI8LxvIBhqiLQABeegDABcG6IwzKEgQVH2E4rhAX2rAlowQKiGCpDsFBMFwvWMaILCMbiaUAF0Qx6jMQ4LhyH2OgQIIzGCSI7AUmAohWtQ1BXPA-C0IIMBXISjBSFcJB0D4nJSSm9GRHJQIKc4fYENaTH6XAvSQREmkwm5pRTM85zrO5DmlE5jHyaxfYAI7CmQeTSgMfxgjA7DSciMYpuJHlEKIpBkulECZdlKbBvlMZ9qgjBECYeRAhljBZdJknJg+XIAL5cgAVkQKiJAA1nArCiMUGS2Fkg3DWNE1fnNECjeNQzVGQdRiFcG21PUohXD09mdMtq0cJMT4iAdJyXdtJxHWA3qnQt60gZt+1XLd12lNpD1ApMu1gQ9SlKFyShoIOWQqHkX4pECJSdHwQhXUCSPCGIX6VKQMCJBIiSiJ5pDkIwkRENjgR1KEmMEkSJL-WA6yE8TpO7lae2hHMtFaCwtBEijAx0MV1NQgkGCoLQBAjaIiCIOLkuTbRUIpKIcAla9oEfY9JGGl0oty1LMvc3UfMY4rip0kNTGlbomj8BNJJwBoMtwL8rAFCoEg2NrOtdCrIgaL5ZqTMWUBXBIA5XFuO7HUBPswMImgwa5rHez7uhglu9F+SxLhm-eqeGnrEsG4g3PUGpk0PoafYiyWPD69LsvFwrseKsrqsKzpz7XTXipF-Lhs8ybLdp9C+SWw0ZZDEl1qaKQDInMQxXfas9WNXkMej5UfuO+Q2faTPKXzztRUlVca9NaQm9b-H1CJyIycuAXOsZzIgcDG5ee6FXio--ntG1GtKIEwmVAhgmtoLfiZsEiJEARAYBmUxSSmJsLPudcxbNxlg3VBhp25qwut3a+o9+4lyNrzNkptCBvVqKwcC-AiFpxIY3MuFc0ityhH-HktEf7dVBkoWgJhdDsBUMTEE2hPxKTKJ0UQhQICsGeOgdgEMAAkghaClBUSrRgvxepKDAD1AAukAA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
  \item{example-2}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXQBBHwCAGWTdVIxMxERGRpbahp8sALaOrp7G4aUAX0UIACsiFTSAazhWUQrc23z+OBMoYVI0gn5aUQI0xeW1jeBoeE3MuQBdN2h0NpVi9gHM3QAvCFcrgBnwhCJRIDdODhGJfhA6nVSDA0hI0hcwuRGHEiCiwoxaFEEUikdQoPQ4NRoQowAECFiZLjSDCCUTaaDEaTdFoWLQKZDoXRRKQSdzkXBuBhULQCCtRN0ZXLNgNxXVMqI4KQVWA+s1aXJOWqkR0lfLurzCQL4arjboCCUlgQxNDNPwNlS4BpunAdGxSioJDYjXa6pqRBo4PxoQRfrqAOKuPC6WlYepeJNVWkBZwADQNIdDMCOtDieiBDhchbtREYe0YUehlec1eNJiJjaBADF6s0c7bxYaB9zTbLzYhedRBPC5sah1zxaPlYqxyqF2qNVqdZMAgXh6Sl+PLfz6JCxaGHU6XUC3R6I6QfX7WAGIEH56Gw57I9GgbHafVEw5FNdQANV7ID-x8ABNPd12NYtqE0MsmycFt925Wt607exUNbNV2yibCez7Vw4O5Wc1Qo0l3xHZgIFEExaxgMJayhIFhVFdDUjSUg6IYpi0nbRCZHPNVDwVRAzTXD92klHgpO6TdtTjHdYJk8SLT5a0dQABXqNMAFlfDUj8NInbhp02cMvXIH9gIAIWcAJhh0vwDXQuoqPI4cvIouY5loExdHYFRsXUTQdBsWwai5UQyggVh6nQdg0FQAASQRaCqVK0s1Rg-TmaYlDAaYXiAA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXQBBHwCAGWTdVIxMxERGRpbahp8sALaOrp7G4aUAX0UIACsiFTSAazhWUQrc23z+OBMoYVI0gn5aUQI0xeW1jeBoeE3MuQBdN2h0NpVi9gHM3QAvCFcrgBnwhCJRIDdODhGJfhA6nVSDA0hI0hcwuRGHEiCiwoxaFEEUikdQoPQ4NRoQowAECFiZLjSDCCUTaaDEaTdFoWLQKZDoXRRKQSdzkXBuBhULQCCtRN0ZXLNgNxXVMqI4KQVWA+s1aXJOWqkR0lfLurzCQL4arjboCCUlgQxNDNPwNlS4BpunAdGxSioJDYjXa6pqRBo4PxoQRfrqAOKuPC6WlYepeJNVWkBZwADQNIdDMCOtDieiBDhchbtREYe0YUehlec1eNJiJjaBADF6s0c7bxYaB9zTbLzYhedRBPC5sah1zxaPlYqxyqF2qNVqdZMAgXh6Sl+PLfz6JCxaGHU6XUC3R6I6QfX7WAGIEH56Gw57I9GgbHafVEw5FNdQANV7ID-x8ABNPd12NYtqE0MsmycFt925Wt607exUNbNV2yibCez7Vw4O5Wc1Qo0l3xHZgIFEExaxgMJayhIFhVFdDUjSUg6IYpi0nbRCZHPNVDwVRAzTXD92klHgpO6TdtTjHdYJk8SLT5a0dQABXqNMAFlfDUj8NInbhp02cMvXIH9gIAIWcAJhh0vwDXQuoqPI4cvIouY5loExdHYFRsXUTQdBsWwai5UQyggVh6nQdg0FQAASQRaCqVK0s1Rg-TmaYlDAaYXiAA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
}
}

