% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{tm_g_scatterplotmatrix}
\alias{tm_g_scatterplotmatrix}
\title{Create a scatterplot matrix}
\usage{
tm_g_scatterplotmatrix(label = "Scatterplot matrix", dataname,
  select_col, plot_height = c(600, 200, 2000), pre_output = NULL,
  post_output = NULL)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{dataname}{Name of datasets used to generate the regression plot (just used for labeling).}

\item{select_col}{(\code{list}) Output of \code{teal.devel::data_extract_spec}
to define the plotting variables from an incoming dataset with filtering and selecting.}

\item{plot_height}{(\code{numeric}) A vector of length three with \code{c(value, min and max)} for a slider
encoding the plot height.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
Create a scatterplot matrix
}
\examples{
library(random.cdisc.data)
library(tern)

ASL <- cadsl
ADTE <- radtte(ASL, seed = 1, event.descr = c("STUDYID", "USUBJID", "PARAMCD"))

keys(ASL) <- c("STUDYID", "USUBJID")
keys(ADTE) <- c("STUDYID", "USUBJID", "PARAMCD")

adte_filters <- filter_spec(
  vars = c("PARAMCD"), #'  only key variables are allowed
  sep = " - ",
  choices = c("OS", "PFS", "EFS"),
  selected = "OS",
  multiple = TRUE, #'  if multiple, then a spread is needed
  label = "Choose endpoint"
)

adte_prep <- data_extract_spec(
  dataname = "ADTE",
  filter = adte_filters,
  columns = columns_spec(
    choices = colnames(ADTE),
    selected = if (all(c('AGE', 'SEX') \%in\% colnames(ADTE))) {
     c('AGE', 'SEX')
    } else {
     colnames(ADTE)[1:2]
    },
    multiple = TRUE,
    fixed = FALSE, #'  Whether the user can select the item
    label = "" #'  Label the column select dropdown (optional)
  )
)

app <- init(
  data = cdisc_data(
       ASL = ASL,
       ADTE = ADTE,
       code = 'ASL <- cadsl
               ADTE <- radtte(ASL, seed = 1, event.descr = c("STUDYID", "USUBJID", "PARAMCD"))
               keys(ASL) <- c("STUDYID", "USUBJID")
               keys(ADTE) <- c("STUDYID", "USUBJID", "PARAMCD")
               ',
       check = FALSE),
  modules = root_modules(
    tm_g_scatterplotmatrix(
      label = "Scatterplot matrix",
      dataname = c("ASL","ADTE"),
      select_col = list(adte_prep)
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
