% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{tm_g_scatterplotmatrix}
\alias{tm_g_scatterplotmatrix}
\title{Create a scatterplot matrix}
\usage{
tm_g_scatterplotmatrix(label = "Scatterplot matrix", dataname, selected,
  plot_height = c(600, 200, 2000), pre_output = NULL,
  post_output = NULL)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{dataname}{Name of datasets used to generate the regression plot (just used for labeling).}

\item{selected}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Plotting variables from an incoming dataset with filtering and selecting.}

\item{plot_height}{(\code{numeric}) A vector of length three with \code{c(value, min and max)} for a slider
encoding the plot height.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
Create a scatterplot matrix
}
\examples{

# datasets: single wide dataset

library(random.cdisc.data)
library(tern)

ASL <- cadsl
ADTE <- radtte(ASL, seed = 1, event.descr = c("STUDYID", "USUBJID", "PARAMCD"))
keys(ASL) <- c("STUDYID", "USUBJID")

app <- init(
  data = cdisc_data(
    ASL = ASL,
    code = 'ASL <- cadsl
keys(ASL) <- c("STUDYID", "USUBJID")',
    check = FALSE
  ),
  modules = root_modules(
    tm_g_scatterplotmatrix(
      label = "Scatterplot matrix",
      dataname = c("ASL"),
      selected = data_extract_spec(
        dataname = "ASL",
        columns = columns_spec(
          label = "Selected columns",
          choices = colnames(ASL),
          selected = c("AGE", "RACE", "SEX"),
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
# datasets: different wide

library(random.cdisc.data)
library(tern)
library(dplyr)

ASL <- cadsl
ASL <- mutate_at(ASL,
  .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
  .funs = funs(as.factor(.))
) \%>\% select(
  "ARM", "ACTARM", "ACTARMCD",
  "SEX", "STRATA1", "AGE", "USUBJID", "STUDYID", "STRATA2"
)
keys(ASL) <- c("STUDYID", "USUBJID")


ASL_2 <- mutate_at(cadsl,
  .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
  .funs = funs(as.factor(.))
) \%>\% select("ACTARM", "AGE", "STRATA2", "COUNTRY", "USUBJID", "STUDYID")
keys(ASL_2) <- c("STUDYID", "USUBJID")


app <- init(
  data = cdisc_data(
    ASL = ASL,
    ASL_2 = ASL_2,
    code = 'ASL <- cadsl
ASL_2 <- mutate_at(cadsl,
.vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
.funs = funs(as.factor(.))) \%>\% select("ACTARM", "AGE", "STRATA2", "COUNTRY", "USUBJID", "STUDYID")
keys(ASL) <- keys(ASL_2) <- c("STUDYID", "USUBJID")',
    check = FALSE
  ),
  modules = root_modules(
    tm_g_scatterplotmatrix(
      label = "Scatterplot matrix",
      dataname = c("ASL", "ASL_2"),
      selected = list(
        data_extract_spec(
          dataname = "ASL",
          columns = columns_spec(
            label = "Selected columns",
            choices = colnames(ASL),
            selected = c("AGE", "ACTARM", "SEX"),
            multiple = TRUE,
            fixed = FALSE
          )
        ),
        data_extract_spec(
          dataname = "ASL_2",
          columns = columns_spec(
            label = "Selected columns",
            choices = colnames(ASL_2),
            selected = c("COUNTRY", "ACTARM", "STRATA2"),
            multiple = TRUE,
            fixed = FALSE
          )
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
# datasets: multiple long datasets
library(random.cdisc.data)

ASL <- cadsl
ADRS <- cadrs
ADTTE <- cadtte

app <- init(
  data = cdisc_data(
    ASL = ASL,
    ADRS = ADRS,
    ADTTE = ADTTE,
    code = "ASL <- cadsl; ADRS <- cadrs; ADTTE <- cadtte",
    check = FALSE
  ),
  modules = root_modules(
    tm_g_scatterplotmatrix(
      label = "Scatterplot matrix on two long datasets",
      dataname = c("ASL", "ADRS", "ADTTE"),
      selected = list(
        data_extract_spec(
          dataname = "ADRS",
          columns = columns_spec(
            choices = names(ADRS),
            selected = c("AVAL", "AVALC"),
            multiple = TRUE,
            fixed = FALSE
          ),
          filter = filter_spec(
            vars = c("PARAMCD", "AVISIT"),
          choices = apply(expand.grid(levels(ADRS$PARAMCD), levels(ADRS$AVISIT)), 1, paste, collapse = " - "),
            selected = "OVRINV - Screening",
            multiple = TRUE
          )
        ),
        data_extract_spec(
          dataname = "ADTTE",
          columns = columns_spec(
            choices = names(ADTTE),
            selected = c("AVAL", "CNSR"),
            multiple = TRUE,
            fixed = FALSE
          ),
          filter = filter_spec(
            vars = c("PARAMCD"),
            choices = unique(ADTTE$PARAMCD),
            selected = "OS",
            multiple = TRUE
          )
        )
      )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: different subsets of long dataset

library(random.cdisc.data)
library(tern)

ASL <- cadsl
ADLB <- cadlb

keys(ASL) <- c("STUDYID", "USUBJID")
keys(ADLB) <- c("STUDYID", "USUBJID", "PARAMCD", "AVISIT")

app <- init(
  data = cdisc_data(
    ASL = ASL,
    ADLB = ADLB,
    code = 'ASL <- cadsl
           ADLB <- cadrs
           keys(ASL) <- c("USUBJID", "STUDYID")
           keys(ADLB) <- c("USUBJID", "STUDYID", "PARAMCD", "AVISIT")',
    check = FALSE
  ),
  modules = root_modules(
    tm_g_scatterplotmatrix(
      label = "Scatterplot matrix",
      dataname = "ADLB",
      selected = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[1],
            multiple = FALSE,
            label = "Choose measurement"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Choose visit"
          )
        ),
        columns = columns_spec(
          choices = c("AVAL", "CHG", "BMRKR1", "BMRKR2", "AGE"),
          selected = c("AVAL", "CHG"),
          multiple = TRUE,
          fixed = FALSE,
          label = "Variable"
        )
      )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
