% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_scatterplotmatrix.R
\name{tm_g_scatterplotmatrix}
\alias{tm_g_scatterplotmatrix}
\title{\code{teal} module: Scatterplot matrix}
\usage{
tm_g_scatterplotmatrix(
  label = "Scatterplot Matrix",
  variables,
  plot_height = c(600, 200, 2000),
  plot_width = NULL,
  pre_output = NULL,
  post_output = NULL
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{variables}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Specifies plotting variables from an incoming dataset with filtering and selecting. In case of
\code{data_extract_spec} use \code{select_spec(..., ordered = TRUE)} if plot elements should be
rendered according to selection order.}

\item{plot_height}{(\code{numeric}) optional, specifies the plot height as a three-element vector of
\code{value}, \code{min}, and \code{max} intended for use with a slider UI element.}

\item{plot_width}{(\code{numeric}) optional, specifies the plot width as a three-element vector of
\code{value}, \code{min}, and \code{max} for a slider encoding the plot width.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Generates a scatterplot matrix from selected \code{variables} from datasets.
Each plot within the matrix represents the relationship between two variables,
providing the overview of correlations and distributions across selected data.
}
\note{
For more examples, please see the vignette "Using scatterplot matrix" via
\code{vignette("using-scatterplot-matrix", package = "teal.modules.general")}.
}
\examples{
# general data example
data <- teal_data()
data <- within(data, {
  countries <- data.frame(
    id = c("DE", "FR", "IT", "ES", "PT", "GR", "NL", "BE", "LU", "AT"),
    government = factor(
      c(2, 2, 2, 1, 2, 2, 1, 1, 1, 2),
      labels = c("Monarchy", "Republic")
    ),
    language_family = factor(
      c(1, 3, 3, 3, 3, 2, 1, 1, 3, 1),
      labels = c("Germanic", "Hellenic", "Romance")
    ),
    population = c(83, 67, 60, 47, 10, 11, 17, 11, 0.6, 9),
    area = c(357, 551, 301, 505, 92, 132, 41, 30, 2.6, 83),
    gdp = c(3.4, 2.7, 2.1, 1.4, 0.3, 0.2, 0.7, 0.5, 0.1, 0.4),
    debt = c(2.1, 2.3, 2.4, 2.6, 2.3, 2.4, 2.3, 2.4, 2.3, 2.4)
  )
  sales <- data.frame(
    id = 1:50,
    country_id = sample(
      c("DE", "FR", "IT", "ES", "PT", "GR", "NL", "BE", "LU", "AT"),
      size = 50,
      replace = TRUE
    ),
    year = sort(sample(2010:2020, 50, replace = TRUE)),
    venue = sample(c("small", "medium", "large", "online"), 50, replace = TRUE),
    cancelled = sample(c(TRUE, FALSE), 50, replace = TRUE),
    quantity = rnorm(50, 100, 20),
    costs = rnorm(50, 80, 20),
    profit = rnorm(50, 20, 10)
  )
})
datanames(data) <- c("countries", "sales")
join_keys(data) <- join_keys(
  join_key("countries", "countries", "id"),
  join_key("sales", "sales", "id"),
  join_key("countries", "sales", c("id" = "country_id"))
)

app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplotmatrix(
      label = "Scatterplot matrix",
      variables = list(
        data_extract_spec(
          dataname = "countries",
          select = select_spec(
            label = "Select variables:",
            choices = variable_choices(data[["countries"]]),
            selected = c("area", "gdp", "debt"),
            multiple = TRUE,
            ordered = TRUE,
            fixed = FALSE
          )
        ),
        data_extract_spec(
          dataname = "sales",
          filter = filter_spec(
            label = "Select variable:",
            vars = "country_id",
            choices = value_choices(data[["sales"]], "country_id"),
            selected = c("DE", "FR", "IT", "ES", "PT", "GR", "NL", "BE", "LU", "AT"),
            multiple = TRUE
          ),
          select = select_spec(
            label = "Select variables:",
            choices = variable_choices(data[["sales"]], c("quantity", "costs", "profit")),
            selected = c("quantity", "costs", "profit"),
            multiple = TRUE,
            ordered = TRUE,
            fixed = FALSE
          )
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# CDISC data example
data <- teal_data()
data <- within(data, {
  ADSL <- rADSL
  ADRS <- rADRS
})
datanames(data) <- c("ADSL", "ADRS")
join_keys(data) <- default_cdisc_join_keys[datanames(data)]

app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplotmatrix(
      label = "Scatterplot matrix",
      variables = list(
        data_extract_spec(
          dataname = "ADSL",
          select = select_spec(
            label = "Select variables:",
            choices = variable_choices(data[["ADSL"]]),
            selected = c("AGE", "RACE", "SEX"),
            multiple = TRUE,
            ordered = TRUE,
            fixed = FALSE
          )
        ),
        data_extract_spec(
          dataname = "ADRS",
          filter = filter_spec(
            label = "Select endpoints:",
            vars = c("PARAMCD", "AVISIT"),
            choices = value_choices(data[["ADRS"]], c("PARAMCD", "AVISIT"), c("PARAM", "AVISIT")),
            selected = "INVET - END OF INDUCTION",
            multiple = TRUE
          ),
          select = select_spec(
            label = "Select variables:",
            choices = variable_choices(data[["ADRS"]]),
            selected = c("AGE", "AVAL", "ADY"),
            multiple = TRUE,
            ordered = TRUE,
            fixed = FALSE
          )
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
\section{Examples in Shinylive}{
\itemize{
  \item example-1\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6usTmpIy0YnH+GRgmzPDsFZW6tPy6ALxVHWAAIq54ugpgAGJYk6WTAJJ2cxNgzgDKK5MACsvjkwDis-tgAHIAMltgAEJj82DnjlcAgntyuJ2VEkQ6jBDwZH6pnUpCIjA6EC6XQI7AATKV4bpEQBGBFo3SojGlTGw96fLrUKD0ODUURAmGTACyJBYBHyrCuWDgqEE9DoBEmikhXTx3MqhIgEkEUCkiRMsFo1FYQPFGjBEKhlRhmIAzKU1boNRqUdj1djeYrdITiaTyUMDjIYFBlByTgAJEkiG2MohWixwTn4g1dVBEFmEzQkM0ADg1ADYAOylMMABlKABYoxi4xjMcik8jMTGMGHSgBOb2VFhwYIDGEqgCsSYrFdVMcxFZjFfzKJViPjdYROdKocLXn4qDNKow8a7SdhGDTI9K2Y12cR2aT2ebumzWZHff4cHopDNE5xGG106R3ZPR9HZ67F4n565lTvulE3AaWRSLTacAVXR6QORiEbHx8lURA1GwiQ-gMT4wJEn74kqQyjFcMxXEsVwbFcuxXEcVwXFctxXI8LxvIBhqiLQABeegDABcG6IwzKEgQVH2E4rhAX2rAlowQKiGCpDsFBMFwvWMaILCMbiaUAF0Qx6jMQ4LhyH2OgQIIzGCSI7AUmAohWtQ1BXPA-C0IIMBXISjBSFcJB0D4nJSSm9GRHJQIKc4fYENaTH6XAvSQREmkwm5pRTM85zrO5DmlE5jHyaxfYAI7CmQeTSgMfxgjA7DSciMYpuJHlEKIpBkulECZdlKbBvlMZ9qgjBECYeRAhljBZdJknJg+XIAL5cik0DwKIxQZLYWTadUZB1GIVxPiIoiehAABWRAqIkADWcCsMNKRjcYK1rZt21fgdEAbVtQyTbU9QLScV3Tbd9w9PZnSnedHCTHNM0nF9j2rM9YDem9R2XSBU03bNz5-dpANApM92sOB-CclyXJKGgg5ZCoeRfikQIlJ0fBCPNQJE8IYhfpUpAwIkEiJKInmkOQjCREQ1OBHUoSUwSRIknDYDrIzzOs7uVrXaEcy0VoLC0ESJMDHQxXc1CKSJGEtQgvTqBwBSQGGgNsDMfDYPXd9JGGl0ogkjru6QdbGhazrysW8afMDJM6z27u0t1HLYiIJLesW3Sq1MaVug+7LbJwIkIe0GHI1BMAwDG6BENgAAuhnfYW4+Xu+WakzFlAVwSAOVxbjuL1B4aMDCJoMGuax5u55UYJbvRfksS4Let01oQFwMoXhWxre6A+ho510qvq8wDuiNrutjwb8D879gdj011DMzKkrM47S9j0avN+O7AtexHMt+wHeC0YaPvh6nU2IwDve53HYdAtL1BqbH+ShxTFIydPpQ0mFne4CMkbVyPnnEQGhB6DEmIhE4yETioROOhE4mETjYROLhE4+ETiEROK8aBR867b1oI3AYbk748jfoqK2cDbawJtgfZ2hpXan1WJ7Zhl9fbR1EDfBhhoP4NAGJHP2f8AE7QyMAnSoDM4Z1KNpJK1pNCkAZHdIqJUrj1UankFGIjGH5y7qo5KGitEQJ0X9SY+imqkDIWPChDcRBNx7nQqE7cZAIOCp4ro-cEHDwiv4ieiown3i9J0VGShaAmF0OwFQzMQTaE-EpMonRRCFAgKwZ46B2AYwACSCFoKUIpVtGC-F6koMAPUM5AA}{Open in Shinylive}\cr
    \if{html}{\out{<iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6usTmpIy0YnH+GRgmzPDsFZW6tPy6ALxVHWAAIq54ugpgAGJYk6WTAJJ2cxNgzgDKK5MACsvjkwDis-tgAHIAMltgAEJj82DnjlcAgntyuJ2VEkQ6jBDwZH6pnUpCIjA6EC6XQI7AATKV4bpEQBGBFo3SojGlTGw96fLrUKD0ODUURAmGTACyJBYBHyrCuWDgqEE9DoBEmikhXTx3MqhIgEkEUCkiRMsFo1FYQPFGjBEKhlRhmIAzKU1boNRqUdj1djeYrdITiaTyUMDjIYFBlByTgAJEkiG2MohWixwTn4g1dVBEFmEzQkM0ADg1ADYAOylMMABlKABYoxi4xjMcik8jMTGMGHSgBOb2VFhwYIDGEqgCsSYrFdVMcxFZjFfzKJViPjdYROdKocLXn4qDNKow8a7SdhGDTI9K2Y12cR2aT2ebumzWZHff4cHopDNE5xGG106R3ZPR9HZ67F4n565lTvulE3AaWRSLTacAVXR6QORiEbHx8lURA1GwiQ-gMT4wJEn74kqQyjFcMxXEsVwbFcuxXEcVwXFctxXI8LxvIBhqiLQABeegDABcG6IwzKEgQVH2E4rhAX2rAlowQKiGCpDsFBMFwvWMaILCMbiaUAF0Qx6jMQ4LhyH2OgQIIzGCSI7AUmAohWtQ1BXPA-C0IIMBXISjBSFcJB0D4nJSSm9GRHJQIKc4fYENaTH6XAvSQREmkwm5pRTM85zrO5DmlE5jHyaxfYAI7CmQeTSgMfxgjA7DSciMYpuJHlEKIpBkulECZdlKbBvlMZ9qgjBECYeRAhljBZdJknJg+XIAL5cik0DwKIxQZLYWTadUZB1GIVxPiIoiehAABWRAqIkADWcCsMNKRjcYK1rZt21fgdEAbVtQyTbU9QLScV3Tbd9w9PZnSnedHCTHNM0nF9j2rM9YDem9R2XSBU03bNz5-dpANApM92sOB-CclyXJKGgg5ZCoeRfikQIlJ0fBCPNQJE8IYhfpUpAwIkEiJKInmkOQjCREQ1OBHUoSUwSRIknDYDrIzzOs7uVrXaEcy0VoLC0ESJMDHQxXc1CKSJGEtQgvTqBwBSQGGgNsDMfDYPXd9JGGl0ogkjru6QdbGhazrysW8afMDJM6z27u0t1HLYiIJLesW3Sq1MaVug+7LbJwIkIe0GHI1BMAwDG6BENgAAuhnfYW4+Xu+WakzFlAVwSAOVxbjuL1B4aMDCJoMGuax5u55UYJbvRfksS4Let01oQFwMoXhWxre6A+ho510qvq8wDuiNrutjwb8D879gdj011DMzKkrM47S9j0avN+O7AtexHMt+wHeC0YaPvh6nU2IwDve53HYdAtL1BqbH+ShxTFIydPpQ0mFne4CMkbVyPnnEQGhB6DEmIhE4yETioROOhE4mETjYROLhE4+ETiEROK8aBR867b1oI3AYbk748jfoqK2cDbawJtgfZ2hpXan1WJ7Zhl9fbR1EDfBhhoP4NAGJHP2f8AE7QyMAnSoDM4Z1KNpJK1pNCkAZHdIqJUrj1UankFGIjGH5y7qo5KGitEQJ0X9SY+imqkDIWPChDcRBNx7nQqE7cZAIOCp4ro-cEHDwiv4ieiown3i9J0VGShaAmF0OwFQzMQTaE-EpMonRRCFAgKwZ46B2AYwACSCFoKUIpVtGC-F6koMAPUM5AA" height="800" width="150\%" allow="fullscreen" scrolling="auto" style="border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem;"></iframe>}}
  \item example-2\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXQBBHwCAGWTdRkaW2oafLAC2jt6ApQBfRQhM6HhRCtzbfIJ2BTBO5uWq5ca+5fGAKyIVNIBrOFYZzPnjfjgTKGFSNIJ+WlECNP3Dk7PgSdgxWfCcgAum5oOg2ipiksIHVMroALwhXK4bp8IQiUQI3Ro4T-bp1UgwNISNKvMLkRhxIiEsKMWhRaF1Jm6ahQehwahY5YBAjkmRU0jY2n09b4plaFi0NkYrF0USkRnMpmZNLRUjMDSk1BwRZipVI8JTPSIzZNNZ4PVK0QcnWCxHWkSa0Ta3Uw-VK1nszkmsABG0aXQSunSsSIUVu91MgglA4EMRYoNS+giR4x2hx865YDAU1dMBAoFyFERyO6B22uD8LG6lYAcVceF0yyw9S8DY2vucAA0dsXS0yYPdaHFjfYnM4+-3dERGNdGJWsQ4XJP+yZ6QvEQAxerNAKuEv68aRouW2G5VVRdXqB7OnWKyO-eBclZDcOr2jUClYtefmRau+nh6bIcs+fqOoKFD8KgBxkKIYYWge7pBpiiI1gACvULYALK+OsTYrAAaoEfh2L2gHMtGsbxoiErUIIcCplRmbhNmprbPmQJVOhmH1DhPh4ZsREBCRva6Nx2ECYRxGkWAcgnoh+rlhoG74X4AByBHOHYujGM4ak+LoADym66OpPiOF4dh+IZalvlOg6fsOIiLuO5G6PJ-ZKXaZb+jeLr3qWnogT6YG2oGkohnBdn9pR6bUeFwbJgxsUZgCUCsS+7EFh5U5eSpNb1PWkn1ARO7FT4ACaZEKUqDmaCOLnLm5dQznOKlLhOzWmOuVZbjue5uUe7pDcyI3ud04zjLQJi6OwKgUte2hwDYtg1G6ohlBArD1Og7BoKgAAkgi0FU+0HdajA6Iw4wjEoYAjECQA}{Open in Shinylive}\cr
    \if{html}{\out{<iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXQBBHwCAGWTdRkaW2oafLAC2jt6ApQBfRQhM6HhRCtzbfIJ2BTBO5uWq5ca+5fGAKyIVNIBrOFYZzPnjfjgTKGFSNIJ+WlECNP3Dk7PgSdgxWfCcgAum5oOg2ipiksIHVMroALwhXK4bp8IQiUQI3Ro4T-bp1UgwNISNKvMLkRhxIiEsKMWhRaF1Jm6ahQehwahY5YBAjkmRU0jY2n09b4plaFi0NkYrF0USkRnMpmZNLRUjMDSk1BwRZipVI8JTPSIzZNNZ4PVK0QcnWCxHWkSa0Ta3Uw-VK1nszkmsABG0aXQSunSsSIUVu91MgglA4EMRYoNS+giR4x2hx865YDAU1dMBAoFyFERyO6B22uD8LG6lYAcVceF0yyw9S8DY2vucAA0dsXS0yYPdaHFjfYnM4+-3dERGNdGJWsQ4XJP+yZ6QvEQAxerNAKuEv68aRouW2G5VVRdXqB7OnWKyO-eBclZDcOr2jUClYtefmRau+nh6bIcs+fqOoKFD8KgBxkKIYYWge7pBpiiI1gACvULYALK+OsTYrAAaoEfh2L2gHMtGsbxoiErUIIcCplRmbhNmprbPmQJVOhmH1DhPh4ZsREBCRva6Nx2ECYRxGkWAcgnoh+rlhoG74X4AByBHOHYujGM4ak+LoADym66OpPiOF4dh+IZalvlOg6fsOIiLuO5G6PJ-ZKXaZb+jeLr3qWnogT6YG2oGkohnBdn9pR6bUeFwbJgxsUZgCUCsS+7EFh5U5eSpNb1PWkn1ARO7FT4ACaZEKUqDmaCOLnLm5dQznOKlLhOzWmOuVZbjue5uUe7pDcyI3ud04zjLQJi6OwKgUte2hwDYtg1G6ohlBArD1Og7BoKgAAkgi0FU+0HdajA6Iw4wjEoYAjECQA" height="800" width="150\%" allow="fullscreen" scrolling="auto" style="border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem;"></iframe>}}
}
}

