% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_simple_regression.R
\name{tm_g_simple_regression}
\alias{tm_g_simple_regression}
\title{Scatterplot and Regression Model}
\usage{
tm_g_simple_regression(label = "Simple Regression Analysis", dataname,
  response_var, regressor_var, plot_height = c(600, 200, 2000),
  pre_output = NULL, post_output = NULL)
}
\description{
This module is designed for horizontal data
}
\examples{

\dontrun{
N <- 100
ASL <- data.frame(
  USUBJID = paste("id", seq_len(N), sep = "-"),
  STUDYID = "study1",
  F1 = factor(sample(paste0("facet_1_", c("A", "B")), N, TRUE)),
  F2 = factor(sample(paste0("facet_2_", c("a", "b", "c")), N, TRUE)),
  cont = rnorm(N),
  disc = factor(sample(letters[1:5], N, TRUE)),
  cont2 = runif(N),
  disc2 = factor(sample(LETTERS[1:5], N, TRUE))
)

ASL$cont3 <- ASL$cont *3 +2 +rnorm(nrow(ASL), sd = .3)

attr(ASL, "source") <- "# ASL is random data"

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_g_simple_regression(
      dataname = "ASL",
      response_var = choices_selected(c("cont", "cont2"), "cont"),
      regressor_var = choices_selected(names(ASL), "cont3"),
      plot_height = c(600, 200, 2000)
    )
  ))

shinyApp(x$ui, x$server)



fit <- lm(cont ~ cont2, data = ASL)

plot(fit)

}



}
