% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_landing_popup.R
\name{tm_landing_popup}
\alias{tm_landing_popup}
\alias{landing_popup}
\title{Landing Popup Module}
\usage{
tm_landing_popup(
  label = "Landing Popup",
  title = NULL,
  content = NULL,
  buttons = modalButton("Accept")
)

landing_popup(title = NULL, content = NULL, buttons = modalButton("Accept"))
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module. Any label possible except
\code{"global_filters"} - read more in \code{mapping} argument of \link[teal]{teal_slices}.}

\item{title}{\code{character(1)} the text to be displayed as a title of the popup.}

\item{content}{\code{character(1)} the content of the popup. Passed to \code{...} of \code{shiny::modalDialog}. Can be a \code{character}
or a text input control (like \code{textInput}) or a list of \code{shiny} tags. See examples.}

\item{buttons}{\code{shiny} tag or a list of tags (\code{tagList}). Typically a \code{modalButton} or \code{actionButton}. See examples.}
}
\value{
A \code{teal} module to be used in \code{teal} applications.
}
\description{
This module creates a simple landing welcome popup for \code{teal} applications.
}
\details{
If you use this module with a default label \code{"Landing Popup"}, it will be used as a first module in
a \code{teal} application and there will not be a tab provided for this module in the application.
}
\examples{
app1 <- teal::init(
  data = teal.data::dataset("iris", iris),
  modules = teal::modules(
    teal.modules.general::tm_landing_popup(
      title = "Welcome",
      content = "A place for the welcome message or a disclaimer statement.",
      buttons = modalButton("Proceed")
    ),
    teal.modules.general::tm_front_page()
  )
)
if (interactive()) {
  shinyApp(app1$ui, app1$server)
}

app2 <- teal::init(
  data = teal.data::dataset("iris", iris),
  modules = teal::modules(
    teal.modules.general::tm_landing_popup(
      title = "Welcome",
      content = div(tags$b("A place for the welcome message or a disclaimer statement.", style = "color: red;")),
      buttons = tagList(
        modalButton("Proceed"),
        actionButton("close", "Read more", onclick = "window.open('http://google.com', '_blank')")
      )
    ),
    teal.modules.general::tm_front_page()
  )
)

if (interactive()) {
  shinyApp(app2$ui, app2$server)
}

}
