% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_missing_data.R
\name{tm_missing_data}
\alias{tm_missing_data}
\title{Missing data module}
\usage{
tm_missing_data(label = "Missing data", plot_height = c(600, 400, 5000))
}
\arguments{
\item{label}{(\code{character}) Label shown in the navigation item for the module.}

\item{plot_height}{optional, (\code{numeric}) A vector of length three with \code{c(value, min and max)}
for a slider encoding the plot height.}
}
\description{
Present analysis of missing observations and patients.
}
\examples{
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADRS <- radrs(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    code = "ADSL <- radsl(cached = TRUE); ADRS <- radrs(cached = TRUE)",
    check = TRUE
  ),
  modules = root_modules(
    tm_missing_data()
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
}
