% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_missing_data.R
\name{tm_missing_data}
\alias{tm_missing_data}
\title{Missing data module}
\usage{
tm_missing_data(
  label = "Missing data",
  plot_height = c(600, 400, 5000),
  plot_width = NULL,
  ggtheme = c("classic", "gray", "bw", "linedraw", "light", "dark", "minimal", "void",
    "test"),
  ggplot2_args = list(`Combinations Hist` = teal.devel::ggplot2_args(labs =
    list(caption = NULL)), `Combinations Main` = teal.devel::ggplot2_args(labs =
    list(title = NULL))),
  pre_output = NULL,
  post_output = NULL
)
}
\arguments{
\item{label}{(\code{character}) Label shown in the navigation item for the module.}

\item{plot_height}{optional, (\code{numeric}) A vector of length three with \code{c(value, min and max)}
for a slider encoding the plot height.}

\item{plot_width}{optional, (\code{numeric}) A vector of length three with \code{c(value, min and max)}
for a slider encoding the plot width.}

\item{ggtheme}{optional, (\code{character}) \code{ggplot} Theme to be used by default.
One of \code{c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void", "test")}.
Each of the theme can be chosen by the user during the session. Defaults to \code{classic}.}

\item{ggplot2_args}{optional, (`ggplot2_args`) object created by [`teal.devel::ggplot2_args()`]
with settings for all the plots or named list of `ggplot2_args` objects for plot-specific settings.
For more details see the help vignette:
`vignette("Custom ggplot2 arguments module", package = "teal.devel")`.
List names should match the following:\cr
`c("default", "Summary Obs", "Summary Patients", "Combinations Main", "Combinations Hist", "By Subject")`.
The argument is merged with options variable `teal.ggplot2_args` and default module setup.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example a title.}

\item{post_output}{optional, \code{shiny.tag} with text placed after the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
Present analysis of missing observations and patients.
}
\examples{
library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl
ADRS <- synthetic_cdisc_data("latest")$adrs

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"),
    cdisc_dataset("ADRS", ADRS, code = "ADRS <- synthetic_cdisc_data(\"latest\")$adrs"),
    check = TRUE
  ),
  modules = root_modules(
    tm_missing_data()
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
}
