% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_outliers.R
\name{tm_outliers}
\alias{tm_outliers}
\title{Outliers Module}
\usage{
tm_outliers(
  label = "Outliers Module",
  outlier_var,
  categorical_var = NULL,
  lineplot_param = NULL,
  key_identifier = c("USUBJID"),
  plot_height = c(600, 200, 2000),
  plot_width = NULL,
  pre_output = NULL,
  post_output = NULL
)
}
\arguments{
\item{label}{(\code{character}) Label shown in the navigation item for the module.}

\item{outlier_var}{(`data_extract_spec` or `list` of multiple `data_extract_spec`)
variable to consider for the outliers analysis.}

\item{categorical_var}{(`data_extract_spec` or `list` of multiple `data_extract_spec`)
categorical factor to split the selected outliers variable on.}

\item{lineplot_param}{(`data_extract_spec` or `list` of multiple `data_extract_spec`)
x-axis variable to be used for the outlier line plot or `NULL` (default) if no line plot
tab to be included.}

\item{key_identifier}{(`character`)
Identifier key variable to be used in the cumulative distribution plot and the line plot.}

\item{plot_height}{optional, (\code{numeric}) A vector of length three with \code{c(value, min and max)}
for a slider encoding the plot height.}

\item{plot_width}{optional, (\code{numeric}) A vector of length three with \code{c(value, min and max)}
for a slider encoding the plot width.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example a title.}

\item{post_output}{optional, \code{shiny.tag} with text placed after the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
Module to analyze and identify outliers using different methods
}
\examples{
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

fact_vars_adsl <- names(Filter(isTRUE, sapply(ADSL, is.factor)))

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- radsl(cached = TRUE)"),
    check = TRUE
  ),
  modules = root_modules(
    tm_outliers(
      outlier_var = list(
        data_extract_spec(
          dataname = "ADSL",
          select = select_spec(
            label = "Select variable:",
            choices = variable_choices(ADSL, c("AGE", "BMRKR1")),
            selected = "AGE",
            multiple = FALSE,
            fixed = FALSE
          )
        )
      ),
      categorical_var = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          choices = variable_choices(ADSL, subset = fact_vars_adsl),
          selected = "RACE",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

}
