% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_rmarkdown.R
\name{tm_rmarkdown}
\alias{tm_rmarkdown}
\title{\code{teal} module: Rmarkdown render}
\usage{
tm_rmarkdown(
  label = "RMarkdown Module",
  rmd_file,
  datanames = "all",
  allow_download = TRUE,
  pre_output = NULL,
  post_output = NULL,
  transformators = list(),
  extra_transform = list()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{rmd_file}{(\code{character}) Path to the R Markdown file to be rendered.
The file must be accessible from the Shiny app environment.}

\item{datanames}{(\code{character}) Names of the datasets relevant to the item.
There are 2 reserved values that have specific behaviors:
\itemize{
\item The keyword \code{"all"} includes all datasets available in the data passed to the teal application.
\item \code{NULL} hides the sidebar panel completely.
\item If \code{transformators} are specified, their \code{datanames} are automatically added to this \code{datanames}
argument.
}}

\item{allow_download}{(\code{logical}) whether to allow downloading of the R Markdown file.
Defaults to \code{TRUE}.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{transformators}{(\code{list} of \code{teal_transform_module}) that will be applied to transform module's data input.
To learn more check \code{vignette("transform-input-data", package = "teal")}.}

\item{extra_transform}{(\code{list}) of \code{\link[teal:teal_transform_module]{teal::teal_transform_module()}} that will be added in the module's UI.
This can be used to create interactive inputs that modify the parameters in R Markdown rendering.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Module to render R Markdown files using the data provided in the \code{teal_data} object.
}
\details{
The R Markdown file should be designed to accept parameters corresponding to the datasets.
See using \code{params} in R Markdown documentation:
\href{https://bookdown.org/yihui/rmarkdown/params-use.html}{bookdown.org/yihui/rmarkdown/params-use.html}

For example, if the \code{teal_data} object contains datasets named "mtcars" and "iris",
the R Markdown file can define parameters as follows:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{---
title: "R Markdown Report"
output: html_document
params:
  mtcars: NULL
  iris: NULL
---
}\if{html}{\out{</div>}}

The libraries used in the R Markdown file must be available in
the Shiny app environment.
}
\examples{

# general data example
data <- teal_data()
data <- within(data, {
  CO2 <- CO2
  CO2[["primary_key"]] <- seq_len(nrow(CO2))
})
join_keys(data) <- join_keys(join_key("CO2", "CO2", "primary_key"))


app <- init(
  data = data,
  modules = modules(
    tm_rmarkdown(
      label = "RMarkdown Module",
      rmd_file = system.file(file.path("sample_files", "test.Rmd"), package = "teal.modules.general")
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

nrow_transform <- teal_transform_module(
  label = "N Rows selector",
  ui = function(id) {
    ns <- NS(id)
    tags$div(
      numericInput(ns("n_rows"), "Show n rows", value = 5, min = 0, max = 200, step = 5)
    )
  },
  server = function(id, data) {
    moduleServer(id, function(input, output, session) {
      reactive({
        req(data())
        within(data(),
          {
            rmd_data$n_rows <- n_rows_value
          },
          n_rows_value = input$n_rows
        )
      })
    })
  }
)

app <- init(
  data = data,
  modules = modules(
    tm_rmarkdown(
      label = "RMarkdown Module",
      rmd_file = system.file(file.path("sample_files", "test.Rmd"), package = "teal.modules.general"),
      allow_download = FALSE,
      extra_transform = list(nrow_transform)
    )
  )
) |> shiny::runApp()
}
\section{Examples in Shinylive}{
\describe{
  \item{example-1}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokrcQAxLqkQ11XfxQpFC6cAAesKgiSoHBBsZc1AD6sVA2MUEhRroA7rSkABYq7Km4uiBKuroAwgDyAEzxNQ2VzfXAwApgqIy0MCysSQDWcKxdALrjTaJwAI5JIhDsEIxEOex19XKKEAC+OwBWRCrDo6IlmbbZRycjrOc3EKccXZtdZa8teLpdPX0Dzy623cSjQqCaKny7FaqV0AF4AplcK0+EIRKJ4bpUcIxNCIFUqqQYElGP1GEN+Gslq0CbpqFB6HB-AiulgALIsClU3RsgQ4940gmk-hJEy0ESY0T3cgwDBikTseVwDCoIIFaFgUSRESi8Vid4-MDkUSkbAwfhAsqqghDKBSTFdRK8PnoyTSWRAwU7Ko7Ha0Ey6dgqcjMSw6Gy2Cr43SiIoQVgAQXQ7DBABJBLQymmZowdIwdrslGBduMgA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokrcQAxLqkQ11XfxQpFC6cAAesKgiSoHBBsZc1AD6sVA2MUEhRroA7rSkABYq7Km4uiBKuroAwgDyAEzxNQ2VzfXAwApgqIy0MCysSQDWcKxdALrjTaJwAI5JIhDsEIxEOex19XKKEAC+OwBWRCrDo6IlmbbZRycjrOc3EKccXZtdZa8teLpdPX0Dzy623cSjQqCaKny7FaqV0AF4AplcK0+EIRKJ4bpUcIxNCIFUqqQYElGP1GEN+Gslq0CbpqFB6HB-AiulgALIsClU3RsgQ4940gmk-hJEy0ESY0T3cgwDBikTseVwDCoIIFaFgUSRESi8Vid4-MDkUSkbAwfhAsqqghDKBSTFdRK8PnoyTSWRAwU7Ko7Ha0Ey6dgqcjMSw6Gy2Cr43SiIoQVgAQXQ7DBABJBLQymmZowdIwdrslGBduMgA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
}
}

