% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_scatterplot.R
\name{tm_scatterplot}
\alias{tm_scatterplot}
\title{Create a simple scatterplot}
\usage{
tm_scatterplot(label, x, y, color_by = NULL, plot_height = c(600, 200,
  2000), alpha = c(1, 0, 1), size = c(4, 1, 12), pre_output = NULL,
  post_output = NULL)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{x}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
x variable}

\item{y}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
y variable}

\item{color_by}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Defines the color encoding. If
\code{NULL} then no color encoding option will be displayed. Note
\code{_none_} is a keyword and means that no color encoding should be used.}

\item{plot_height}{if scalar then the plot will have a fixed height. If a
slider should be presented to adjust the plot height dynamically then it
can be a vector of length three with \code{c(value, min and max)}.}

\item{alpha}{if scalar then the plot points will have a fixed opacity. If a
slider should be presented to adjust the plot point opacity dynamically
then it can be a vector of length three with vlaue, min and max.}

\item{size}{if scalar then the plot points sizes will have a fixed opacity.
If a slider should be presented to adjust the plot point sizes dynamically
then it can be a vector of length three with vlaue, min and max.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
Create a plot with the \code{\link{plot}[base]} function
}
\examples{

# datasets: single wide
# Scatterplot of variables from ADSL dataset
library(random.cdisc.data)

ADSL <- cadsl

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- cadsl",
    check = TRUE
  ),
  modules = root_modules(
    tm_scatterplot(
      label = "Scatterplot Choices",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = c("AGE", "BMRKR1", "BMRKR2"),
          selected = "AGE",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = c("AGE", "BMRKR1", "BMRKR2"),
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      color_by = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variables:",
          choices = c("RACE", "SEX"),
          selected = NULL,
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
  shinyApp(app$ui, app$server)
}

# datasets: different wide
# Scatterplot of AGE and BMRKR1 colored by COUNTRY
library(random.cdisc.data)
library(dplyr)

ADSL <- cadsl
ADSL <- mutate_at(ADSL,
  .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
  .funs = list(~as.factor(.))
) \%>\% select(
  "ARM", "ACTARM", "ACTARMCD",
  "SEX", "BMRKR1", "BMRKR2", "AGE", "USUBJID", "STUDYID"
)

ADSL_2 <- mutate_at(cadsl,
  .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
  .funs = list(~as.factor(.))
) \%>\% select("ACTARM", "AGE", "STRATA2", "COUNTRY", "USUBJID", "STUDYID")

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    dataset("ADSL_2", ADSL_2),
    code = 'ADSL <- cadsl
ADSL_2 <- mutate_at(cadsl,
.vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
.funs = list(~as.factor(.))) \%>\%
select("ACTARM", "AGE", "STRATA2", "COUNTRY", "USUBJID", "STUDYID")',
    check = FALSE #TODO
  ),
  modules = root_modules(
    tm_scatterplot("Scatterplot for different wide data",
     x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
         label = "Select variable:",
          choices = c("BMRKR1", "BMRKR2"),
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = FALSE
        )),
      y = data_extract_spec(
        dataname = "ADSL_2",
        select = select_spec(
          label = "Select variable:",
          choices = c("AGE", "RACE"),
          selected = "AGE",
          multiple = FALSE,
          fixed = FALSE
        )),
      color_by = data_extract_spec(
        dataname = "ADSL_2",
        select = select_spec(
          label = "Select variable:",
          choices = c("COUNTRY", "AGE", "RACE"),
          selected = "COUNTRY",
          multiple = FALSE,
          fixed = FALSE
        ))
     )
   )
)

\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: multiple long datasets
#' Scatterplot of parameters from ADTTE and ADRS datasets.
library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs
ADTTE <- cadtte

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    cdisc_dataset("ADTTE", ADTTE),
    code = "ADSL <- cadsl; ADRS <- cadrs; ADTTE <- cadtte",
    check = TRUE
  ),
  modules = root_modules(
    tm_scatterplot(
      label = "Scatterplot on two long datasets",
      x = data_extract_spec(
        dataname = "ADRS",
        select = select_spec(
          label = "Select variable:",
          choices = names(ADRS),
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE
        ),
        filter = filter_spec(
          label = "Select endpoint:",
          vars = c("PARAMCD", "AVISIT"),
          choices = apply(expand.grid(
          levels(ADRS$PARAMCD), levels(ADRS$AVISIT)), 1, paste, collapse = " - "),
          selected = "OVRINV - Screening",
          multiple = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADTTE",
        select = select_spec(
          label = "Select variable:",
          choices = names(ADTTE),
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE
        ),
        filter = filter_spec(
          label = "Select parameters:",
          vars = c("PARAMCD"),
          choices = unique(ADTTE$PARAMCD),
          selected = "OS",
          multiple = TRUE
        )
      ),
      color_by = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = c("AGE", "SEX"),
          selected = "AGE",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: wide, long
# Scatterplot of response duration and patient age
library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    code = "ADSL <- cadsl; ADRS <- cadrs",
    check = TRUE
  ),
  modules = root_modules(
   tm_scatterplot("Scatterplot for wide and long data",
     x = data_extract_spec(
          dataname = "ADSL",
          select = select_spec(
            label = "Select variable:",
            choices = c("SEX", "AGE", "RACE", "COUNTRY"),
            selected = "AGE",
            multiple = FALSE,
            fixed = FALSE
          )
        ),
     y = data_extract_spec(
          dataname = "ADRS",
          filter = list(
            filter_spec(
              vars = "PARAM",
              choices = levels(ADRS$PARAM),
              selected = levels(ADRS$PARAM)[1],
              multiple = FALSE,
              label = "Select response:"
            ),
            filter_spec(
              vars = "AVISIT",
              choices = levels(ADRS$AVISIT),
              selected = levels(ADRS$AVISIT)[1],
              multiple = FALSE,
              label = "Select visit:"
            )
          ),
          select = select_spec(
            label = "Selected variable:",
            choices = "AVAL",
            selected = "AVAL",
            multiple = FALSE,
            fixed = TRUE
          )
        ),
     color_by = data_extract_spec(
          dataname = "ADSL",
          select = select_spec(
            label = "Select variable:",
            choices = c("SEX", "AGE", "RACE", "COUNTRY"),
            selected = NULL,
            multiple = FALSE,
            fixed = FALSE
          )
        )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: same long
# Scatterplot of different continuous variables from ADRS dataset

library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    code = "ADSL <- cadsl; ADRS <- cadrs",
    check = TRUE
  ),
  modules = root_modules(
    tm_scatterplot(
      "Scatterplot for same long dataset",
      x = data_extract_spec(
        dataname = "ADRS",
        select = select_spec(
          choices = c("AVAL", "BMRKR1", "BMRKR2"),
          selected = "AVAL",
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      ),
      y = data_extract_spec(
        dataname = "ADRS",
        select = select_spec(
          choices = c("AVAL", "BMRKR1", "BMRKR2"),
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      ),
      color_by = data_extract_spec(
        dataname = "ADRS",
        select = select_spec(
          choices = c("AGE", "SEX", "RACE"),
          selected = NULL,
          multiple = FALSE, #todo: 0 or 1
          fixed = FALSE,
          label = "Select variable:"
        )
      )
    )
  )
)
\dontrun{
  shinyApp(app$ui, app$server)
}



# datasets: different subsets of long dataset
# Scatterplot of different continuous variables from ADLB dataset

library(random.cdisc.data)

ADSL <- cadsl
ADLB <- cadlb

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADLB", ADLB),
    code = "ADSL <- cadsl; ADLB <- cadlb",
    check = TRUE
  ),
  modules = root_modules(
    tm_scatterplot(
      "Scatterplot Choices",
      x = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[1],
            multiple = FALSE,
            label = "Select lab:"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Select visit:"
          )
        ),
        select = select_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = TRUE
        )
      ),
      y = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[2],
            multiple = FALSE,
            label = "Select lab:"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Select visit:"
          )
        ),
        select = select_spec(
          choices = "AVAL",
          selected = "AVAL",
          multiple = FALSE,
          fixed = TRUE
        )
      ),
      color_by = data_extract_spec(
        dataname = "ADLB",
        filter = list(
          filter_spec(
            vars = "PARAMCD",
            choices = levels(ADLB$PARAMCD),
            selected = levels(ADLB$PARAMCD)[1],
            multiple = FALSE,
            label = "Select lab:"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADLB$AVISIT),
            selected = levels(ADLB$AVISIT)[1],
            multiple = FALSE,
            label = "Select visit:"
          )
        ),
        select = select_spec(
          choices = c("RACE", "SEX"),
          selected = "SEX",
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      ),
    )
  )
)
\dontrun{
  shinyApp(app$ui, app$server)
}
}
