% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_bep_summary.R
\name{tm_t_bep_summary}
\alias{tm_t_bep_summary}
\title{Summary Table for analysis}
\usage{
tm_t_bep_summary(label, dataname, arm_var, arm_var_choices = arm_var,
  summarize_vars, summarize_vars_choices = summarize_vars, bep_var,
  bep_var_choices = bep_var, pre_output = NULL, post_output = NULL,
  code_data_processing = NULL)
}
\arguments{
\item{label}{menue item label of the module in the teal app}

\item{dataname}{analysis data used in teal module, needs to be available in
the list passed to the \code{data} argument of \code{\link[teal]{init}}.
Note that the data is expected to be in vertical form with the
\code{PARAMCD} variable filtering to one observation per patient.}

\item{arm_var}{single name of variable in analysis data that is used as
\code{col_by} argument for the respective \code{tern} function.}

\item{arm_var_choices}{vector with variable names that can be used as
\code{arm_var}}

\item{summarize_vars}{character vector with variable names that are selected
by default}

\item{summarize_vars_choices}{character vector with variable names that can
be selected (for summary)}

\item{bep_var}{TODO}

\item{bep_var_choices}{TODO}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{code_data_processing}{string with data preprocessing before the teal
app is initialized}
}
\description{
TODO: descrioption (one paragraph what is this  module doing)
}
\examples{

\dontrun{
library(random.cdisc.data)
library(dplyr)

ASL <- radsl(N = 100) \%>\% mutate(BEP = sample(c(TRUE, FALSE), n(), TRUE))


x <- teal::init(
   data = list(ASL = ASL),
   modules = root_modules(
     tm_t_bep_summary(
        label = "Demographic",
        dataname = "ASL",
        arm_var = "ARM",
        arm_var_choices = c("ARM", "SEX"),
        summarize_vars = "AGE",
        summarize_vars_choices = names(ASL),
        bep_var = "BEP"
     )
  )
)

shinyApp(x$ui, x$server)
}
}
