% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_crosstable.R
\name{tm_t_crosstable}
\alias{tm_t_crosstable}
\title{Create a simple cross-table}
\usage{
tm_t_crosstable(label = "Cross Table", x, y, show_percentage = TRUE,
  show_total = TRUE, pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{(\code{character}) Label of the app in the teal menu}

\item{x}{(\code{choices_selected}) object with all available
choices with pre-selected option for variable X}

\item{y}{(\code{choices_selected}) object with all available
choices with pre-selected option for variable Y}

\item{show_percentage}{whether to show percentages (relevant only when \code{x} is a \code{factor})}

\item{show_total}{whether to show total column}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{...}{Arguments passed on to \code{teal.devel::standard_layout}
\describe{
  \item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
  \item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}}
}
\description{
Create a simple cross-table
}
\examples{
# Percentage cross table of variables from ADSL dataset
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- radsl(cached = TRUE)",
    check = TRUE
  ),
  modules = root_modules(
    tm_t_crosstable(
      label = "Cross Table",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(ADSL, c("COUNTRY", "STUDYID")),
          selected = "COUNTRY",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(ADSL, c("SEX","RACE")),
          selected = "SEX",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
}
\author{
wolfs25 waddella
}
