% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_crosstable.R
\name{tm_t_crosstable}
\alias{tm_t_crosstable}
\title{\code{teal} module: Cross-table}
\usage{
tm_t_crosstable(
  label = "Cross Table",
  x,
  y,
  show_percentage = TRUE,
  show_total = TRUE,
  pre_output = NULL,
  post_output = NULL,
  basic_table_args = teal.widgets::basic_table_args()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{x}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Object with all available choices with pre-selected option for variable X - row values.
In case of \code{data_extract_spec} use \code{select_spec(..., ordered = TRUE)} if table elements should be
rendered according to selection order.}

\item{y}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Object with all available choices with pre-selected option for variable Y - column values.

\code{data_extract_spec} must not allow multiple selection in this case.}

\item{show_percentage}{(\code{logical(1)})
Indicates whether to show percentages (relevant only when \code{x} is a \code{factor}).
Defaults to \code{TRUE}.}

\item{show_total}{(\code{logical(1)})
Indicates whether to show total column.
Defaults to \code{TRUE}.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{basic_table_args}{(\code{basic_table_args}) object created by \code{\link[teal.widgets:basic_table_args]{teal.widgets::basic_table_args()}}
with settings for the module table.
The argument is merged with options variable \code{teal.basic_table_args} and default module setup.

For more details see the vignette: \code{vignette("custom-basic-table-arguments", package = "teal.widgets")}}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Generates a simple cross-table of two variables from a dataset with custom
options for showing percentages and sub-totals.
}
\note{
For more examples, please see the vignette "Using cross table" via
\code{vignette("using-cross-table", package = "teal.modules.general")}.
}
\examples{
# general data example
data <- teal_data()
data <- within(data, {
  mtcars <- mtcars
  for (v in c("cyl", "vs", "am", "gear")) {
    mtcars[[v]] <- as.factor(mtcars[[v]])
  }
  mtcars[["primary_key"]] <- seq_len(nrow(mtcars))
})
datanames(data) <- "mtcars"
join_keys(data) <- join_keys(join_key("mtcars", "mtcars", "primary_key"))

app <- init(
  data = data,
  modules = modules(
    tm_t_crosstable(
      label = "Cross Table",
      x = data_extract_spec(
        dataname = "mtcars",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["mtcars"]], c("cyl", "vs", "am", "gear")),
          selected = c("cyl", "gear"),
          multiple = TRUE,
          ordered = TRUE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "mtcars",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["mtcars"]], c("cyl", "vs", "am", "gear")),
          selected = "vs",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# CDISC data example
data <- teal_data()
data <- within(data, {
  ADSL <- rADSL
})
datanames(data) <- "ADSL"
join_keys(data) <- default_cdisc_join_keys[datanames(data)]

app <- init(
  data = data,
  modules = modules(
    tm_t_crosstable(
      label = "Cross Table",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], subset = function(data) {
            idx <- !vapply(data, inherits, logical(1), c("Date", "POSIXct", "POSIXlt"))
            return(names(data)[idx])
          }),
          selected = "COUNTRY",
          multiple = TRUE,
          ordered = TRUE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], subset = function(data) {
            idx <- vapply(data, is.factor, logical(1))
            return(names(data)[idx])
          }),
          selected = "SEX",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
\section{Examples in Shinylive}{
\itemize{
  \item example-1\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6ujCkBCyicVU1dRWmRIy67Fq6KroE7AqErNQDpQNaoiO6A7CTA1IsA3K25RCVldW1jKLAwFoAunsNUOIm6qRt7Bt1O-t7iqu6AL4tV1s7A6iMtDAsrIkA1nBWAMDg1RHAAI6JEQQdgQRhEbKXJpbJZKR73FLQeCiYoZWxZAavCZgJQAKyIKgBQNxKQJxgpVMBrFxjIg1I4RJRJNGYGJszAn2+vw5i3uSjQqAaKjy-QeKV0AF5-BlcC8BMIxEqqhqRLiWpVSDBEqREgQEaJREF6CI5Ws1tQoPQ4H5lQMAMIW+p2J1RPAGtahbUpRJhUjMDSJUSoOB9AP2rGwPRuvnckbxtbgkQabVZ2Om6Oxu32kuO52uqZgADKLvzui0LFovrgiHTDxLawI+UpBC1yobX2bZu7tF7tIy71TmxJB1KccGwzwlfGApmS7mcAWYCWavbHd0eY0cH42vnBCGAvmjEWu-3axgwk0kWT9iczlvd90bX4MmP2ocLgfneJi0KEf7KgAYgAggAMlWrh7va9wdnIQFrKwwYZKGoThmcUYxnGiFrIm8Daly05tneh6kLmtaRoWhGfroZYumR1Z0TRA5NjaLaUZ+XY9n29aNkOAmjmIeJBJO-JgLOvT9AuAoruuKAwJem7XtuqEZva1HgcuPI6fej60M+2rQXB75GZUIFgSekGwfBRnISWLmVC59z3LQJgdCo5ARpoOg2MsLSiIUECsFB6DsJKAAkgi0KUcXgowOiMPczwQGAjx7EAA}{Open in Shinylive}\cr
    \if{html}{\out{<iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6ujCkBCyicVU1dRWmRIy67Fq6KroE7AqErNQDpQNaoiO6A7CTA1IsA3K25RCVldW1jKLAwFoAunsNUOIm6qRt7Bt1O-t7iqu6AL4tV1s7A6iMtDAsrIkA1nBWAMDg1RHAAI6JEQQdgQRhEbKXJpbJZKR73FLQeCiYoZWxZAavCZgJQAKyIKgBQNxKQJxgpVMBrFxjIg1I4RJRJNGYGJszAn2+vw5i3uSjQqAaKjy-QeKV0AF5-BlcC8BMIxEqqhqRLiWpVSDBEqREgQEaJREF6CI5Ws1tQoPQ4H5lQMAMIW+p2J1RPAGtahbUpRJhUjMDSJUSoOB9AP2rGwPRuvnckbxtbgkQabVZ2Om6Oxu32kuO52uqZgADKLvzui0LFovrgiHTDxLawI+UpBC1yobX2bZu7tF7tIy71TmxJB1KccGwzwlfGApmS7mcAWYCWavbHd0eY0cH42vnBCGAvmjEWu-3axgwk0kWT9iczlvd90bX4MmP2ocLgfneJi0KEf7KgAYgAggAMlWrh7va9wdnIQFrKwwYZKGoThmcUYxnGiFrIm8Daly05tneh6kLmtaRoWhGfroZYumR1Z0TRA5NjaLaUZ+XY9n29aNkOAmjmIeJBJO-JgLOvT9AuAoruuKAwJem7XtuqEZva1HgcuPI6fej60M+2rQXB75GZUIFgSekGwfBRnISWLmVC59z3LQJgdCo5ARpoOg2MsLSiIUECsFB6DsJKAAkgi0KUcXgowOiMPczwQGAjx7EAA" height="800" width="150\%" allow="fullscreen" scrolling="auto" style="border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem;"></iframe>}}
  \item example-2\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXQBBHwCAGWTdRkaWpQBfRQhM6HhRCtzbfIUwTuaJpQArIhU0gGs4VmHMseN+OBMoYVI0gn5aUQI0+cWVteAB2DER8LkAXTdodDaVYvZakNzdAF5fuFcD8+EIRKIAbowcJ7j86qQYGkDgRGERRKJwvQRN8IHV8bpqFB6HBqFCJl40Rj7MT4nh4fiolDMmloqRmBo0qJUHACLiCQTbvByZMmtN6XiBXVRKTeaQoTKRJzubz+VL8USSWTARMArKNLotCxaLS4IgJiDJerdAQSgsCGIoUbGCbsXBDnbaA71rlgMAJlMJk8nlVRIJ6DL5YCzBZNNYNtUGdbaPwmfkAIRG9DUDiVXQqEoyYqiKrUIgSL3cdgARjkVT5Ex8YTgFt0EwACgB5AJ+AAaGlbHe7feopAmcj61oJjDgpEEjAg7EG9w2wBTUSek6nvUtU90irlcH4Iq8nccADkHABNC1JgUwfa0OJ6QEOFy7qdERjbGfH19OZwP2tExaCiI8oQAMXqZoAlcK0BS3Ak6zvVhmVyVkonZdQDhVBt4MFXJlxFQMJWtA8DUBcicJ5PC901UkRT1JV5WdV0RHNUip1te1HUBVjTQ9HifXCP0AzFIMQ33cNIyhGNLHjUZE3w9V1zaLM4lzXIqhODBdg0L9S3LStqBrCc7ylGc5wXJc7mEqA5DXVNN3M-Edxc6V9XIP82zAWDe1vZSCQfUcnxESDoNgoD1RAsDvKgmC4OtRD8WSrc+j6WgTF0dgVHIDlNB0GxbBqSVRDKCBWHqdB2DQVAABJBFoKparqmVGB0Rg+m6JQwG6J4gA}{Open in Shinylive}\cr
    \if{html}{\out{<iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXQBBHwCAGWTdRkaWpQBfRQhM6HhRCtzbfIUwTuaJpQArIhU0gGs4VmHMseN+OBMoYVI0gn5aUQI0+cWVteAB2DER8LkAXTdodDaVYvZakNzdAF5fuFcD8+EIRKIAbowcJ7j86qQYGkDgRGERRKJwvQRN8IHV8bpqFB6HBqFCJl40Rj7MT4nh4fiolDMmloqRmBo0qJUHACLiCQTbvByZMmtN6XiBXVRKTeaQoTKRJzubz+VL8USSWTARMArKNLotCxaLS4IgJiDJerdAQSgsCGIoUbGCbsXBDnbaA71rlgMAJlMJk8nlVRIJ6DL5YCzBZNNYNtUGdbaPwmfkAIRG9DUDiVXQqEoyYqiKrUIgSL3cdgARjkVT5Ex8YTgFt0EwACgB5AJ+AAaGlbHe7feopAmcj61oJjDgpEEjAg7EG9w2wBTUSek6nvUtU90irlcH4Iq8nccADkHABNC1JgUwfa0OJ6QEOFy7qdERjbGfH19OZwP2tExaCiI8oQAMXqZoAlcK0BS3Ak6zvVhmVyVkonZdQDhVBt4MFXJlxFQMJWtA8DUBcicJ5PC901UkRT1JV5WdV0RHNUip1te1HUBVjTQ9HifXCP0AzFIMQ33cNIyhGNLHjUZE3w9V1zaLM4lzXIqhODBdg0L9S3LStqBrCc7ylGc5wXJc7mEqA5DXVNN3M-Edxc6V9XIP82zAWDe1vZSCQfUcnxESDoNgoD1RAsDvKgmC4OtRD8WSrc+j6WgTF0dgVHIDlNB0GxbBqSVRDKCBWHqdB2DQVAABJBFoKparqmVGB0Rg+m6JQwG6J4gA" height="800" width="150\%" allow="fullscreen" scrolling="auto" style="border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem;"></iframe>}}
}
}

