% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_crosstable.R
\name{tm_t_crosstable}
\alias{tm_t_crosstable}
\title{\code{teal} module: Cross-table}
\usage{
tm_t_crosstable(
  label = "Cross Table",
  x,
  y,
  show_percentage = TRUE,
  show_total = TRUE,
  pre_output = NULL,
  post_output = NULL,
  basic_table_args = teal.widgets::basic_table_args()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{x}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Object with all available choices with pre-selected option for variable X - row values.
In case of \code{data_extract_spec} use \code{select_spec(..., ordered = TRUE)} if table elements should be
rendered according to selection order.}

\item{y}{(\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
Object with all available choices with pre-selected option for variable Y - column values.

\code{data_extract_spec} must not allow multiple selection in this case.}

\item{show_percentage}{(\code{logical(1)})
Indicates whether to show percentages (relevant only when \code{x} is a \code{factor}).
Defaults to \code{TRUE}.}

\item{show_total}{(\code{logical(1)})
Indicates whether to show total column.
Defaults to \code{TRUE}.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{basic_table_args}{(\code{basic_table_args}) object created by \code{\link[teal.widgets:basic_table_args]{teal.widgets::basic_table_args()}}
with settings for the module table.
The argument is merged with options variable \code{teal.basic_table_args} and default module setup.

For more details see the vignette: \code{vignette("custom-basic-table-arguments", package = "teal.widgets")}}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Generates a simple cross-table of two variables from a dataset with custom
options for showing percentages and sub-totals.
}
\note{
For more examples, please see the vignette "Using cross table" via
\code{vignette("using-cross-table", package = "teal.modules.general")}.
}
\examples{
\dontshow{if (require("rtables", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# general data example
data <- teal_data()
data <- within(data, {
  mtcars <- mtcars
  for (v in c("cyl", "vs", "am", "gear")) {
    mtcars[[v]] <- as.factor(mtcars[[v]])
  }
  mtcars[["primary_key"]] <- seq_len(nrow(mtcars))
})
datanames(data) <- "mtcars"
join_keys(data) <- join_keys(join_key("mtcars", "mtcars", "primary_key"))

app <- init(
  data = data,
  modules = modules(
    tm_t_crosstable(
      label = "Cross Table",
      x = data_extract_spec(
        dataname = "mtcars",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["mtcars"]], c("cyl", "vs", "am", "gear")),
          selected = c("cyl", "gear"),
          multiple = TRUE,
          ordered = TRUE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "mtcars",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["mtcars"]], c("cyl", "vs", "am", "gear")),
          selected = "vs",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
\dontshow{\}) # examplesIf}
\dontshow{if (require("rtables", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# CDISC data example
data <- teal_data()
data <- within(data, {
  ADSL <- rADSL
})
datanames(data) <- "ADSL"
join_keys(data) <- default_cdisc_join_keys[datanames(data)]

app <- init(
  data = data,
  modules = modules(
    tm_t_crosstable(
      label = "Cross Table",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], subset = function(data) {
            idx <- !vapply(data, inherits, logical(1), c("Date", "POSIXct", "POSIXlt"))
            return(names(data)[idx])
          }),
          selected = "COUNTRY",
          multiple = TRUE,
          ordered = TRUE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(data[["ADSL"]], subset = function(data) {
            idx <- vapply(data, is.factor, logical(1))
            return(names(data)[idx])
          }),
          selected = "SEX",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
\dontshow{\}) # examplesIf}
}
\section{Examples in Shinylive}{
\itemize{
  \item example-1\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNrv2CsIykUPQiov64ul60cKTUrLoAvPZOzooQAMS6UhBq1Lr8UEG6cAAesKgiSgVFRrpc1AD61VA2VYVQBsYA7rSkABYq7M0RIEq6ujCkBCyinRNTM2OmRIy67Fq6KroEfoSs1OG6-lpheEcoMIf+Uiz+crajEOPjk9OMosDAWgC633NQ4hM6lIK3Yrxmnx+3zS4wAvktwe9Pv5UIxaDAWKwGgBrOCsfy-OaiTwNEQQdgQRhELpghbve5KWFpZrQeCiIbtWy1fyI05KABWRBUOLx7OaXOMguFuNY7KlEBFHB5dNOEWVb1V51R6MxiruaSUaFQcxUvT8T3y7USlqCuARAmEYmtfCEoXNzzqMAapAaBCpolEQRCcHdHt01GCcDySX8AGF-bM7MFKnglh6Stbmg1SqRmBoGqJUHAdmmwyzYHoY2BeeFSx7iSINNaG8WfYXi6Gw88I-Qo9b-ABlKOt3RaFi0ZNwRC1i1d8YEPpCghOpJjtGT32L2jLsXtZHVlUE74REt7A5nY6a-ywK5gG6MfV22dzlsaOD8a2ngj7W-3u5PudnhgYRNAqStkhcADAN0FZ+Bkd9rQcSC6zDExaBKBCkgAMQAQQAGQHVxnw9GEwzkKCPXiJIsxzPM2yLEtiOect4H7A8NRnQDX1IZth3zdtGOg8NI2jc4h0bHi1wnYNp1TJiPQXJcV1HccN0U7cxA5IJ9xrMBfhPXZv3PNUwBOW8bwvO84FuMB7gol8+PID8qzMuShOA6hQJEa1cII5x7K7NCMOc3RfMIlDxlIkjS1ItI0loEw1hUcg6O0EN7l0R5xlEAYIFYHD0HYI0ABJvAiEriUYHRGDSeEIDAWFviAA}{Open in Shinylive}\cr
    \if{html}{
      \out{        <div class = "iframe-wrapper" style="height: 800px; width: 100\%">          <script type="text/javascript">$(function() {  var if_pkgdown = [...document.scripts].filter(x => x.src.includes("pkgdown.js")).length > 0;  if (if_pkgdown) {    $(".iframe-wrapper").css("width", "140\%");  }});          </script>          <iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNrv2CsIykUPQiov64ul60cKTUrLoAvPZOzooQAMS6UhBq1Lr8UEG6cAAesKgiSgVFRrpc1AD61VA2VYVQBsYA7rSkABYq7M0RIEq6ujCkBCyinRNTM2OmRIy67Fq6KroEfoSs1OG6-lpheEcoMIf+Uiz+crajEOPjk9OMosDAWgC633NQ4hM6lIK3Yrxmnx+3zS4wAvktwe9Pv5UIxaDAWKwGgBrOCsfy-OaiTwNEQQdgQRhELpghbve5KWFpZrQeCiIbtWy1fyI05KABWRBUOLx7OaXOMguFuNY7KlEBFHB5dNOEWVb1V51R6MxiruaSUaFQcxUvT8T3y7USlqCuARAmEYmtfCEoXNzzqMAapAaBCpolEQRCcHdHt01GCcDySX8AGF-bM7MFKnglh6Stbmg1SqRmBoGqJUHAdmmwyzYHoY2BeeFSx7iSINNaG8WfYXi6Gw88I-Qo9b-ABlKOt3RaFi0ZNwRC1i1d8YEPpCghOpJjtGT32L2jLsXtZHVlUE74REt7A5nY6a-ywK5gG6MfV22dzlsaOD8a2ngj7W-3u5PudnhgYRNAqStkhcADAN0FZ+Bkd9rQcSC6zDExaBKBCkgAMQAQQAGQHVxnw9GEwzkKCPXiJIsxzPM2yLEtiOect4H7A8NRnQDX1IZth3zdtGOg8NI2jc4h0bHi1wnYNp1TJiPQXJcV1HccN0U7cxA5IJ9xrMBfhPXZv3PNUwBOW8bwvO84FuMB7gol8+PID8qzMuShOA6hQJEa1cII5x7K7NCMOc3RfMIlDxlIkjS1ItI0loEw1hUcg6O0EN7l0R5xlEAYIFYHD0HYI0ABJvAiEriUYHRGDSeEIDAWFviAA" allow="fullscreen" scrolling="auto" style="height: 100\%; width: 100\%; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem"></iframe>        </div>      }
    }
  \item example-2\cr
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNrv2CsIykUPQiov64ul60cKTUrLoAvPZOzooQAMS6AMIAIgCSAMpZuvxQQbpwAB6wqCJKpeVGulzUAPoNUDb1ZVAGxgDutKQAFirsHREgSrq6AII5BQAyfbqM80tKAL5pHdDwouM9tk3+64v+SgBWRCqtANZwrAcdx8b8cCZQwqStBPy0ogIrWutweT2Au1gYkOQTkAF0lEo0KgViohn4IDMOokSj1cNNdHwhKEcUThNCCTNSDBWj8CIwiKJREEQnAMTMObpqME4NQcf4sgymfZgnU8JSOZUcR1WlVSMwNK1RKg4AR2ZyOZD4PywGdwhLOaJearSDijSJFcrVeqNRzufReTqCsaNLotCxaKK4Ih9ZjbRyCMMbgQxDj3YxPazfkHaCHnj1gMBTgtzmA4XCIqJBPQjaakmYLJprC9dFM-f6ObR+FKmgBCd3oOIwqARFTDGRDUQRahECSx7jsACMcgiav8OTKcHCun8AAUAPIFPIADQ007ni5X1FI-jkaQrnPcpEEjAg7D20JewCrlTh+4P23x5f95pNcH4Oqy88cADkHABNX0D0Jb5aFqPQkgcFwn2AohGHedwP0glIYIPExaEqd8cQAMVmRYClcZ9OXvYjUM5eIkhlOUFR+K0xyIzUegvHU9XFBiZlfV0kk42iVXo4D7UdJJ-GdC1TXDSMRB9NjgMDYNQySCSvWjeT4yCRNkw2NMM10LMcxiHEC0sYsjlLA1-RvFYG1qDgJl0AEME+DQ4O7Xt+2oIc93M20jxPM8LzUqA5Gvas728jlH3CjiXXIJCZzAAjlyAg8YFA8CcLwgiyP9dDMLi3D8MIisSI5Er7zSNJaBMXR2BUcgaO0Nk9zMv1RFGCBWFmdB2GRAASbwIj6o1GB0Rg0k2JQwE2OEgA}{Open in Shinylive}\cr
    \if{html}{
      \out{        <div class = "iframe-wrapper" style="height: 800px; width: 100\%">          <script type="text/javascript">$(function() {  var if_pkgdown = [...document.scripts].filter(x => x.src.includes("pkgdown.js")).length > 0;  if (if_pkgdown) {    $(".iframe-wrapper").css("width", "140\%");  }});          </script>          <iframe src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAoksZwAjoNrv2CsIykUPQiov64ul60cKTUrLoAvPZOzooQAMS6AMIAIgCSAMpZuvxQQbpwAB6wqCJKpeVGulzUAPoNUDb1ZVAGxgDutKQAFirsHREgSrq6AII5BQAyfbqM80tKAL5pHdDwouM9tk3+64v+SgBWRCqtANZwrAcdx8b8cCZQwqStBPy0ogIrWutweT2Au1gYkOQTkAF0lEo0KgViohn4IDMOokSj1cNNdHwhKEcUThNCCTNSDBWj8CIwiKJREEQnAMTMObpqME4NQcf4sgymfZgnU8JSOZUcR1WlVSMwNK1RKg4AR2ZyOZD4PywGdwhLOaJearSDijSJFcrVeqNRzufReTqCsaNLotCxaKK4Ih9ZjbRyCMMbgQxDj3YxPazfkHaCHnj1gMBTgtzmA4XCIqJBPQjaakmYLJprC9dFM-f6ObR+FKmgBCd3oOIwqARFTDGRDUQRahECSx7jsACMcgiav8OTKcHCun8AAUAPIFPIADQ007ni5X1FI-jkaQrnPcpEEjAg7D20JewCrlTh+4P23x5f95pNcH4Oqy88cADkHABNX0D0Jb5aFqPQkgcFwn2AohGHedwP0glIYIPExaEqd8cQAMVmRYClcZ9OXvYjUM5eIkhlOUFR+K0xyIzUegvHU9XFBiZlfV0kk42iVXo4D7UdJJ-GdC1TXDSMRB9NjgMDYNQySCSvWjeT4yCRNkw2NMM10LMcxiHEC0sYsjlLA1-RvFYG1qDgJl0AEME+DQ4O7Xt+2oIc93M20jxPM8LzUqA5Gvas728jlH3CjiXXIJCZzAAjlyAg8YFA8CcLwgiyP9dDMLi3D8MIisSI5Er7zSNJaBMXR2BUcgaO0Nk9zMv1RFGCBWFmdB2GRAASbwIj6o1GB0Rg0k2JQwE2OEgA" allow="fullscreen" scrolling="auto" style="height: 100\%; width: 100\%; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem"></iframe>        </div>      }
    }
}
}

