% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_crosstable.R
\name{tm_t_crosstable}
\alias{tm_t_crosstable}
\title{\code{teal} module: Cross-table}
\usage{
tm_t_crosstable(
  label = "Cross Table",
  x,
  y,
  show_percentage = TRUE,
  show_total = TRUE,
  remove_zero_columns = FALSE,
  pre_output = NULL,
  post_output = NULL,
  basic_table_args = teal.widgets::basic_table_args(),
  transformators = list(),
  decorators = list()
)
}
\arguments{
\item{x}{(\code{picks} or \code{list} of \code{picks})
Object with all available choices with pre-selected option for variable X - row values.
In case of \code{picks} use \code{variables(..., ordered = TRUE)} if table elements should be
rendered according to selection order.}

\item{y}{(\code{picks} or \code{list} of multiple \code{picks})
Object with all available choices with pre-selected option for variable Y - column values.

\code{picks} must not allow multiple selection in this case.}

\item{show_percentage}{(\code{logical(1)})
Indicates whether to show percentages (relevant only when \code{x} is a \code{factor}).
Defaults to \code{TRUE}.}

\item{show_total}{(\code{logical(1)})
Indicates whether to show total column.
Defaults to \code{TRUE}.}

\item{remove_zero_columns}{(\code{logical(1)})
Indicates whether to remove columns that contain only zeros from the output table.
Defaults to \code{FALSE}.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{basic_table_args}{(\code{basic_table_args}) object created by \code{\link[teal.widgets:basic_table_args]{teal.widgets::basic_table_args()}}
with settings for the module table.
The argument is merged with options variable \code{teal.basic_table_args} and default module setup.

For more details see the vignette: \code{vignette("custom-basic-table-arguments", package = "teal.widgets")}}

\item{decorators}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
(named \code{list} of lists of \code{teal_transform_module}) optional,
decorator for tables or plots included in the module output reported.
The decorators are applied to the respective output objects.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Generates a simple cross-table of two variables from a dataset with custom
options for showing percentages and sub-totals.
}
\note{
For more examples, please see the vignette "Using cross table" via
\code{vignette("using-cross-table", package = "teal.modules.general")}.
}
\section{Table Settings}{

The module provides several table settings that can be adjusted:
\itemize{
\item \code{Show column percentage}: Shows column percentages when enabled
\item \code{Show total column}: Shows a total column when enabled
\item \code{Remove zero-only columns}: Removes columns that contain only zeros from the output table
}
}

\section{Decorating Module}{


This module generates the following objects, which can be modified in place using decorators:
\itemize{
\item \code{table} (\code{ElementaryTable} - output of \code{rtables::build_table})
}

A Decorator is applied to the specific output using a named list of \code{teal_transform_module} objects.
The name of this list corresponds to the name of the output to which the decorator is applied.
See code snippet below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tm_t_crosstable(
   ..., # arguments for module
   decorators = list(
     table = teal_transform_module(...) # applied to the `table` output
   )
)
}\if{html}{\out{</div>}}

For additional details and examples of decorators, refer to the vignette
\code{vignette("decorate-module-output", package = "teal.modules.general")}.

To learn more please refer to the vignette
\code{vignette("transform-module-output", package = "teal")} or the \code{\link[teal:teal_transform_module]{teal::teal_transform_module()}} documentation.
}

\examples{
# general data example
data <- teal_data()
data <- within(data, {
  mtcars <- mtcars
  for (v in c("cyl", "vs", "am", "gear")) {
    mtcars[[v]] <- as.factor(mtcars[[v]])
  }
  mtcars[["primary_key"]] <- seq_len(nrow(mtcars))
})
join_keys(data) <- join_keys(join_key("mtcars", "mtcars", "primary_key"))

app <- init(
  data = data,
  modules = modules(
    tm_t_crosstable(
      label = "Cross Table",
      x = picks(
        datasets("mtcars"),
        variables(
          choices = variable_choices(data[["mtcars"]], c("cyl", "vs", "am", "gear")),
          selected = c("cyl", "gear"),
          multiple = TRUE,
          ordered = TRUE,
          fixed = FALSE
        ),
        values()
      ),
      y = picks(
        datasets("mtcars"),
        variables(
          choices = variable_choices(data[["mtcars"]], c("cyl", "vs", "am", "gear")),
          selected = "vs",
          multiple = FALSE,
          fixed = FALSE
        ),
        values()
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# CDISC data example
data <- teal_data()
data <- within(data, {
  ADSL <- teal.data::rADSL
})
join_keys(data) <- default_cdisc_join_keys[names(data)]

app <- init(
  data = data,
  modules = modules(
    tm_t_crosstable(
      label = "Cross Table",
      x = picks(
        datasets("ADSL"),
        variables(
          choices = variable_choices(data[["ADSL"]], subset = function(data) {
            idx <- !vapply(data, inherits, logical(1), c("Date", "POSIXct", "POSIXlt"))
            return(names(data)[idx])
          }),
          selected = "COUNTRY",
          multiple = TRUE,
          ordered = TRUE,
          fixed = FALSE
        ),
        values()
      ),
      y = picks(
        datasets("ADSL"),
        variables(
          choices = variable_choices(data[["ADSL"]], subset = function(data) {
            idx <- vapply(data, is.factor, logical(1))
            return(names(data)[idx])
          }),
          selected = "SEX",
          multiple = FALSE,
          fixed = FALSE
        ),
        values()
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

}
\section{Examples in Shinylive}{
\describe{
  \item{example-1}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6ujCkBCyicVU1dRWmRIy67Fq6KroE7AqErNQDpQNaoiO6A7CTA1IsA3K25RCVldW1jKLAwFoAunsNUOIm6qRt7Bt1O-t7iqu6AL4tV1s7A6iMtDAsrIkA1nBWAMDg1RHAAI6JEQQdgQRhEbKXJpbJZKR73ABWRBUAKBomKGVsWWxuMBrAJpIgeI4A1eEzwUzA9NmYE+31+NMW9yUaFQDRUeX6DxSugAvP4MrgXgJhGJxVVZSICS1KqQYIlSIkCAjRKIgvQRMK1mtqFB6HA-BKBgBhXX1OzmqJ4VVrUIK1C0Aj-FUPE2VFLg0gq5kohlyaV+-1aFi0J1iY3+k0EfI4gjyiUxr7x7Wpr0JlLvUObBkHUp9AYEIas8asmaMuZwBZgJaRpMm8EiDRwfgKiuDYYNsDzRiLNvt9bCTSRPQShwuccTtr8GQ9hXz5yL9smWihNcSgBiAEEADIAZVcUZNEddJpj1EECfuSZvV90rA9Xp9iaTgbgwf6YtmhbLc71jeNfQnSoUzTDNdCzONDTgXNYIJQtgDpMMQT2ctAKrQdRjAWsh3rQiR25UD-U7OBu17a0iIZSiTRgKdaBnBVj3PTdb39Hc9zo3ROIvHjKlfCd70fAln39aTRJae57loEwOhUchmEsHQbGWFpREKCBWCPdB2D5AASQRaFKUzwUYHRGHuZ4IDAR49iAA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXVIhrqu-lCkULpwAB6wqCJKAUEGxlzUAPoxUDbRgcFGugDutKQAFirsKbi6IEq6ujCkBCyicVU1dRWmRIy67Fq6KroE7AqErNQDpQNaoiO6A7CTA1IsA3K25RCVldW1jKLAwFoAunsNUOIm6qRt7Bt1O-t7iqu6AL4tV1s7A6iMtDAsrIkA1nBWAMDg1RHAAI6JEQQdgQRhEbKXJpbJZKR73ABWRBUAKBomKGVsWWxuMBrAJpIgeI4A1eEzwUzA9NmYE+31+NMW9yUaFQDRUeX6DxSugAvP4MrgXgJhGJxVVZSICS1KqQYIlSIkCAjRKIgvQRMK1mtqFB6HA-BKBgBhXX1OzmqJ4VVrUIK1C0Aj-FUPE2VFLg0gq5kohlyaV+-1aFi0J1iY3+k0EfI4gjyiUxr7x7Wpr0JlLvUObBkHUp9AYEIas8asmaMuZwBZgJaRpMm8EiDRwfgKiuDYYNsDzRiLNvt9bCTSRPQShwuccTtr8GQ9hXz5yL9smWihNcSgBiAEEADIAZVcUZNEddJpj1EECfuSZvV90rA9Xp9iaTgbgwf6YtmhbLc71jeNfQnSoUzTDNdCzONDTgXNYIJQtgDpMMQT2ctAKrQdRjAWsh3rQiR25UD-U7OBu17a0iIZSiTRgKdaBnBVj3PTdb39Hc9zo3ROIvHjKlfCd70fAln39aTRJae57loEwOhUchmEsHQbGWFpREKCBWCPdB2D5AASQRaFKUzwUYHRGHuZ4IDAR49iAA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
  \item{example-2}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXQBBHwCAGWTdVIxMxERGRpalAF9FCAArIhU0gGs4VlEK3Nt8-jgTKGFSNIJ+WlECNLGJ6dngaHg5zLkAXTdodDaVYvZakNzdAF4X8NxnviERUXeul+wjETwgdTqpBgaQ2BEYRFEonC9BEYIhEOoUHocGogIUYC88MR9ix8Twz3RUUBqFoBEmcwp6M+UFEcFIDLAfWa+Lk33BTLqWhYtFJoMZAoIJXGBDEgKFjBFKLgmyltNBmWAwHxXPxl0uVVEgnorNIgLMFk01gu1XFAt0tH4VPyAEIhehqBxKvaICUZMVRFVqEQJLTuOwAIy83QEJ5gHxhOD4qr4gAKAHkAn4ABoaJO6VMZ7PUUg84Z2pmMNmCRgQdindULYAOqKXMvl3RDPnt3SskQaOD8PEEtOOAByDgAmknbeiYOtaHE9B8HC4u+2iIxlpXB8unM41+WTLQogPAQAxerNAKufkC3kz3RC6iCUFt9H328Q1jU2n0tF2zITQ5HUwA-ct5UVf5-3LSVpVlD4INFFU4POXJNW1JpuTAPUDSNE0zXMSwrQWG1P3LZs2jdOJPVyKodgwVYNA3QNg1DagIzkN920rUhq1retUPCOQm0dVsHzqTtxJ7HE4H7Hd8zAa8s2nMimTnYsFxEc9L2vA87SPE95IvK8b3LMC7SfF85i4uouLbYZhloExdHYFRyGYSwdBsWwan5UQyggVh6nQdg0FQAASQRaCqMLwtZRgdEYYYBiUMABkuIA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYy5qAH1EqBsEsIijXQB3WlIACxV2TNxdECVdXQBBHwCAGWTdVIxMxERGRpalAF9FCAArIhU0gGs4VlEK3Nt8-jgTKGFSNIJ+WlECNLGJ6dngaHg5zLkAXTdodDaVYvZakNzdAF4X8NxnviERUXeul+wjETwgdTqpBgaQ2BEYRFEonC9BEYIhEOoUHocGogIUYC88MR9ix8Twz3RUUBqFoBEmcwp6M+UFEcFIDLAfWa+Lk33BTLqWhYtFJoMZAoIJXGBDEgKFjBFKLgmyltNBmWAwHxXPxl0uVVEgnorNIgLMFk01gu1XFAt0tH4VPyAEIhehqBxKvaICUZMVRFVqEQJLTuOwAIy83QEJ5gHxhOD4qr4gAKAHkAn4ABoaJO6VMZ7PUUg84Z2pmMNmCRgQdindULYAOqKXMvl3RDPnt3SskQaOD8PEEtOOAByDgAmknbeiYOtaHE9B8HC4u+2iIxlpXB8unM41+WTLQogPAQAxerNAKufkC3kz3RC6iCUFt9H328Q1jU2n0tF2zITQ5HUwA-ct5UVf5-3LSVpVlD4INFFU4POXJNW1JpuTAPUDSNE0zXMSwrQWG1P3LZs2jdOJPVyKodgwVYNA3QNg1DagIzkN920rUhq1retUPCOQm0dVsHzqTtxJ7HE4H7Hd8zAa8s2nMimTnYsFxEc9L2vA87SPE95IvK8b3LMC7SfF85i4uouLbYZhloExdHYFRyGYSwdBsWwan5UQyggVh6nQdg0FQAASQRaCqMLwtZRgdEYYYBiUMABkuIA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
}
}

