% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_percentage_cross_table.R
\name{tm_t_percentage_cross_table}
\alias{tm_t_percentage_cross_table}
\title{Cross table based on rtables}
\usage{
tm_t_percentage_cross_table(label = "Cross Table", dataname, x, y,
  pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{(\code{character}) Label of the app in the teal menu}

\item{dataname}{(\code{character}) Name of the dataset used in the teal app. Just a single
dataset is allowed!}

\item{x}{(\code{choices_selected}) object with all available
choices with preselected option for variable X}

\item{y}{(\code{choices_selected}) object with all available
choices with preselected option for variable Y}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{...}{Arguments passed on to \code{teal.devel::standard_layout}
\describe{
  \item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
  \item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}}
}
\description{
Cross table based on rtables
}
\examples{
library(random.cdisc.data)

ASL <- cadsl

app <- init(
  data = cdisc_data(
    ASL = ASL,
    code = "ASL <- cadsl",
    check = TRUE
  ),
  modules = root_modules(
    tm_t_percentage_cross_table(
      label = "Cross Table",
      dataname = "ASL",
      x = data_extract_spec(
        dataname = "ASL",
        columns = columns_spec(
          label = "Select X Variable",
          choices = c("STRATA1", "STRATA2"),
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ASL",
        columns = columns_spec(
          label = "Select Y Variable",
          choices = c("STRATA2", "STRATA1"),
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# multiple long datasets
library(random.cdisc.data)

ASL <- cadsl
ADRS <- cadrs
ADTTE <- cadtte

app <- init(
  data = cdisc_data(
    ASL = ASL,
    ADRS = ADRS,
    ADTTE = ADTTE,
    code = "ASL <- cadsl; ADRS <- cadrs; ADTTE <- cadtte",
    check = FALSE
  ),
  modules = root_modules(
    tm_t_percentage_cross_table(
      label = "Cross Table",
      dataname = "ASL",
      x = data_extract_spec(
        dataname = "ADRS",
        columns = columns_spec(
          choices = names(ADRS),
          selected = "AVALC",
          multiple = FALSE,
          fixed = FALSE
        ),
        filter = filter_spec(
          vars = c("PARAMCD", "AVISIT"),
          choices = apply(expand.grid(unique(ADRS$PARAMCD), unique(ADRS$AVISIT)),
                          1, paste, collapse = " - "),
          selected = "OVRINV - Screening",
          multiple = TRUE
        )
      ),
      y = data_extract_spec(
        dataname = "ADTTE",
        columns = columns_spec(
          choices = names(ADTTE),
          selected = "CNSR",
          multiple = FALSE,
          fixed = FALSE
        ),
        filter = filter_spec(
          vars = c("PARAMCD"),
          choices = unique(ADTTE$PARAMCD),
          selected = "OS",
          multiple = TRUE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
}
\author{
wolfs25 waddella
}
