% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_percentage_cross_table.R
\name{tm_t_percentage_cross_table}
\alias{tm_t_percentage_cross_table}
\title{Cross table based on rtables}
\usage{
tm_t_percentage_cross_table(label = "Cross Table", dataname, x, y,
  pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{(\code{character}) Label of the app in the teal menu}

\item{dataname}{(\code{character}) Name of the dataset used in the teal app. Just a single
dataset is allowed!}

\item{x}{(\code{choices_selected}) object with all available
choices with preselected option for variable X}

\item{y}{(\code{choices_selected}) object with all available
choices with preselected option for variable Y}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{...}{Arguments passed on to \code{teal.devel::standard_layout}
\describe{
  \item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
  \item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}}
}
\description{
Cross table based on rtables
}
\examples{
library(random.cdisc.data)
# Percentage cross table of variables from ADSL dataset
# datasets: single wide

ADSL <- cadsl

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- cadsl",
    check = TRUE
  ),
  modules = root_modules(
    tm_t_percentage_cross_table(
      label = "Cross Table",
      dataname = "ADSL",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = c("COUNTRY", "STUDYID"),
          selected = "COUNTRY",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = c("SEX", "RACE"),
          selected = "SEX",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: different wide
# Percentage cross table with AGE groups over RACE
library(random.cdisc.data)
library(dplyr)

ADSL <- cadsl
ADSL <- mutate_at(ADSL,
  .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
  .funs = list(~as.factor(.))
) \%>\% select(
  "ARM", "ACTARM", "ACTARMCD",
  "SEX", "STRATA1", "AGE", "USUBJID", "STUDYID", "STRATA2"
)

ADSL_2 <- mutate_at(cadsl,
  .vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
  .funs = list(~as.factor(.))
) \%>\% select("ACTARM", "AGE", "STRATA2", "COUNTRY", "USUBJID", "STUDYID")

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    dataset("ADSL_2", ADSL_2),
    code = 'ADSL <- cadsl
ADSL_2 <- mutate_at(cadsl,
.vars = vars(c("ARM", "ACTARM", "ACTARMCD", "SEX", "STRATA1", "STRATA2")),
.funs = list(~as.factor(.))) \%>\%
select("ACTARM", "AGE", "STRATA2", "COUNTRY", "USUBJID", "STUDYID")',
    check = FALSE #TODO
  ),
  modules = root_modules(
    tm_t_percentage_cross_table("Cross Table",
     dataname = c("ADSL", "ADSL_2"),
     x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = c("AGE", "SEX", "STRATA1", "RACE"),
          selected = c("AGE"),
          multiple = FALSE,
          fixed = FALSE
        )),
      y = data_extract_spec(
        dataname = "ADSL_2",
        select = select_spec(
          label = "Select variable:",
          choices = c("COUNTRY", "AGE", "RACE"),
          selected = "RACE",
          multiple = FALSE,
          fixed = FALSE
        )),
     )
   )
)

\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: multiple long datasets
#' Percentage cross table of parameters from ADTTE and ADRS datasets.
library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs
ADTTE <- cadtte

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    cdisc_dataset("ADTTE", ADTTE),
    code = "ADSL <- cadsl; ADRS <- cadrs; ADTTE <- cadtte",
    check = TRUE
  ),
  modules = root_modules(
    tm_t_percentage_cross_table(
      label = "Cross Table",
      dataname = c("ADSL", "ADRS", "ADTTE"),
      x = data_extract_spec(
        dataname = "ADRS",
        filter = filter_spec(
          label = "Select endpoints:",
          vars = c("PARAMCD", "AVISIT"),
          choices = apply(expand.grid(levels(ADRS$PARAMCD), levels(ADRS$AVISIT)),
                          1, paste, collapse = " - "),
          selected = "OVRINV - Screening",
          multiple = TRUE
        ),
        select = select_spec(
          label = "Select variable:",
          choices = names(ADRS),
          selected = "AVALC",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADTTE",
        filter = filter_spec(
          label = "Select parameters:",
          vars = "PARAMCD",
          choices = unique(ADTTE$PARAMCD),
          selected = "OS",
          multiple = TRUE
        ),
        select = select_spec(
          label = "Select variable:",
          choices = names(ADTTE),
          selected = "CNSR",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: wide, long
# Percentage cross table with AGE groups per STUDYID
#   filtered by response type and VISIT
library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    code = "ADSL <- cadsl; ADRS <- cadrs",
    check = TRUE
  ),
  modules = root_modules(
    tm_t_percentage_cross_table("Cross Table",
      dataname = c("ADSL", "ADRS"),
      x = data_extract_spec(
        dataname = "ADRS",
        filter = list(
          filter_spec(
            vars = "PARAM",
            choices = levels(ADRS$PARAM),
            selected = levels(ADRS$PARAM)[1],
            multiple = FALSE,
            label = "Choose response:"
          ),
          filter_spec(
            vars = "AVISIT",
            choices = levels(ADRS$AVISIT),
            selected = levels(ADRS$AVISIT)[1],
            multiple = FALSE,
            label = "Choose visit:"
          )
        ),
        select = select_spec(
          label = "Select variable:",
          choices = "STUDYID",
          selected = "STUDYID",
          multiple = FALSE,
          fixed = TRUE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = c("SEX", "AGE", "RACE", "COUNTRY"),
          selected = "AGE",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}

# datasets: same long
# Contingency table of variables from ADRS dataset

library(random.cdisc.data)

ADSL <- cadsl
ADRS <- cadrs

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADRS", ADRS),
    code = "ADSL <- cadsl; ADRS <- cadrs",
    check = TRUE
  ),
  modules = root_modules(
    tm_t_percentage_cross_table(
      "Scatterplot for same long dataset",
      dataname = "ADRS",
      x = data_extract_spec(
        dataname = "ADRS",
        select = select_spec(
          choices = c("PARAMCD", "AVISIT"),
          selected = c("PARAMCD"),
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      ),
      y = data_extract_spec(
        dataname = "ADRS",
        select = select_spec(
          choices = c("AVISIT", "PARAMCD"),
          selected = c("AVISIT"),
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      )
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}


# datasets: different subsets of long dataset
# Contingency table of variables from Lab dataset (ADLB)

library(random.cdisc.data)

ADSL <- cadsl
ADLB <- cadlb

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADLB", ADLB),
    code = "ADSL <- cadsl; ADLB <- cadlb",
    check = TRUE
  ),
  modules = root_modules(
    tm_t_percentage_cross_table(
      label = "Cross Table",
      dataname = "ADLB",
      x = data_extract_spec(
        dataname = "ADLB",
        filter = filter_spec(
          vars = "PARAMCD",
          choices = levels(ADLB$PARAMCD),
          selected = levels(ADLB$PARAMCD)[1],
          multiple = TRUE,
          label = "Choose lab:"
        ),
        select = select_spec(
          choices = names(ADLB),
          selected = "AVISIT",
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      ),
      y = data_extract_spec(
        dataname = "ADLB",
        filter = filter_spec(
          vars = "PARAMCD",
          choices = levels(ADLB$PARAMCD),
          selected = levels(ADLB$PARAMCD)[2],
          multiple = TRUE,
          label = "Choose lab:"
        ),
        select = select_spec(
          choices = names(ADLB),
          selected = "ARMCD",
          multiple = FALSE,
          fixed = FALSE,
          label = "Select variable:"
        )
      )
    )
  )
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
\author{
wolfs25 waddella
}
