% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_percentage_cross_table.R
\name{tm_t_percentage_cross_table}
\alias{tm_t_percentage_cross_table}
\title{Cross table based on rtables}
\usage{
tm_t_percentage_cross_table(label = "Cross Table", dataname, x_var,
  y_var)
}
\arguments{
\item{label}{(\code{character}) Label of the app in the teal menu}

\item{dataname}{(\code{character}) Name of the dataset used in the teal app. Just a single
dataset is allowed!}

\item{x_var}{\code{\link[teal]{choices_selected}} object with all available
choices with preselected option for variable X}

\item{y_var}{\code{\link[teal]{choices_selected}} object with all available
choices with preselected option for variable Y}

\item{...}{Arguments passed on to \code{teal.devel::standard_layout}
\describe{
  \item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
  \item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}}
}
\description{
Cross table based on rtables
}
\examples{
\dontrun{
library(random.cdisc.data)

ASL <- radsl(seed = 1)

x <- teal::init(
  data = cdisc_data(
    ASL = ASL,
    code = "ASL <- radsl(seed = 1)",
    check = TRUE
  ),
  modules = teal::root_modules(
    tm_t_percentage_cross_table(
      label = "Cross Table",
      dataname = "ASL",
      x_var = choices_selected(c("STRATA1"), "STRATA1"),
      y_var = choices_selected(c("STRATA2"), "STRATA2")
    )
  )
)
shinyApp(x$ui, x$server)
}
}
\author{
wolfs25 waddella
}
