% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_t_reactable.R
\name{tm_t_reactables}
\alias{tm_t_reactables}
\title{\code{teal} module: Reactable}
\usage{
tm_t_reactables(
  label = "Table",
  datanames = "all",
  colnames = list(),
  transformators = list(),
  decorators = list(),
  reactable_args = list()
)
}
\arguments{
\item{decorators}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
(named \code{list} of lists of \code{teal_transform_module}) optional,
decorator for tables or plots included in the module output reported.
The decorators are applied to the respective output objects.}

\item{reactable_args}{(\code{list}) any argument of \code{\link[reactable:reactable]{reactable::reactable()}}.}
}
\description{
Wrapper module on \code{\link[reactable:reactable]{reactable::reactable()}}
}
