% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_table.R
\name{tm_table}
\alias{tm_table}
\title{Create a simple cross-table}
\usage{
tm_table(label, dataname, xvar, yvar, useNA = c("ifany", "no", "always"),
  pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{(\code{chracter}) menu label}

\item{dataname}{(\code{chracter}) name of dataset used to generate table}

\item{xvar}{(\code{list} of \code{data_extract_spec}) Specification how the
user can select data to get encoded in the rows of the cross table. Please just use
single selections inside the \code{columns_spec}.}

\item{yvar}{(\code{list} of \code{data_extract_spec}) Specification how the
user can select data to get encoded in the columns of the cross table. Please just use
single selections inside the \code{columns_spec}.}

\item{useNA}{(\code{character}) optional pre-selected option indicating how to utilize NA in
table display. One of \code{'ifany'}, \code{'always'}, \code{'no'}. If
missing then \code{'ifany'} will be used. If vector then only the first
one will be used.}

\item{pre_output}{(\code{shiny.tag}) html tags appended below the output}

\item{post_output}{(\code{shiny.tag}) html tags appended after the output}
}
\description{
Create a table with the \code{\link{table}[base]} function
}
\examples{
library(random.cdisc.data)

ASL <- radsl(seed = 1)

asl_extract_xvar <- data_extract_spec(
    "ASL",
    columns = columns_spec(
        choices = base::setdiff(names(ASL), utils.nest::keys(ASL)),
        selected = names(ASL)[5],
        multiple = FALSE,
        fixed = FALSE
    )
)

asl_extract_yvar <- data_extract_spec(
    "ASL",
    columns = columns_spec(
        choices = base::setdiff(names(ASL), utils.nest::keys(ASL)),
        selected = names(ASL)[6],
        multiple = FALSE,
        fixed = FALSE
    )
)

app <- init(
 data = cdisc_data(
   ASL = ASL,
   code = 'library(random.cdisc.data)
           ASL <- radsl(seed = 1)',
   check = FALSE
 ),
 root_modules(
   tm_table("Table Choices",
            dataname =  "ASL",
            xvar = list(asl_extract_xvar),
            yvar = list(asl_extract_yvar))
 )
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
