% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_tbl_summary.R
\name{tm_tbl_summary}
\alias{tm_tbl_summary}
\title{\code{teal} module: GT Summary table}
\usage{
tm_tbl_summary(
  label = "Summary table",
  by = NULL,
  include = teal.picks::picks(teal.picks::datasets(), teal.picks::variables(selected =
    dplyr::everything(), multiple = TRUE)),
  dataname = NULL,
  ...,
  col_label = NULL,
  pre_output = NULL,
  post_output = NULL,
  transformators = list(),
  decorators = list()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{by}{(\code{picks} or \link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select})
When using picks object it should have \code{\link[teal.picks:variables]{teal.picks::variables()}} to identify a single column from the data.}

\item{include}{(\code{picks} or \link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select})
When using picks object it should have \code{\link[teal.picks:variables]{teal.picks::variables()}} to identify one or more columns from the data.}

\item{dataname}{(\code{string} or \code{NULL}) Name of the dataset to be used in the module if and only if
\code{picks} is not used for other arguments.}

\item{...}{
  Arguments passed on to \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary}}
  \describe{
    \item{\code{statistic}}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies summary statistics to display for each variable.  The default is
\code{list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~ "{n} ({p}\%)")}.
See below for details.}
    \item{\code{digits}}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies how summary statistics are rounded. Values may be either integer(s)
or function(s). If not specified, default formatting is assigned
via \code{assign_summary_digits()}. See below for details.}
    \item{\code{type}}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies the summary type. Accepted value are
\code{c("continuous", "continuous2", "categorical", "dichotomous")}.
If not specified, default type is assigned via
\code{assign_summary_type()}. See below for details.}
    \item{\code{value}}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies the level of a variable to display on a single row.
The gtsummary type selectors, e.g. \code{all_dichotomous()}, cannot be used
with this argument. Default is \code{NULL}. See below for details.}
    \item{\code{missing,missing_text,missing_stat}}{Arguments dictating how and if missing values are presented:
\itemize{
\item \code{missing}: must be one of \code{c("ifany", "no", "always")}.
\item \code{missing_text}: string indicating text shown on missing row. Default is \code{"Unknown"}.
\item \code{missing_stat}: statistic to show on missing row. Default is \code{"{N_miss}"}.
Possible values are \code{N_miss}, \code{N_obs}, \code{N_nonmiss}, \code{p_miss}, \code{p_nonmiss}.
}}
    \item{\code{sort}}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies sorting to perform for categorical variables.
Values must be one of \code{c("alphanumeric", "frequency")}.
Default is \code{all_categorical(FALSE) ~ "alphanumeric"}.}
  }}

\item{col_label}{Used to override default labels in summary table, e.g. \code{list(age = "Age, years")}.
The default for each variable is the column label attribute, \code{attr(., 'label')}.
If no label has been set, the column name is used.}

\item{pre_output}{(\code{shiny.tag}) optional, text or UI element to be displayed before the module's output,
providing context or a title.
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}) optional, text or UI element to be displayed after the module's output,
adding context or further instructions. Elements like \code{shiny::helpText()} are useful.}

\item{transformators}{(\code{list} of \code{teal_transform_module}) that will be applied to transform module's data input.
To learn more check \code{vignette("transform-input-data", package = "teal")}.}

\item{decorators}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
(named \code{list} of lists of \code{teal_transform_module}) optional,
decorator for tables or plots included in the module output reported.
The decorators are applied to the respective output objects.}
}
\value{
Object of class \code{teal_module} to be used in \code{teal} applications.
}
\description{
Summary table from a given dataset, using \code{gtsummary}.
}
\section{Decorating Module}{


This module generates the following objects, which can be modified in place using decorators:
\itemize{
\item \code{table} (\code{gtsummary} - output of \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}})
}

A Decorator is applied to the specific output using a named list of \code{teal_transform_module} objects.
The name of this list corresponds to the name of the output to which the decorator is applied.
See code snippet below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tm_tbl_summary(
   ..., # arguments for module
   decorators = list(
     table = teal_transform_module(...) # applied to the `table` output
   )
)
}\if{html}{\out{</div>}}

For additional details and examples of decorators, refer to the vignette
\code{vignette("decorate-module-output", package = "teal.modules.general")}.

To learn more please refer to the vignette
\code{vignette("transform-module-output", package = "teal")} or the \code{\link[teal:teal_transform_module]{teal::teal_transform_module()}} documentation.
}

\section{statistic argument}{


The statistic argument specifies the statistics presented in the table. The
input dictates the summary statistics presented in the table. For example,
\code{statistic = list(age ~ "{mean} ({sd})")} would report the mean and
standard deviation for age; \code{statistic = list(all_continuous() ~ "{mean} ({sd})")}
would report the mean and standard deviation for all continuous variables.

The values are interpreted using  \code{\link[glue:glue]{glue::glue()}} syntax:
a name that appears between curly brackets will be interpreted as a function
name and the formatted result of that function will be placed in the table.

For categorical variables, the following statistics are available to display:
\code{{n}} (frequency), \code{{N}} (denominator), \code{{p}} (percent).

For continuous variables, \strong{any univariate function may be used}.
The most commonly used functions are \code{{median}}, \code{{mean}}, \code{{sd}}, \code{{min}},
and \code{{max}}.
Additionally, \verb{{p##}} is available for percentiles, where \verb{##} is an integer from 0 to 100.
For example, \code{p25: quantile(probs=0.25, type=2)}.

When the summary type is \code{"continuous2"}, pass a vector of statistics.
Each element of the vector will result in a separate row in the summary table.

For both categorical and continuous variables, statistics on the number of
missing and non-missing observations and their proportions are available to
display.
\itemize{
\item \code{{N_obs}} total number of observations
\item \code{{N_miss}} number of missing observations
\item \code{{N_nonmiss}} number of non-missing observations
\item \code{{p_miss}} percentage of observations missing
\item \code{{p_nonmiss}} percentage of observations not missing
}

}

\section{digits argument}{


The digits argument specifies the the number of digits (or formatting function)
statistics are rounded to.

The values passed can either be a single integer, a vector of integers, a
function, or a list of functions. If a single integer or function is passed,
it is recycled to the length of the number of statistics presented.
For example, if the statistic is \code{"{mean} ({sd})"}, it is equivalent to
pass \code{1}, \code{c(1, 1)}, \code{label_style_number(digits=1)}, and
\code{list(label_style_number(digits=1), label_style_number(digits=1))}.

Named lists are also accepted to change the default formatting for a single
statistic, e.g. \code{list(sd = label_style_number(digits=1))}.

}

\section{type and value arguments}{


There are four summary types. Use the \code{type} argument to change the default summary types.
\itemize{
\item \code{"continuous"} summaries are shown on a \emph{single row}. Most numeric
variables default to summary type continuous.
\item \code{"continuous2"} summaries are shown on \emph{2 or more rows}
\item \code{"categorical"} \emph{multi-line} summaries of nominal data. Character variables,
factor variables, and numeric variables with fewer than 10 unique levels default to
type categorical. To change a numeric variable to continuous that
defaulted to categorical, use \code{type = list(varname ~ "continuous")}
\item \code{"dichotomous"} categorical variables that are displayed on a \emph{single row},
rather than one row per level of the variable.
Variables coded as \code{TRUE}/\code{FALSE}, \code{0}/\code{1}, or \code{yes}/\code{no} are assumed to be dichotomous,
and the \code{TRUE}, \code{1}, and \code{yes} rows are displayed.
Otherwise, the value to display must be specified in the \code{value}
argument, e.g. \code{value = list(varname ~ "level to show")}
}

}

\section{Reporting}{



This module returns an object of class \code{teal_module}, that contains a \code{server} function.
Since the server function returns a \code{teal_report} object, this makes this module reportable, which means that
the reporting functionality will be turned on automatically by the \code{teal} framework.

For more information on reporting in \code{teal}, see the vignettes:
\itemize{
\item \code{vignette("reportable-shiny-application", package = "teal.reporter")}
\item \code{vignette("adding-support-for-reporting-to-custom-modules", package = "teal")}
}

}

\examples{
# General example
data <- teal_data()
data <- within(data, CO2 <- CO2)
app <- init(
  data = data,
  modules = modules(
    tm_tbl_summary(
      by = teal.picks::picks(
        datasets("CO2", "CO2"),
        variables(selected = "Plant")
      ),
      include = teal.picks::picks(
        datasets("CO2", "CO2"),
        variables(selected = c("Type", "Treatment"), multiple = TRUE)
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# CDISC data example
data <- within(teal.data::teal_data(), {
  ADSL <- teal.data::rADSL
})
join_keys(data) <- default_cdisc_join_keys[names(data)]
app <- init(
  data = data,
  modules = modules(
    tm_tbl_summary(
      by = teal.picks::picks(
        datasets(c("ADSL", "ADTTE"), "ADTTE"),
        variables(c("SEX", "COUNTRY", "SITEID", "ACTARM", "CNSR", "PARAMCD"), "SEX")
      ),
      include = teal.picks::picks(
        datasets(c("ADSL", "ADTTE"), "ADSL"),
        variables(c("SITEID", "COUNTRY", "ACTARM", "SEX"), "SITEID", multiple = TRUE)
      )
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
\section{Examples in Shinylive}{
\describe{
  \item{example-1}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQHFps3XAAPWFQRJX4oUigDYy5qAH1wyJswiKijXQB3WlIACxV2RKhcXQBhAHkAJmjSysVodGqVbPYlXV1C3QBedtTcVt0+IRFRLoGBYTEWiDa20hg40np40UEYGBYOfpndelZR2IxUWgIAa1FERCPT0Snt7cLROFIbhTByitfi1-fXuT7pu5tLQsWhQJaTR4iDRwfijV4ABWoUDIvy2Mz+aLaKgI1EE-D03QOVzOF2JLwBgIeTxeb0qn103zpYAxFLuwMYoPBN0hcGhsO6BBaYDsrFQcHprzsjDgEXgKOZxRgwk0IQJ9iczjqgK16P6dTqtBMunYKnIzEsOhsthA-VEeQgrAAguh2GhUAASQS0Ypu92PRg6Rh1AC+SjAwYAukA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQHFps3XAAPWFQRJX4oUigDYy5qAH1wyJswiKijXQB3WlIACxV2RKhcXQBhAHkAJmjSysVodGqVbPYlXV1C3QBedtTcVt0+IRFRLoGBYTEWiDa20hg40np40UEYGBYOfpndelZR2IxUWgIAa1FERCPT0Snt7cLROFIbhTByitfi1-fXuT7pu5tLQsWhQJaTR4iDRwfijV4ABWoUDIvy2Mz+aLaKgI1EE-D03QOVzOF2JLwBgIeTxeb0qn103zpYAxFLuwMYoPBN0hcGhsO6BBaYDsrFQcHprzsjDgEXgKOZxRgwk0IQJ9iczjqgK16P6dTqtBMunYKnIzEsOhsthA-VEeQgrAAguh2GhUAASQS0Ypu92PRg6Rh1AC+SjAwYAukA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
  \item{example-2}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYwB3WlIACxVOOG4MRKhERC5qAH0Cm1xdECVdXQBBHwCAGWTdEvywwsRGRpalAF9FCAArIhVSgGs4VlF2AtsjELgTKGFSUoJ+WlECUrGJ6dngaHg5hYBdJTRUNpV09lqQrt0AXmfw3Ce+IRFRN90P2EYkeEDqdVIMFKpHoZVEghgMBYHCe4N09FYAI6qFoBEmoiKOLxc1RaI+UFEcFIcwIjzAfWaCjwuiZjTsdlcYDkVVZPnZnO5pLRWhYtCgsJBtKZAWcAA0mTywF4APKOAByDgAmgqWWAAn4OX4fDrWV47PUsABZE1KtUBLA2gAKFvqlt8TO5upl8q5Qt0grBZJUBGogn4ene2Nx+MJ0ZJgbJ5Mp1PYUvpTUZzN5-I9ioZub9dRFjDFEppdP1huNWaVqo1WG1NfqZot1pr3tzXoNziNOpg61ocQj9iczmGifH4Mnw2GtBMunYKnIzEsOhsthqgdEWQgrHq6HYNwAJIJaFVj5TGDpGMMBmABhcgA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAGwEsAjAJykYE8AKcqajGIgEwCu1OAGcMAcwpxm1AJQAdCLTIyoBUrQBucAAQAeALS6AZoIgbaJdnN0AVLAFUAokqUBiXQGEAIgEkAZS9dfihSKF04AA9YVBElUPCDYwB3WlIACxVOOG4MRKhERC5qAH0Cm1xdECVdXQBBHwCAGWTdEvywwsRGRpalAF9FCAArIhVSgGs4VlF2AtsjELgTKGFSUoJ+WlECUrGJ6dngaHg5hYBdJTRUNpV09lqQrt0AXmfw3Ce+IRFRN90P2EYkeEDqdVIMFKpHoZVEghgMBYHCe4N09FYAI6qFoBEmoiKOLxc1RaI+UFEcFIcwIjzAfWaCjwuiZjTsdlcYDkVVZPnZnO5pLRWhYtCgsJBtKZAWcAA0mTywF4APKOAByDgAmgqWWAAn4OX4fDrWV47PUsABZE1KtUBLA2gAKFvqlt8TO5upl8q5Qt0grBZJUBGogn4ene2Nx+MJ0ZJgbJ5Mp1PYUvpTUZzN5-I9ioZub9dRFjDFEppdP1huNWaVqo1WG1NfqZot1pr3tzXoNziNOpg61ocQj9iczmGifH4Mnw2GtBMunYKnIzEsOhsthqgdEWQgrHq6HYNwAJIJaFVj5TGDpGMMBmABhcgA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
}
}

