% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{tm_variable_browser}
\alias{tm_variable_browser}
\title{Variable Browser Teal Module}
\usage{
tm_variable_browser(
  label = "Variable Browser",
  pre_output = NULL,
  post_output = NULL
)
}
\arguments{
\item{label}{(\code{character}) Label shown in the navigation item for the module.}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example a title.}

\item{post_output}{optional, \code{shiny.tag} with text placed after the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
The variable browser provides a table with variable names and labels and a
plot that visualizes the content of a particular variable.
}
\details{
Numeric columns with fewer than 30 distinct values can be treated as either factors
or numbers with a checkbox allowing users to switch how they are treated (if < 6 unique values
then the default is categorical, otherwise it is numeric).
}
\examples{
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADTTE <- radtte(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- radsl(cached = TRUE)"),
    cdisc_dataset("ADTTE", ADTTE, code = "ADTTE <- radtte(cached = TRUE)"),
    check = TRUE
  ),
  root_modules(
    tm_variable_browser(label = "Variable browser")
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
}
