% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{tm_variable_browser}
\alias{tm_variable_browser}
\alias{tm_variable_browser_ui,}
\alias{tm_variable_browser_srv,}
\alias{tm_variable_browser,}
\alias{variable_browser_ui,}
\alias{variable_browser_srv,}
\alias{variable_browser}
\title{Variable browser \code{teal} module}
\usage{
tm_variable_browser(
  label = "Variable Browser",
  datasets_selected = character(0),
  parent_dataname = "ADSL",
  pre_output = NULL,
  post_output = NULL,
  ggplot2_args = teal.widgets::ggplot2_args()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{datasets_selected}{(\code{character}) vector of datasets which should be
shown, in order. Names must correspond with datasets names.
If vector of length zero (default) then all datasets are shown.
Note: Only \code{data.frame} objects are compatible; using other types will cause an error.}

\item{parent_dataname}{(\code{character(1)}) string specifying a parent dataset.
If it exists in \code{datasets_selected}then an extra checkbox will be shown to
allow users to not show variables in other datasets which exist in this \code{dataname}.
This is typically used to remove \code{ADSL} columns in \code{CDISC} data.
In non \code{CDISC} data this can be ignored. Defaults to \code{"ADSL"}.}

\item{pre_output}{(\code{shiny.tag}, optional)\cr
with text placed before the output to put the output into context. For example a title.}

\item{post_output}{(\code{shiny.tag}, optional) with text placed after the output to put the output
into context. For example the \code{\link[shiny:helpText]{shiny::helpText()}} elements are useful.}

\item{ggplot2_args}{(\code{ggplot2_args}) object created by \code{\link[teal.widgets:ggplot2_args]{teal.widgets::ggplot2_args()}}
with settings for the module plot.
The argument is merged with options variable \code{teal.ggplot2_args} and default module setup.

For more details see the vignette: \code{vignette("custom-ggplot2-arguments", package = "teal.widgets")}}
}
\description{
Module provides provides a detailed summary and visualization of variable distributions
for \code{data.frame} objects, with interactive features to customize analysis.
}
\details{
Numeric columns with fewer than 30 distinct values can be treated as either categorical
or numeric with a checkbox allowing users to switch how they are treated (if < 6 unique values
then the default is categorical, otherwise it is numeric).
}
\examples{
library(teal.widgets)

data <- teal_data()
data <- within(data, {
  ADSL <- teal.modules.general::rADSL
  ADTTE <- teal.modules.general::rADTTE
})
datanames <- c("ADSL", "ADTTE")
datanames(data) <- datanames
join_keys(data) <- default_cdisc_join_keys[datanames]

app <- init(
  data = data,
  modules(
    tm_variable_browser(
      label = "Variable browser",
      ggplot2_args = ggplot2_args(
        labs = list(subtitle = "Plot generated by Variable Browser Module")
      ),
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
