% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_venn.R
\name{tm_venn2}
\alias{tm_venn2}
\title{venn2 teal module}
\usage{
tm_venn2(label, dataname, bm1, bm2, plot_height = c(600, 200, 2000),
  alpha = c(1, 0, 1), pre_output = tags$p("NAs get currently removed"),
  post_output = NULL)
}
\description{
venn2 teal module
}
\examples{


N <- 100
var_biomarkers <- paste0("B", 1:10)
sample_bm_data <- lapply(1:10, function(x)sample(c(TRUE, FALSE), N, replace = TRUE))
names(sample_bm_data) <- var_biomarkers

ASL <- do.call(data.frame, c(
  list(USUBJID = paste("ID", 1:N), STUDYID = "1"), sample_bm_data
))


app <- init(
  data = cdisc_data(
    ASL = ASL,
    code = '
    N <- 100
    var_biomarkers <- paste0("B", 1:10)
    sample_bm_data <- lapply(1:10, function(x)sample(c(TRUE, FALSE), N, replace = TRUE))
    names(sample_bm_data) <- var_biomarkers
    ASL <- do.call(data.frame, c(
    list(USUBJID = paste("ID", 1:N), STUDYID = "1"), sample_bm_data
    ))
    ',
    check = FALSE),
  modules = root_modules(
    teal.modules.general:::tm_venn2(
      "Venn Diagram", "ASL",
      bm1 = data_extract_spec(
        dataname = "ASL",
        column = columns_spec(
          label = "Select Biomarker",
          choices = var_biomarkers,
          selected = "B1",
          fixed = FALSE)
      ),
      bm2 = data_extract_spec(
        dataname = "ASL",
        column = columns_spec(
          label = "Select Biomarker",
          choices = var_biomarkers,
          selected = "B2",
          fixed = FALSE,)
      )
    )
  )
)
\dontRun{
 shinyApp(app$ui, app$server)
}
}
