% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_basic_table_args}
\alias{validate_basic_table_args}
\title{Additional validation for \code{basic_table_args} argument}
\usage{
validate_basic_table_args(basic_table_args, table_names = NULL)
}
\arguments{
\item{basic_table_args}{(\code{basic_table_args}) class list, generated by \code{basic_table_args_fun()}.
Or list of such objects.}

\item{table_names}{(`character`) vector of all tables names.
Defaults to NULL so not multi-plot in the module.}
}
\value{
nothing
}
\description{
A validation function.
}
