% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_ggplot2_args}
\alias{validate_ggplot2_args}
\title{Additional validation for \code{ggplot2_args} argument}
\usage{
validate_ggplot2_args(ggplot2_args, plot_names = NULL)
}
\description{
Additional validation for \code{ggplot2_args} argument
}
