% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{validate_input}
\alias{validate_input}
\title{Validates the variable browser inputs}
\usage{
validate_input(input, plot_var, data)
}
\arguments{
\item{input}{(\code{session$input}) the shiny session input}

\item{plot_var}{(\code{list}) list of a data frame and an array of variable names}

\item{data}{(\code{tdata}) the datasets passed to the module}
}
\value{
\code{logical} TRUE if validations pass; a Shiny validation error otherwise
}
\description{
Validates the variable browser inputs
}
\keyword{internal}
