% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{validate_input}
\alias{validate_input}
\title{Validates the variable browser inputs}
\usage{
validate_input(
  inputId,
  condition = function(x) TRUE,
  message = "",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{input}{(\code{session$input}) the \code{shiny} session input}

\item{plot_var}{(\code{list}) list of a data frame and an array of variable names}

\item{data}{(\code{teal_data}) the datasets passed to the module}
}
\value{
\code{logical} TRUE if validations pass; a \code{shiny} validation error otherwise
}
\description{
Validates the variable browser inputs
}
\keyword{internal}
