% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{var_missings_info}
\alias{var_missings_info}
\title{Summarizes missings occurrence}
\usage{
var_missings_info(x)
}
\arguments{
\item{x}{vector of any type and length}
}
\value{
text describing \code{NA} occurrence.
}
\description{
Summarizes missings occurrence in vector
}
\keyword{internal}
