% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{var_summary_table}
\alias{var_summary_table}
\title{Summarizes variable}
\usage{
var_summary_table(x, numeric_as_factor)
}
\arguments{
\item{x}{vector of any type}

\item{numeric_as_factor}{\code{logical} should the numeric variable be treated as a factor}
}
\value{
text with simple statistics.
}
\description{
Creates html summary with statistics relevant to data type. For numeric values it returns central
tendency measures, for factor returns level counts, for Date  date range, for other just
number of levels.
}
