% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_variable_browser.R
\name{var_summary_table}
\alias{var_summary_table}
\title{Summarizes variable}
\usage{
var_summary_table(x, numeric_as_factor, dt_rows, outlier_definition)
}
\arguments{
\item{x}{vector of any type}

\item{numeric_as_factor}{\code{logical} should the numeric variable be treated as a factor}

\item{dt_rows}{\code{numeric} current/latest \code{DT} page length}

\item{outlier_definition}{If 0 no outliers are removed, otherwise
outliers (those more than outlier_definition*IQR below/above Q1/Q3 be removed)}
}
\value{
text with simple statistics.
}
\description{
Creates html summary with statistics relevant to data type. For numeric values it returns central
tendency measures, for factor returns level counts, for Date  date range, for other just
number of levels.
}
\keyword{internal}
