% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varname_w_label}
\alias{varname_w_label}
\title{Generate a string for a variable including its label}
\usage{
varname_w_label(
  var_names,
  dataset,
  wrap_width = 80,
  prefix = NULL,
  suffix = NULL
)
}
\arguments{
\item{var_names}{(\code{character}) Name of variable to extract labels from.}

\item{dataset}{(\code{dataset}) Name of analysis dataset.}

\item{wrap_width}{(\code{numeric}) Number of characters to wrap original label to. Defaults to 80.}

\item{prefix, suffix}{(\code{character}) String to paste to the beginning/end of the variable name with label.}
}
\value{
(\code{character}) String with variable name and label.
}
\description{
Generate a string for a variable including its label
}
\keyword{internal}
